### 2022-05-17 #4400 @editorialbot
[PRE REVIEW]: PyAutoGalaxy: Open-Source Multiwavelength Galaxy Structure & Morphology
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** https://github.com/Jammy2211/PyAutoGalaxy
**Branch with paper.md** (empty if default branch):
**Version:** 2022.05.02.1
