### 2022-05-18 60578c4
moved inversion plane funcitonality to new module
