+++ autogalaxy/plane/plane.py
@@ -6,13 +6,8 @@ import autoarray as aa
 
 from autoconf.dictable import Dictable
 
-from autoarray.inversion.inversion.factory import inversion_imaging_unpacked_from
-from autoarray.inversion.inversion.factory import inversion_interferometer_unpacked_from
 from autogalaxy import exc
 from autogalaxy.galaxy.galaxy import Galaxy
-from autogalaxy.profiles.light_profiles.light_profiles_linear import (
-    LightProfileLinearObjFunc,
-)
 from autogalaxy.profiles.light_profiles.light_profiles_snr import LightProfileSNR
 from autogalaxy.operate.image import OperateImageGalaxies
 from autogalaxy.operate.deflections import OperateDeflections
@@ -243,24 +238,6 @@ class Plane(OperateImageGalaxies, OperateDeflections, Dictable):
         """Trace this plane's grid_stacks to the next plane, using its deflection angles."""
         return grid - self.deflections_yx_2d_from(grid=grid)
 
-    def sparse_image_plane_grid_list_from(
-        self, grid: aa.type.Grid2DLike, settings_pixelization=aa.SettingsPixelization()
-    ) -> Optional[List[aa.type.Grid2DLike]]:
-
-        if not self.has_pixelization:
-            return None
-
-        return [
-            pixelization.data_pixelization_grid_from(
-                data_grid_slim=grid,
-                hyper_image=hyper_galaxy_image,
-                settings=settings_pixelization,
-            )
-            for pixelization, hyper_galaxy_image in zip(
-                self.pixelization_list, self.hyper_galaxies_with_pixelization_image_list
-            )
-        ]
-
     @property
     def has_pixelization(self) -> bool:
         return any([galaxy.pixelization for galaxy in self.galaxies])
@@ -288,169 +265,6 @@ class Plane(OperateImageGalaxies, OperateDeflections, Dictable):
     @property
     def regularization_list(self) -> List:
         return [galaxy.regularization for galaxy in self.galaxies_with_pixelization]
-
-    def light_profile_linear_func_list_from(
-        self,
-        source_grid_slim,
-        source_blurring_grid_slim,
-        convolver: Optional[aa.Convolver] = None,
-    ):
-
-        if not self.has_light_profile_linear:
-            return []
-
-        light_profile_linear_list = []
-
-        for galaxy in self.galaxies:
-            if galaxy.has_light_profile_linear:
-                for light_profile_linear in galaxy.light_profile_linear_list:
-
-                    light_profile_linear_func = LightProfileLinearObjFunc(
-                        grid=source_grid_slim,
-                        blurring_grid=source_blurring_grid_slim,
-                        convolver=convolver,
-                        light_profile=light_profile_linear,
-                    )
-
-                    light_profile_linear_list.append(light_profile_linear_func)
-
-        return light_profile_linear_list
-
-    def mapper_from(
-        self,
-        source_grid_slim,
-        source_pixelization_grid,
-        pixelization,
-        hyper_galaxy_image,
-        data_pixelization_grid=None,
-        settings_pixelization=aa.SettingsPixelization(),
-        preloads=aa.Preloads(),
-    ):
-
-        return pixelization.mapper_from(
-            source_grid_slim=source_grid_slim,
-            source_pixelization_grid=source_pixelization_grid,
-            data_pixelization_grid=data_pixelization_grid,
-            hyper_image=hyper_galaxy_image,
-            settings=settings_pixelization,
-            preloads=preloads,
-            profiling_dict=self.profiling_dict,
-        )
-
-    def mapper_list_from(
-        self,
-        grid,
-        settings_pixelization=aa.SettingsPixelization(),
-        preloads=aa.Preloads(),
-    ):
-
[…]
