+++ autogalaxy/interferometer/fit_interferometer.py
@@ -8,7 +8,7 @@ import autoarray as aa
 from autogalaxy.galaxy.galaxy import Galaxy
 from autogalaxy.hyper.hyper_data import HyperBackgroundNoise
 from autogalaxy.plane.plane import Plane
-from autogalaxy.plane.inversion_maker import InversionMaker
+from autogalaxy.plane.to_inversion import PlaneToInversion
 
 
 class FitInterferometer(aa.FitInterferometer):
@@ -95,9 +95,9 @@ class FitInterferometer(aa.FitInterferometer):
             else:
                 w_tilde = None
 
-            inversion_maker = InversionMaker(plane=self.plane)
+            plane_to_inversion = PlaneToInversion(plane=self.plane)
 
-            return inversion_maker.inversion_interferometer_from(
+            return plane_to_inversion.inversion_interferometer_from(
                 dataset=self.dataset,
                 visibilities=self.profile_subtracted_visibilities,
                 noise_map=self.noise_map,

[…]
