+++ autogalaxy/profiles/light_profiles/light_profiles_linear.py
@@ -95,6 +95,67 @@ class EllSersic(lp.EllSersic, LightProfileLinear):
         )
 
 
+class EllExponential(lp.EllExponential, LightProfileLinear):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        effective_radius: float = 0.6,
+    ):
+        """
+        The elliptical Exponential light profile.
+
+        See `autogalaxy.profiles.light_profiles.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        effective_radius
+            The circular radius containing half the light of this profile.
+        """
+        super().__init__(
+            centre=centre,
+            elliptical_comps=elliptical_comps,
+            intensity=1.0,
+            effective_radius=effective_radius,
+        )
+
+
+class EllDevVaucouleurs(lp.EllDevVaucouleurs, LightProfileLinear):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        effective_radius: float = 0.6,
+    ):
+        """
+        The elliptical DevVaucouleurs light profile.
+
+        See `autogalaxy.profiles.light_profiles.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        effective_radius
+            The circular radius containing half the light of this profile.
+        """
+        super().__init__(
+            centre=centre,
+            elliptical_comps=elliptical_comps,
+            intensity=1.0,
+            effective_radius=effective_radius,
+        )
+
+
+
 class EllGaussian(lp.EllGaussian, LightProfileLinear):
     def __init__(
         self,
