+++ autogalaxy/imaging/fit_imaging.py
@@ -22,7 +22,7 @@ class FitImaging(aa.FitImaging):
         use_hyper_scaling: bool = True,
         settings_pixelization: aa.SettingsPixelization = aa.SettingsPixelization(),
         settings_inversion: aa.SettingsInversion = aa.SettingsInversion(),
-        preloads=aa.Preloads(),
+        preloads:aa.Preloads=aa.Preloads(),
         profiling_dict: Optional[Dict] = None,
     ):
         """ An lens fitter, which contains the plane's used to perform the fit and functions to manipulate \
