+++ test_autogalaxy/interferometer/model/test_visualizer_interferometer.py
@@ -32,9 +32,7 @@ def test__visualizes_fit_interferometer__uses_configs(
 
     plot_path = path.join(plot_path, "fit_interferometer")
 
-    assert (
-        path.join(plot_path, "subplot_fit_interferometer.png") in plot_patch.paths
-    )
+    assert path.join(plot_path, "subplot_fit_interferometer.png") in plot_patch.paths
     assert path.join(plot_path, "visibilities.png") in plot_patch.paths
     assert path.join(plot_path, "noise_map.png") not in plot_patch.paths
     assert path.join(plot_path, "signal_to_noise_map.png") not in plot_patch.paths
