+++ autogalaxy/profiles/light_profiles/light_profiles_linear.py
@@ -155,7 +155,6 @@ class EllDevVaucouleurs(lp.EllDevVaucouleurs, LightProfileLinear):
         )
 
 
-
 class EllGaussian(lp.EllGaussian, LightProfileLinear):
     def __init__(
         self,
