### 2022-05-23 5c4c994
more fit imaging / interferomete tests for linear objs
