+++ test_autogalaxy/interferometer/test_fit_interferometer.py
@@ -95,11 +95,6 @@ def test__fit_figure_of_merit(interferometer_7):
     assert fit.log_evidence == pytest.approx(-283424.48941, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-283424.48941, 1.0e-4)
 
-
-def test__fit_figure_of_merit__different_linear_obj_lists_for_inversion(
-    interferometer_7
-):
-
     g0_linear_light = ag.Galaxy(
         redshift=0.5, light_profile=ag.lp_linear.EllSersic(sersic_index=1.0)
     )
@@ -118,6 +113,17 @@ def test__fit_figure_of_merit__different_linear_obj_lists_for_inversion(
 
     assert fit.log_likelihood == pytest.approx(-23.44419, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-23.44419, 1.0e-4)
+
+    plane = ag.Plane(redshift=0.5, galaxies=[g0_linear_light, galaxy_pix])
+
+    fit = ag.FitInterferometer(
+        dataset=interferometer_7,
+        plane=plane,
+        settings_inversion=ag.SettingsInversion(use_w_tilde=False),
+    )
+
+    assert fit.log_evidence == pytest.approx(-35.16806296, 1e-4)
+    assert fit.figure_of_merit == pytest.approx(-35.16806296, 1.0e-4)
 
 
 def test__fit_figure_of_merit__include_hyper_methods(interferometer_7):
