### 2022-05-23 f1076b5
linear obj func tests dont require use_w_tilde false, automatically occurxs in autoarray
