### 2022-05-23 #4400 @christinahedges
Hi @Jammy2211 thanks for your submission to JOSS! I'm excited to get this review kicked off, this seems like an exciting tool. I'm going to start by reading through your submission, and then I will ping some potential reviewers. We will find at least two reviewers, and then start the review. They will go through a checklist and then provide you feedback on `PyAutoGalaxy`.

I have some potential reviewers in mind that I can start with, but sometimes it can be hard to find reviewers who have no [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). If you know of researchers who have relevant expertise and are not conflicted with you, please feel free to suggest their names here, otherwise I will search for reviewers and get back to you.

It can sometimes take a while to get together reviewers, but after we find them the review is usually complete in 4-6 weeks.
