### 2022-05-24 aa13d6e
intensity linear light profile dict in fit imaging
