### 2022-05-26 6aa75a8
inversion intensity dict in interferometr
