+++ autogalaxy/profiles/mass_profiles/mass_profiles.py
@@ -55,18 +55,14 @@ class MassProfile(EllProfile, OperateDeflections):
         raise NotImplementedError
 
     @aa.grid_dec.grid_1d_to_structure
-    def convergence_1d_from(
-        self, grid: aa.type.Grid1D2DLike, radial_grid_shape_slim: Optional[int] = None
-    ) -> aa.type.Grid1D2DLike:
+    def convergence_1d_from(self, grid: aa.type.Grid1D2DLike) -> aa.type.Grid1D2DLike:
         return self.convergence_2d_from(grid=grid)
 
     def potential_2d_from(self, grid):
         raise NotImplementedError
 
     @aa.grid_dec.grid_1d_to_structure
-    def potential_1d_from(
-        self, grid: aa.type.Grid1D2DLike, radial_grid_shape_slim: Optional[int] = None
-    ) -> aa.type.Grid1D2DLike:
+    def potential_1d_from(self, grid: aa.type.Grid1D2DLike) -> aa.type.Grid1D2DLike:
         return self.potential_2d_from(grid=grid)
 
     def potential_func(self, u, y, x):
