### 2022-05-30 9df2bd2
new API for slam
