+++ autogalaxy/interferometer/model/analysis.py
@@ -405,7 +405,7 @@ class AnalysisInterferometer(AnalysisDataset):
             )
 
     def make_result(
-        self, samples: af.PDFSamples, model: af.Collection, search: af.NonLinearSearch
+        self, samples: af.PDFSamples, model: af.Collection, sigma=1.0, use_errors=True, use_widths=False
     ) -> ResultInterferometer:
         """
         After the non-linear search is complete create its `Result`, which includes:
@@ -438,7 +438,7 @@ class AnalysisInterferometer(AnalysisDataset):
             The result of fitting the model to the interferometer dataset, via a non-linear search.
         """
         return ResultInterferometer(
-            samples=samples, model=model, analysis=self, search=search
+            samples=samples, model=model, analysis=self,
         )
 
     def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
