+++ autogalaxy/imaging/model/analysis.py
@@ -352,7 +352,7 @@ class AnalysisImaging(AnalysisDataset):
                 pass
 
     def make_result(
-        self, samples: af.PDFSamples, model: af.Collection, search: af.NonLinearSearch
+        self, samples: af.PDFSamples, model: af.Collection, sigma=1.0, use_errors=True, use_widths=False
     ) -> ResultImaging:
         """
         After the non-linear search is complete create its `Result`, which includes:
@@ -384,7 +384,7 @@ class AnalysisImaging(AnalysisDataset):
         ResultImaging
             The result of fitting the model to the imaging dataset, via a non-linear search.
         """
-        return ResultImaging(samples=samples, model=model, analysis=self, search=search)
+        return ResultImaging(samples=samples, model=model, analysis=self)
 
     def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
         """
