+++ autogalaxy/profiles/light_profiles/light_profiles_linear.py
@@ -1,18 +1,48 @@
+import inspect
 import numpy as np
 from typing import Dict, Optional, Tuple, Union
 
 from autoconf import cached_property
 import autoarray as aa
+import autofit as af
 
 from autogalaxy.profiles.light_profiles import light_profiles as lp
+from autogalaxy.profiles import light_and_mass_profiles as lmp
 
 
 class LightProfileLinear(lp.LightProfile):
     @property
-    def not_linear_cls(self):
+    def lp_cls(self):
         raise NotImplementedError
 
-    def not_linear_instance_from(self, intensity: float) -> lp.LightProfile:
+    @property
+    def lmp_cls(self):
+        raise NotImplementedError
+
+    def parameters_dict_from(self, intensity: float) -> Dict[str, float]:
+        """
+        Returns a dictionary of the parameters of the linear light profile with the `intensity` added.
+
+        This `intenisty` will likely have come from the value inferred via the linear inversion.
+
+        Parameters
+        ----------
+        intensity
+            Overall intensity normalisation of the not linear light profile that is created (units are dimensionless
+            and derived from the data the light profile's image is compared too, which is expected to be electrons
+            per second).
+        """
+        parameters_dict = vars(self)
+
+        args = inspect.getfullargspec(self.lp_cls.__init__).args
+        args.remove("self")
+
+        parameters_dict = {key: parameters_dict[key] for key in args}
+        parameters_dict["intensity"] = intensity
+
+        return parameters_dict
+
+    def lp_instance_from(self, intensity: float) -> lp.LightProfile:
         """
         Creates an instance of a linear light profile using its parent normal light profile (e.g. the non linear
         variant which has an `intensity` parameter).
@@ -26,10 +56,27 @@ class LightProfileLinear(lp.LightProfile):
             and derived from the data the light profile's image is compared too, which is expected to be electrons
             per second).
         """
-        parameters = vars(self)
-        parameters["intensity"] = intensity
+        parameters_dict = self.parameters_dict_from(intensity=intensity)
 
-        return self.not_linear_cls(**parameters)
+        return self.lp_cls(**parameters_dict)
+
+    def lmp_model_from(self, intensity: float) -> af.Model(lmp.LightMassProfile):
+        """
+        Creates an instance of a linear light profile using its parent light and mass profile (e.g. the non linear
+        variant which has `mass_to_light_ratio` and `intensity` parameters).
+
+        The `intensity` value of the profile created is passed into this function and used.
+
+        Parameters
+        ----------
+        intensity
+            Overall intensity normalisation of the not linear light profile that is created (units are dimensionless
+            and derived from the data the light profile's image is compared too, which is expected to be electrons
+            per second).
+        """
+        parameters_dict = self.parameters_dict_from(intensity=intensity)
+
+        return af.Model(self.lmp_cls, **parameters_dict)
 
 
 class LightProfileLinearObjFunc(aa.LinearObjFunc):
@@ -115,8 +162,12 @@ class EllSersic(lp.EllSersic, LightProfileLinear):
         )
 
     @property
-    def not_linear_cls(self):
+    def lp_cls(self):
         return lp.EllSersic
+
+    @property
+    def lmp_cls(self):
+        return lmp.EllSersic
 
 
 class EllExponential(lp.EllExponential, LightProfileLinear):
@@ -149,8 +200,12 @@ class EllExponential(lp.EllExponential, LightProfileLinear):
         )
 
     @property
-    def not_linear_cls(self):
+    def lp_cls(self):
         return lp.EllExponential
+
+    @property
+    def lmp_cls(self):
+        return lmp.EllExponential
 
 
 class EllDevVaucouleurs(lp.EllDevVaucouleurs, LightProfileLinear):
[…]
