+++ autogalaxy/quantity/model/analysis.py
@@ -151,7 +151,7 @@ class AnalysisQuantity(Analysis):
         visualizer.visualize_fit_quantity(fit=fit)
 
     def make_result(
-        self, samples: af.PDFSamples, model: af.Collection, search: af.NonLinearSearch
+        self, samples: af.PDFSamples, model: af.Collection, sigma=1.0, use_errors=True, use_widths=False
     ) -> ResultQuantity:
         """
         After the non-linear search is complete create its `ResultQuantity`, which includes:
@@ -184,7 +184,7 @@ class AnalysisQuantity(Analysis):
             The result of fitting the model to the imaging dataset, via a non-linear search.
         """
         return ResultQuantity(
-            samples=samples, model=model, analysis=self, search=search
+            samples=samples, model=model, analysis=self,
         )
 
     def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
