+++ test_autogalaxy/imaging/model/test_result_imaging.py
@@ -15,7 +15,6 @@ class TestResultImaging:
             samples=samples_with_result,
             analysis=analysis_imaging_7x7,
             model=None,
-            search=None,
         )
 
         assert (result.mask == masked_imaging_7x7.mask).all()
@@ -33,7 +32,7 @@ class TestResultImaging:
         samples = ag.m.MockSamples(max_log_likelihood_instance=instance)
 
         result = ResultImaging(
-            samples=samples, analysis=analysis_imaging_7x7, model=None, search=None
+            samples=samples, analysis=analysis_imaging_7x7, model=None,
         )
 
         image_dict = result.image_galaxy_dict
