### 2022-05-30 77bc507
tests on galaxy model image dict for different linear light profiles
