+++ autogalaxy/interferometer/model/analysis.py
@@ -405,7 +405,12 @@ class AnalysisInterferometer(AnalysisDataset):
             )
 
     def make_result(
-        self, samples: af.PDFSamples, model: af.Collection, sigma=1.0, use_errors=True, use_widths=False
+        self,
+        samples: af.PDFSamples,
+        model: af.Collection,
+        sigma=1.0,
+        use_errors=True,
+        use_widths=False,
     ) -> ResultInterferometer:
         """
         After the non-linear search is complete create its `Result`, which includes:
@@ -437,9 +442,7 @@ class AnalysisInterferometer(AnalysisDataset):
         ResultInterferometer
             The result of fitting the model to the interferometer dataset, via a non-linear search.
         """
-        return ResultInterferometer(
-            samples=samples, model=model, analysis=self,
-        )
+        return ResultInterferometer(samples=samples, model=model, analysis=self)
 
     def save_attributes_for_aggregator(self, paths: af.DirectoryPaths):
         """
