+++ autogalaxy/imaging/model/analysis.py
@@ -352,7 +352,12 @@ class AnalysisImaging(AnalysisDataset):
                 pass
 
     def make_result(
-        self, samples: af.PDFSamples, model: af.Collection, sigma=1.0, use_errors=True, use_widths=False
+        self,
+        samples: af.PDFSamples,
+        model: af.Collection,
+        sigma=1.0,
+        use_errors=True,
+        use_widths=False,
     ) -> ResultImaging:
         """
         After the non-linear search is complete create its `Result`, which includes:
