+++ autogalaxy/plane/to_inversion.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import Dict, List, Optional, Union
 
 import autoarray as aa
 
@@ -7,12 +7,40 @@ from autoarray.inversion.inversion.factory import inversion_interferometer_unpac
 from autogalaxy.profiles.light_profiles.light_profiles_linear import (
     LightProfileLinearObjFunc,
 )
+from autogalaxy.galaxy.galaxy import Galaxy
 
 
 class PlaneToInversion:
-    def __init__(self, plane):
+    def __init__(self, plane: "Plane"):
 
         self.plane = plane
+
+    def lp_linear_func_galaxy_dict_from(
+        self,
+        source_grid_slim: aa.type.Grid2DLike,
+        source_blurring_grid_slim: Optional[aa.type.Grid2DLike],
+        convolver: Optional[aa.Convolver] = None,
+    ) -> Dict[LightProfileLinearObjFunc, Galaxy]:
+
+        if not self.plane.has_light_profile_linear:
+            return {}
+
+        lp_linear_func_galaxy_dict = {}
+
+        for galaxy in self.plane.galaxies:
+            if galaxy.has_light_profile_linear:
+                for light_profile_linear in galaxy.light_profile_linear_list:
+
+                    lp_linear_func = LightProfileLinearObjFunc(
+                        grid=source_grid_slim,
+                        blurring_grid=source_blurring_grid_slim,
+                        convolver=convolver,
+                        light_profile=light_profile_linear,
+                    )
+
+                    lp_linear_func_galaxy_dict[lp_linear_func] = galaxy
+
+        return lp_linear_func_galaxy_dict
 
     def sparse_image_plane_grid_list_from(
         self, grid: aa.type.Grid2DLike, settings_pixelization=aa.SettingsPixelization()
@@ -33,43 +61,16 @@ class PlaneToInversion:
             )
         ]
 
-    def light_profile_linear_func_list_from(
-        self,
-        source_grid_slim,
-        source_blurring_grid_slim,
-        convolver: Optional[aa.Convolver] = None,
-    ):
-
-        if not self.plane.has_light_profile_linear:
-            return []
-
-        light_profile_linear_list = []
-
-        for galaxy in self.plane.galaxies:
-            if galaxy.has_light_profile_linear:
-                for light_profile_linear in galaxy.light_profile_linear_list:
-
-                    light_profile_linear_func = LightProfileLinearObjFunc(
-                        grid=source_grid_slim,
-                        blurring_grid=source_blurring_grid_slim,
-                        convolver=convolver,
-                        light_profile=light_profile_linear,
-                    )
-
-                    light_profile_linear_list.append(light_profile_linear_func)
-
-        return light_profile_linear_list
-
     def mapper_from(
         self,
-        source_grid_slim,
-        source_pixelization_grid,
-        pixelization,
-        hyper_galaxy_image,
-        data_pixelization_grid=None,
+        source_grid_slim: aa.type.Grid2DLike,
+        source_pixelization_grid: aa.type.Grid2DLike,
+        pixelization: aa.AbstractPixelization,
+        hyper_galaxy_image: aa.Array2D,
+        data_pixelization_grid: aa.Grid2D = None,
         settings_pixelization=aa.SettingsPixelization(),
         preloads=aa.Preloads(),
-    ):
+    ) -> aa.AbstractMapper:
 
         return pixelization.mapper_from(
             source_grid_slim=source_grid_slim,
@@ -81,21 +82,22 @@ class PlaneToInversion:
             profiling_dict=self.plane.profiling_dict,
         )
 
-    def mapper_list_from(
[…]
