+++ test_autogalaxy/plane/test_inversion_maker.py
@@ -1,239 +1,0 @@
-import numpy as np
-import pytest
-
-import autogalaxy as ag
-
-
-def test__light_profile_linear_func_list_from(lp_0):
-
-    plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5)], redshift=None)
-
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
-
-    light_profile_linear_func_list = plane_to_inversion.light_profile_linear_func_list_from(
-        source_grid_slim=1, source_blurring_grid_slim=2, convolver=3
-    )
-
-    assert light_profile_linear_func_list == []
-
-    lp_linear_0 = ag.lp_linear.LightProfileLinear()
-    lp_linear_1 = ag.lp_linear.LightProfileLinear()
-    lp_linear_2 = ag.lp_linear.LightProfileLinear()
-
-    plane = ag.Plane(
-        galaxies=[
-            ag.Galaxy(
-                redshift=0.5,
-                lp_0=lp_0,
-                light_linear_0=lp_linear_0,
-                light_linear_1=lp_linear_1,
-            ),
-            ag.Galaxy(redshift=0.5, light_linear=lp_linear_2),
-        ],
-        redshift=None,
-    )
-
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
-
-    light_profile_linear_func_list = plane_to_inversion.light_profile_linear_func_list_from(
-        source_grid_slim=1, source_blurring_grid_slim=2, convolver=3
-    )
-
-    assert light_profile_linear_func_list[0].light_profile == lp_linear_0
-    assert light_profile_linear_func_list[1].light_profile == lp_linear_1
-    assert light_profile_linear_func_list[2].light_profile == lp_linear_2
-
-
-def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
-
-    galaxy_pix = ag.Galaxy(
-        redshift=0.5,
-        pixelization=ag.m.MockPixelization(data_pixelization_grid=[[1.0, 1.0]]),
-        regularization=ag.m.MockRegularization(),
-    )
-
-    plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
-
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
-
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
-        grid=sub_grid_2d_7x7
-    )
-
-    assert (sparse_grid == np.array([[1.0, 1.0]])).all()
-
-    # In the ag.m.MockPixelization class the grid is returned if hyper image=None, and grid*hyper image is
-    # returned otherwise.
-
-    galaxy_pix = ag.Galaxy(
-        redshift=0.5,
-        pixelization=ag.m.MockPixelization(
-            data_pixelization_grid=np.array([[1.0, 1.0]])
-        ),
-        regularization=ag.m.MockRegularization(),
-        hyper_galaxy_image=2,
-    )
-
-    plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
-
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
-
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
-        grid=sub_grid_2d_7x7
-    )
-
-    assert (sparse_grid == np.array([[2.0, 2.0]])).all()
-
-    # No Galalxies
-
-    galaxy_no_pix = ag.Galaxy(redshift=0.5)
-
-    plane = ag.Plane(galaxies=[galaxy_no_pix], redshift=0.5)
-
[…]

[…]
