### 2022-05-30 b6c5b26
galaxy model image dict in fit interfereomter
