+++ test_autogalaxy/interferometer/test_fit_interferometer.py
@@ -50,9 +50,9 @@ def test__fit_figure_of_merit(interferometer_7):
         mass_profile=ag.mp.SphIsothermal(einstein_radius=1.0),
     )
 
-    g1 = ag.Galaxy(redshift=1.0, light_profile=ag.lp.EllSersic(intensity=1.0))
+    g1 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
 
-    plane = ag.Plane(redshift=0.75, galaxies=[g0, g1])
+    plane = ag.Plane(redshift=0.5, galaxies=[g0, g1])
 
     fit = ag.FitInterferometer(dataset=interferometer_7, plane=plane)
 
@@ -81,9 +81,9 @@ def test__fit_figure_of_merit(interferometer_7):
 
     pix = ag.pix.Rectangular(shape=(3, 3))
     reg = ag.reg.Constant(coefficient=1.0)
-    galaxy_pix = ag.Galaxy(redshift=1.0, pixelization=pix, regularization=reg)
+    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
 
-    plane = ag.Plane(redshift=0.75, galaxies=[galaxy_light, galaxy_pix])
+    plane = ag.Plane(redshift=0.5, galaxies=[galaxy_light, galaxy_pix])
 
     fit = ag.FitInterferometer(
         dataset=interferometer_7,
@@ -128,9 +128,9 @@ def test__fit_figure_of_merit__include_hyper_methods(interferometer_7):
         mass_profile=ag.mp.SphIsothermal(einstein_radius=1.0),
     )
 
-    g1 = ag.Galaxy(redshift=1.0, light_profile=ag.lp.EllSersic(intensity=1.0))
+    g1 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
 
-    plane = ag.Plane(redshift=0.75, galaxies=[g0, g1])
+    plane = ag.Plane(redshift=0.5, galaxies=[g0, g1])
 
     fit = ag.FitInterferometer(
         dataset=interferometer_7,
@@ -175,9 +175,9 @@ def test__fit_figure_of_merit__include_hyper_methods(interferometer_7):
 
     pix = ag.pix.Rectangular(shape=(3, 3))
     reg = ag.reg.Constant(coefficient=1.0)
-    galaxy_pix = ag.Galaxy(redshift=1.0, pixelization=pix, regularization=reg)
+    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
 
-    plane = ag.Plane(redshift=0.75, galaxies=[galaxy_light, galaxy_pix])
+    plane = ag.Plane(redshift=0.5, galaxies=[galaxy_light, galaxy_pix])
 
     fit = ag.FitInterferometer(
         dataset=interferometer_7,
@@ -217,14 +217,17 @@ def test___fit_figure_of_merit__different_settings(
 
 def test___galaxy_model_image_dict(interferometer_7):
 
-    g0 = ag.Galaxy(
-        redshift=0.5,
-        light_profile=ag.lp.EllSersic(intensity=1.0),
-        mass_profile=ag.mp.SphIsothermal(einstein_radius=1.0),
-    )
-    g1 = ag.Galaxy(redshift=1.0, light_profile=ag.lp.EllSersic(intensity=1.0))
+    # Normal Light Profiles Only
 
-    plane = ag.Plane(redshift=0.75, galaxies=[g0, g1])
[…]
