### 2022-05-30 6f84004
fit interferometer now works
