+++ test_autogalaxy/interferometer/test_fit_interferometer.py
@@ -285,13 +285,14 @@ def test___galaxy_model_image_dict(interferometer_7):
         settings=ag.SettingsInversion(use_w_tilde=False),
     )
 
-    assert (fit.galaxy_model_image_dict[g0].native == np.zeros((7, 7))).all()
+    assert (fit.galaxy_model_image_dict[g0].native == 0.0 + 0.0j * np.zeros((7,))).all()
 
     assert fit.galaxy_model_image_dict[galaxy_pix_0] == pytest.approx(
         inversion.mapped_reconstructed_image.slim, 1.0e-4
     )
 
     # Linear Light PRofiles + Pixelization + Regularizaiton
+    reg = ag.reg.Constant(coefficient=2.0)
 
     galaxy_pix_1 = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
 
@@ -304,14 +305,14 @@ def test___galaxy_model_image_dict(interferometer_7):
     )
 
     assert fit.galaxy_model_image_dict[g0_linear][4] == pytest.approx(
-        -1.65495642e03, 1.0e-4
+        -1.65495642e03, 1.0e-2
     )
     assert fit.galaxy_model_image_dict[g1] == pytest.approx(g1_image, 1.0e-4)
     assert fit.galaxy_model_image_dict[galaxy_pix_0][4] == pytest.approx(
-        -0.00016439460, 1.0e-4
+        -0.000164926, 1.0e-2
     )
     assert fit.galaxy_model_image_dict[galaxy_pix_1][4] == pytest.approx(
-        -0.00016439343, 1.0e-4
+        -0.000153471881, 1.0e-2
     )
 
     mapped_reconstructed_image = (
@@ -327,14 +328,17 @@ def test___galaxy_model_image_dict(interferometer_7):
 
 def test___galaxy_model_visibilities_dict(interferometer_7):
 
-    g0 = ag.Galaxy(
-        redshift=0.5,
-        light_profile=ag.lp.EllSersic(intensity=1.0),
-        mass_profile=ag.mp.SphIsothermal(einstein_radius=1.0),
-    )
-    g1 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
+    # Normal Light Profiles Only
 
-    plane = ag.Plane(redshift=0.5, galaxies=[g0, g1])
+    g0 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
+    g1 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=2.0))
+    g2 = ag.Galaxy(
+        redshift=0.5,
+        light_profile_0=ag.lp.EllSersic(intensity=1.0),
+        light_profile_1=ag.lp.EllSersic(intensity=2.0),
+    )
+
+    plane = ag.Plane(redshift=0.5, galaxies=[g0, g1, g2])
 
     fit = ag.FitInterferometer(
         dataset=interferometer_7,
@@ -349,26 +353,41 @@ def test___galaxy_model_visibilities_dict(interferometer_7):
         grid=interferometer_7.grid, transformer=interferometer_7.transformer
     )
 
-    assert fit.galaxy_model_visibilities_dict[g0].slim == pytest.approx(
+    assert fit.galaxy_model_visibilities_dict[g0] == pytest.approx(
         g0_visibilities, 1.0e-4
     )
-    assert fit.galaxy_model_visibilities_dict[g1].slim == pytest.approx(
+    assert fit.galaxy_model_visibilities_dict[g1] == pytest.approx(
         g1_visibilities, 1.0e-4
     )
-
-    assert fit.model_visibilities.slim == pytest.approx(
[…]
