### 2022-05-30 b91b0b1
extend simualter and fit interferometer method with galaxy dicts
