+++ test_autogalaxy/interferometer/test_simulate_and_fit_interferometer.py
@@ -213,3 +213,17 @@ def test__linear_light_profiles_agree_with_standard_light_profiles():
         galaxy_linear.disk
     ] == pytest.approx(0.2, 1.0e-2)
     assert fit.log_likelihood == pytest.approx(fit_linear.log_likelihood, 1.0e-4)
+
+    galaxy_image = galaxy.image_2d_from(grid=interferometer.grid)
+
+    assert fit_linear.galaxy_model_image_dict[galaxy_linear] == pytest.approx(
+        galaxy_image, 1.0e-4
+    )
+
+    galaxy_visibilities = galaxy.visibilities_via_transformer_from(
+        grid=interferometer.grid, transformer=interferometer.transformer
+    )
+
+    assert fit_linear.galaxy_model_visibilities_dict[galaxy_linear] == pytest.approx(
+        galaxy_visibilities, 1.0e-4
+    )
