### 2022-05-30 03d6932
cleaned up fit imaging tests
