+++ autogalaxy/interferometer/fit_interferometer.py
@@ -160,9 +160,9 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
 
         return {**galaxy_model_visibilities_dict, **galaxy_linear_obj_data_dict}
 
-    def model_visibilities_of_galaxies(self):
+    def model_visibilities_of_galaxies_list(self):
 
-        model_visibilities_of_galaxies = self.plane.visibilities_list_via_transformer_from(
+        model_visibilities_of_galaxies_list = self.plane.visibilities_list_via_transformer_from(
             grid=self.interferometer.grid, transformer=self.interferometer.transformer
         )
 
@@ -170,11 +170,11 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
 
             if galaxy.has_pixelization:
 
-                model_visibilities_of_galaxies[
+                model_visibilities_of_galaxies_list[
                     galaxy_index
                 ] += self.inversion.mapped_reconstructed_image
 
-        return model_visibilities_of_galaxies
+        return model_visibilities_of_galaxies_list
 
     @property
     def total_mappers(self):
