### 2022-05-30 d89abd1
fixed galaxy_model_visibilities_dict
