+++ autogalaxy/galaxy/__init__.py
@@ -1,2 +1,0 @@
-from .galaxy import Galaxy
-from .galaxy import Redshift
