+++ autogalaxy/aggregator/imaging.py
@@ -32,7 +32,7 @@ def _imaging_from(fit: af.Fit, settings_imaging: Optional[aa.SettingsImaging] = 
         settings_imaging.sub_size_pixelized = settings_imaging.sub_size_inversion
 
     if hasattr(settings_imaging, "grid_inversion_class"):
-        settings_imaging.grid_pixelized_class = settings_imaging.grid_inversion_class
+        settings_imaging.grid_pixelized_class = settings_imaging.grid_pixelized_class
 
     imaging = aa.Imaging(
         image=data,
