+++ autogalaxy/analysis/analysis.py
@@ -285,10 +285,6 @@ class AnalysisDataset(Analysis):
            The input instance with images associated with galaxies where possible.
         """
 
-        import copy
-
-        instance = copy.copy(instance)
-
         if self.hyper_galaxy_image_path_dict is not None:
 
             for galaxy_path, galaxy in instance.path_instance_tuples_for_class(Galaxy):
