+++ autogalaxy/analysis/preloads.py
@@ -132,7 +132,7 @@ class Preloads(aa.Preloads):
 
         if preloads.mapper_list is not None:
             preloads.mapper_galaxy_dict = fit_0.plane.to_inversion.mapper_galaxy_dict_from(
-                grid=fit_0.dataset.grid_pixelized
+                grid=fit_0.dataset.grid_inversion
             )
 
         preloads.set_operated_mapping_matrix_with_preloads(fit_0=fit_0, fit_1=fit_1)
