+++ autogalaxy/analysis/result.py
@@ -2,8 +2,6 @@ import numpy as np
 from typing import Dict, Union
 
 from autoconf import conf
-from autoconf import cached_property
-
 import autofit as af
 import autoarray as aa
 
