### 2022-06-02 a3028bb
Cosmology util refactored in prep for modeling
