+++ test_autogalaxy/conftest.py
@@ -291,6 +291,11 @@ def make_plane_x2_gal_7x7():
     return fixtures.make_plane_x2_gal_7x7()
 
 
+@pytest.fixture(name="planck15")
+def make_planck15():
+    return fixtures.make_planck15()
+
+
 @pytest.fixture(name="hyper_model_image_7x7")
 def make_hyper_model_image_7x7():
     return fixtures.make_hyper_model_image_7x7()

[…]
