+++ autogalaxy/cosmology/lensing.py
@@ -5,14 +5,59 @@ import numpy as np
 
 
 class LensingCosmology(cosmo.FLRW):
+    """
+    Class containing specific functions for performing gravitational lensing cosmology calculations.
+
+    By inheriting from the astropy `cosmo.FLRW` class this provides many additional methods for performing cosmological
+    calculations.
+    """
+
     def arcsec_per_kpc_from(self, redshift: float) -> float:
+        """
+        Angular separation in arcsec corresponding to a proper kpc at redshift `z`.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
+        This function therefore returns only the value of the astropy function it wraps, omitting the units instance.
+
+        Parameters
+        ----------
+        redshift
+            Input redshift from which the angular separation is calculated at.
+        """
         return self.arcsec_per_kpc_proper(z=redshift).value
 
     def kpc_per_arcsec_from(self, redshift: float) -> float:
+        """
+        Separation in transverse proper kpc corresponding to an arcminute at redshift `z`.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
+        This function therefore returns only the value of the astropy function it wraps, omitting the units instance.
+
+        Parameters
+        ----------
+        redshift
+            Input redshift from which the transverse proper kpc value is calculated at.
+        """
         return 1.0 / self.arcsec_per_kpc_proper(z=redshift).value
 
     def angular_diameter_distance_to_earth_in_kpc_from(self, redshift: float) -> float:
+        """
+        Angular diameter distance from the input `redshift` to redshift zero (e.g. us, the observer on earth) in
+        kiloparsecs.
 
+        This gives the proper (sometimes called 'physical') transverse distance corresponding to an angle of 1 radian
+        for an object at redshift `z`.
+
+        Weinberg, 1972, pp 421-424; Weedman, 1986, pp 65-67; Peebles, 1993, pp 325-327.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
+        This function therefore returns only the value of the astropy function it wraps, omitting the units instance.
+
+        Parameters
+        ----------
+        redshift
+            Input redshift from which the angular diameter distance to Earth is calculated.
+        """
         angular_diameter_distance_kpc = self.angular_diameter_distance(z=redshift).to(
             "kpc"
         )
@@ -21,7 +66,20 @@ class LensingCosmology(cosmo.FLRW):
 
     def angular_diameter_distance_between_redshifts_in_kpc_from(
         self, redshift_0: float, redshift_1: float
-    ):
+    ) -> float:
+        """
+        Angular diameter distance from an input `redshift_0` to another input `redshift_1`.
+
+        For simplicity, **PyAutoLens** internally uses only certain units to perform lensing cosmology calculations.
+        This function therefore returns only the value of the astropy function it wraps, omitting the units instance.
+
+        Parameters
+        ----------
+        redshift_0
+            Redshift from which the angular diameter distance to the other redshift is calculated.
+        redshift_1
+            Redshift from which the angular diameter distance to the other redshift is calculated.
+        """
         angular_diameter_distance_between_redshifts_kpc = self.angular_diameter_distance_z1z2(
             redshift_0, redshift_1
         ).to(
@@ -30,7 +88,18 @@ class LensingCosmology(cosmo.FLRW):
 
         return angular_diameter_distance_between_redshifts_kpc.value
 
-    def cosmic_average_density_from(self, redshift: float):
+    def cosmic_average_density_from(self, redshift: float) -> float:
+        """
[…]
