+++ autogalaxy/cosmology/__init__.py
@@ -2,3 +2,4 @@ from .lensing import LensingCosmology
 from .wrap import LambdaCDM
 from .wrap import FlatLambdaCDM
 from .wrap import Planck15
+from . import model
