### 2022-06-06 fe3e6e0
set none rather than delete hyper attributes
