+++ autogalaxy/analysis/analysis.py
@@ -285,12 +285,13 @@ class AnalysisDataset(Analysis):
            The input instance with images associated with galaxies where possible.
         """
 
-        instance = copy.copy(instance)
+        # instance = copy.deepcopy(instance)
 
         if self.hyper_galaxy_image_path_dict is not None:
 
             for galaxy_path, galaxy in instance.path_instance_tuples_for_class(Galaxy):
                 if galaxy_path in self.hyper_galaxy_image_path_dict:
+
                     galaxy.hyper_model_image = self.hyper_model_image
 
                     galaxy.hyper_galaxy_image = self.hyper_galaxy_image_path_dict[
