### 2022-06-06 1853095
replace instance reference so that function would be allowed to return a new instance rather than rely on side effects
