+++ autogalaxy/galaxy/galaxy.py
@@ -36,12 +36,12 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
     """
 
     def __init__(
-            self,
-            redshift: float,
-            pixelization: Optional[AbstractPixelization] = None,
-            regularization: Optional[AbstractRegularization] = None,
-            hyper_galaxy: Optional["HyperGalaxy"] = None,
-            **kwargs,
+        self,
+        redshift: float,
+        pixelization: Optional[AbstractPixelization] = None,
+        regularization: Optional[AbstractRegularization] = None,
+        hyper_galaxy: Optional["HyperGalaxy"] = None,
+        **kwargs,
     ):
         """
         Class representing a galaxy, which is composed of attributes used for fitting hyper_galaxies (e.g. light profiles, \
@@ -597,10 +597,10 @@ class HyperGalaxy:
     _ids = count()
 
     def __init__(
-            self,
-            contribution_factor: float = 0.0,
-            noise_factor: float = 0.0,
-            noise_power: float = 1.0,
+        self,
+        contribution_factor: float = 0.0,
+        noise_factor: float = 0.0,
+        noise_power: float = 1.0,
     ):
         """
         If a `Galaxy` is given a *HyperGalaxy* as an attribute, the noise-map in \
@@ -637,9 +637,9 @@ class HyperGalaxy:
     def __eq__(self, other):
         if isinstance(other, HyperGalaxy):
             return (
-                    self.contribution_factor == other.contribution_factor
-                    and self.noise_factor == other.noise_factor
-                    and self.noise_power == other.noise_power
+                self.contribution_factor == other.contribution_factor
+                and self.noise_factor == other.noise_factor
+                and self.noise_power == other.noise_power
             )
         return False
 
@@ -675,10 +675,10 @@ class HyperGalaxy:
             raise
 
     def hyper_noise_map_via_hyper_images_from(
-            self,
-            hyper_model_image: aa.Array2D,
-            hyper_galaxy_image: aa.Array2D,
-            noise_map: aa.Array2D,
+        self,
+        hyper_model_image: aa.Array2D,
+        hyper_galaxy_image: aa.Array2D,
+        noise_map: aa.Array2D,
     ) -> aa.Array2D:
         contribution_map = self.contribution_map_from(
             hyper_model_image=hyper_model_image, hyper_galaxy_image=hyper_galaxy_image
@@ -689,7 +689,7 @@ class HyperGalaxy:
         )
 
     def hyper_noise_map_from(
-            self, noise_map: aa.Array2D, contribution_map: aa.Array2D
+        self, noise_map: aa.Array2D, contribution_map: aa.Array2D
     ) -> aa.Array2D:
         """
         Returns a hyper galaxy hyper_galaxies noise-map from a baseline noise-map.
