### 2022-06-07 02f54e7
Dont log exception
