### 2022-06-08 6aedafc
analysis does not contain search in model util
