### 2022-06-14 #4475 @editorialbot
[REVIEW]: PyAutoGalaxy: Open-Source Multiwavelength Galaxy Structure & Morphology
**Submitting author:** @Jammy2211 (<a href="http://orcid.org/0000-0002-8987-7401">James Nightingale</a>)
**Repository:** https://github.com/Jammy2211/PyAutoGalaxy
**Branch with paper.md** (empty if default branch):
**Version:** 2023.1.15.1
