+++ docs/overview/overview_3_modeling.rst
@@ -60,6 +60,40 @@ In this example we therefore fit a model where:
 
 The redshifts of the galaxy (z=0.5) is fixed.
 
+Printing the ``info`` attribute of the model shows us this is the model we are fitting, and shows us the free
+parameters and their priors:
+
+.. code-block:: python
+
+    print(model.info)
+
+This gives the following output:
+
+.. code-block:: bash
+
+    galaxies
+        galaxy
+            redshift                                   0.5
+            bulge
+                centre
+                    centre_0                           GaussianPrior, mean = 0.0, sigma = 0.3
+                    centre_1                           GaussianPrior, mean = 0.0, sigma = 0.3
+                elliptical_comps
+                    elliptical_comps_0                 GaussianPrior, mean = 0.0, sigma = 0.5
+                    elliptical_comps_1                 GaussianPrior, mean = 0.0, sigma = 0.5
+                intensity                              LogUniformPrior, lower_limit = 1e-06, upper_limit = 1000000.0
+                effective_radius                       UniformPrior, lower_limit = 0.0, upper_limit = 30.0
+                sersic_index                           UniformPrior, lower_limit = 0.8, upper_limit = 5.0
+            disk
+                centre
+                    centre_0                           GaussianPrior, mean = 0.0, sigma = 0.3
+                    centre_1                           GaussianPrior, mean = 0.0, sigma = 0.3
+                elliptical_comps
+                    elliptical_comps_0                 GaussianPrior, mean = 0.0, sigma = 0.5
+                    elliptical_comps_1                 GaussianPrior, mean = 0.0, sigma = 0.5
+                intensity                              LogUniformPrior, lower_limit = 1e-06, upper_limit = 1000000.0
+                effective_radius                       UniformPrior, lower_limit = 0.0, upper_limit = 30.0
+
 Non-linear Search
 -----------------
 
@@ -120,6 +154,106 @@ by the search so far.
 
 The fit above returns a ``Result`` object, which includes lots of information on the model. 
 
+The ``info`` attribute can be printed to give the results in a readable format:
+
+.. code-block:: python
+
+    print(result_list.info)
+
+This gives the following output:
+
+.. code-block:: bash
+
+    Bayesian Evidence                              4910.81446407
+    Maximum Log Likelihood                         5010.64422962
+    Maximum Log Posterior                          975179.18825227
+    
+    model                                          CollectionPriorModel (N=13)
+        galaxies                                   CollectionPriorModel (N=13)
+            galaxy                                 Galaxy (N=13)
+                bulge                              EllSersic (N=7)
+                disk                               EllExponential (N=6)
+    
+    Maximum Log Likelihood Model:
+    
+    galaxies
+        galaxy
+            bulge
+                centre
+                    centre_0                       -0.002
+                    centre_1                       0.001
+                elliptical_comps
+                    elliptical_comps_0             0.056
+                    elliptical_comps_1             -0.009
+                intensity                          0.757
+                effective_radius                   0.708
+                sersic_index                       3.554
+            disk
+                centre
+                    centre_0                       0.001
+                    centre_1                       -0.004
+                elliptical_comps
+                    elliptical_comps_0             0.155
+                    elliptical_comps_1             0.091
+                intensity                          0.500
+                effective_radius                   1.554
+    
+    
+    Summary (3.0 sigma limits):
+    
+    galaxies
+        galaxy
+            bulge
+                centre
+                    centre_0                       -0.0028 (-0.0051, 0.0005)
+                    centre_1                       0.0014 (-0.0013, 0.0038)
[…]
