### 2022-06-24 b0a63a9
added operated light profile and method to galaxy
