+++ test_autogalaxy/plane/test_plane.py
@@ -386,6 +386,31 @@ def test__image_2d_list_from(sub_grid_2d_7x7):
     assert image_of_galaxies[1].binned[1] == lp1_image_pixel_1
 
 
+def test__image_2d_operated_from(sub_grid_2d_7x7):
+
+    light = ag.lp.EllSersic(intensity=1.0)
+    light_operated = ag.lp_operated.EllGaussian(intensity=1.0)
+
+    galaxy = ag.Galaxy(redshift=0.5, light=light, light_operated=light_operated)
+    g1 = ag.Galaxy(redshift=0.5, light_operated=light_operated)
+
+    plane = ag.Plane(galaxies=[galaxy])
+
+    image_2d_not_operated = plane.image_2d_not_operated_from(grid=sub_grid_2d_7x7)
+
+    light_image_2d = light.image_2d_from(grid=sub_grid_2d_7x7)
+
+    assert (image_2d_not_operated == light_image_2d).all()
+
+    image_2d = plane.image_2d_from(grid=sub_grid_2d_7x7)
+
+    light_operated_image_2d = 2.0 * light_operated.image_2d_from(grid=sub_grid_2d_7x7)
+
+    image_2d_via_light = light_image_2d + light_operated_image_2d
+
+    assert (image_2d == image_2d_via_light).all()
+
+
 def test__galaxy_image_2d_dict_from(sub_grid_2d_7x7):
 
     g0 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
