### 2022-07-04 baecf02
operated light profiles for PSF convolution
