+++ docs/requirements.txt
@@ -5,7 +5,7 @@ matplotlib>=3.0.3
 numpy>=1.18.0
 pyprojroot==0.2.0
 scikit-learn==0.21.3
-autoconf==2022.05.02.1
+autoconf==2022.06.29.1
 autofit==2022.05.02.1
 autoarray==2022.05.02.1
 sphinx==4.0.0
