### 2022-07-05 1d736d9
implementing bool in image
