+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -7,6 +7,7 @@ from autogalaxy.profiles.light_profiles import light_profiles_linear as lp_linea
 
 from autogalaxy import exc
 
+
 class LightProfileOperated:
 
     pass
@@ -20,6 +21,7 @@ class EllGaussian(lp.EllGaussian, LightProfileOperated):
         raise exc.ProfileException(
             "Cannot call `image_2d_not_operated_from() method for a LightProfileOperated object."
         )
+
 
 class EllGaussianLinear(lp_linear.EllGaussian, LightProfileOperated):
 
