### 2022-07-05 ff2a831
only_operated input works in Galaxy
