+++ autogalaxy/profiles/light_profiles/light_profiles.py
@@ -38,10 +38,6 @@ class LightProfile(EllProfile, OperateImage):
         super().__init__(centre=centre, elliptical_comps=elliptical_comps)
         self.intensity = intensity
 
-    @property
-    def is_operated(self):
-        return False
-
     def image_2d_from(self, grid: aa.type.Grid2DLike) -> aa.Array2D:
         """
         Returns the light profile's 2D image from a 2D grid of Cartesian (y,x) coordinates, which may have been
@@ -80,7 +76,6 @@ class LightProfile(EllProfile, OperateImage):
         -------
 
         """
-
 
         return self.image_2d_from(grid=grid)
 
