+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -7,19 +7,18 @@ from autogalaxy.profiles.light_profiles import light_profiles_linear as lp_linea
 
 from autogalaxy import exc
 
+
 class LightProfileOperated:
 
     pass
 
 
 class EllGaussian(lp.EllGaussian, LightProfileOperated):
-
-    pass
-
     def image_2d_not_operated_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
         raise exc.ProfileException(
             "Cannot call `image_2d_not_operated_from() method for a LightProfileOperated object."
         )
+
 
 class EllGaussianLinear(lp_linear.EllGaussian, LightProfileOperated):
 
