+++ autogalaxy/galaxy/galaxy.py
@@ -219,11 +219,17 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
         )
 
     @aa.grid_dec.grid_2d_to_structure
-    def image_2d_from(self, grid: aa.type.Grid2DLike, operated_only:Optional[bool]) -> Union[np.ndarray, aa.Array2D]:
+    def image_2d_from(
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+    ) -> Union[np.ndarray, aa.Array2D]:
         """
         Returns the summed 2D image of the galaxy's light profiles from a 2D grid of Cartesian (y,x) coordinates.
 
         If the galaxy has no light profiles, a numpy array of zeros is returned.
+
+        If the `operated_only` input is included, the function omits light profiles which are parents of
+        the `LightProfileOperated` object, which signifies that the light profile represents emission that has
+        already had the instrument operations (e.g. PSF convolution, a Fourier transform) applied to it.
 
         See the `autogalaxy.profiles.light_profiles` package for details of how images are computed from a light 
         profile. 
@@ -235,13 +241,19 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
         ----------
         grid
             The 2D (y, x) coordinates where values of the image are evaluated.
+        operated_only
+            By default, the image is the sum of light profile images (irrespective of whether they have been operatd on
+            or not). If this input is included as a bool, only images which are or are not already operated are summed
+            and returned.
         """
         if self.has_light_profile:
             return sum(self.image_2d_list_from(grid=grid, operated_only=operated_only))
 
         return np.zeros((grid.shape[0],))
 
-    def image_2d_list_from(self, grid: aa.type.Grid2DLike, operated_only:Optional[bool]) -> List[aa.Array2D]:
+    def image_2d_list_from(
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+    ) -> List[aa.Array2D]:
         """
         Returns a list of the 2D images of the galaxy's light profiles from a 2D grid of Cartesian (y,x) coordinates.
 
@@ -250,26 +262,35 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
         If the galaxy has no light profiles, a numpy array of zeros is returned.
 
+        If the `operated_only` input is included, the function omits light profiles which are parents of
+        the `LightProfileOperated` object, which signifies that the light profile represents emission that has
+        already had the instrument operations (e.g. PSF convolution, a Fourier transform) applied to it.
+
         See the `autogalaxy.profiles.light_profiles` package for details of how images are computed from a light
         profile.
-
 
         Parameters
         ----------
         grid
             The 2D (y, x) coordinates where values of the image are evaluated.
+        operated_only
+            By default, the returnd list contains all light profile images (irrespective of whether they have been
+            operated on or not). If this input is included as a bool, only images which are or are not already
+            operated are included in the list, with the images of other light profiles created as a numpy array of
+            zeros.
         """
         if operated_only is None:
             return [
                 light_profile.image_2d_from(grid=grid)
                 for light_profile in self.light_profile_list
             ]
+
         return [
-                    light_profile.image_2d_from(grid=grid)
-                    if light_profile.is_operated is operated_only
-                    else np.zeros((grid.shape[0],))
-                    for light_profile in self.light_profile_list
-                ]
+            light_profile.image_2d_from(grid=grid)
+            if isinstance(light_profile, LightProfileOperated) is operated_only
+            else np.zeros((grid.shape[0],))
[…]
