### 2022-07-06 da51c90
operated_only added to plane
