### 2022-07-06 70b0363
light profiles now have decorator to check if operated
