+++ autogalaxy/profiles/light_profiles/light_profiles.py
@@ -1,9 +1,10 @@
 import numpy as np
 from scipy.integrate import quad
-from typing import Tuple, Union
+from typing import Optional, Tuple, Union
 
 import autoarray as aa
 
+from autogalaxy.profiles.light_profiles.light_profile_decorators import check_operated_only
 from autogalaxy.operate.image import OperateImage
 from autogalaxy.profiles.geometry_profiles import EllProfile
 
@@ -38,7 +39,11 @@ class LightProfile(EllProfile, OperateImage):
         super().__init__(centre=centre, elliptical_comps=elliptical_comps)
         self.intensity = intensity
 
-    def image_2d_from(self, grid: aa.type.Grid2DLike) -> aa.Array2D:
+    @property
+    def is_operated(self):
+        return False
+
+    def image_2d_from(self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None) -> aa.Array2D:
         """
         Returns the light profile's 2D image from a 2D grid of Cartesian (y,x) coordinates, which may have been
         transformed using the light profile's geometry.
@@ -56,28 +61,6 @@ class LightProfile(EllProfile, OperateImage):
             The image of the `LightProfile` evaluated at every (y,x) coordinate on the transformed grid.
         """
         raise NotImplementedError()
-
-    def image_2d_not_operated_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
-        """
-        The `LightProfileOperated` objects in the `light_profiles_operated.py` package represent light profiles
-        who have already had instrumental operations applied to them.
-
-        For example, they may already been convolved with imaging dataset's PSF or Fourier transformed by the
-        transformer of an interferometer.
-
-        When the images of an object are used for operation (E.g. in the `LightProfileOperated` object), this function
-        is called in order to create the image from only light profiles that have not already been operated.
-
-        Parameters
-        ----------
-        grid
-
-        Returns
-        -------
-
-        """
-
-        return self.image_2d_from(grid=grid)
 
     def image_2d_via_radii_from(self, grid_radii: np.ndarray) -> np.ndarray:
         """
@@ -208,9 +191,10 @@ class EllGaussian(LightProfile):
         )
 
     @aa.grid_dec.grid_2d_to_structure
+    @check_operated_only
     @aa.grid_dec.transform
     @aa.grid_dec.relocate_to_radial_minimum
-    def image_2d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
+    def image_2d_from(self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None) -> np.ndarray:
         """
         Returns the Gaussian light profile's 2D image from a 2D grid of Cartesian (y,x) coordinates.
 
