### 2022-07-06 a58a668
galaxy ponly_operated simplified thanks to light profile wrapper
