+++ autogalaxy/galaxy/galaxy.py
@@ -279,52 +279,10 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
             operated are included in the list, with the images of other light profiles created as a numpy array of
             zeros.
         """
-        if operated_only is None:
-            return [
-                light_profile.image_2d_from(grid=grid)
-                for light_profile in self.light_profile_list
-            ]
-
         return [
-            light_profile.image_2d_from(grid=grid)
-            if isinstance(light_profile, LightProfileOperated) is operated_only
-            else np.zeros((grid.shape[0],))
+            light_profile.image_2d_from(grid=grid, operated_only=operated_only)
             for light_profile in self.light_profile_list
         ]
-
-    @aa.grid_dec.grid_2d_to_structure
-    def image_2d_not_operated_from(self, grid: aa.type.Grid2DLike) -> aa.Array2D:
-        """
-        Returns the summed 2D image of the galaxy's light profiles from a 2D grid of Cartesian (y,x) coordinates.
-
-        This function omits light profiles which are parents of the `LightProfileOperated` object, which signifies
-        that the light profile represents emission that has already had the instrument operations (e.g. PSF
-        convolution, a Fourier transform) applied to it.
-
-        If the galaxy has no light profiles, a numpy array of zeros is returned.
-
-        See the `autogalaxy.profiles.light_profiles` package for details of how images are computed from a light
-        profile.
-
-        The decorator `grid_2d_to_structure` converts the output arrays from ndarrays to an `Array2D` data structure
-        using the input `grid`'s attributes.
-
-        Parameters
-        ----------
-        grid
-            The 2D (y, x) coordinates where values of the image are evaluated.
-        """
-        if self.has_light_profile:
-            return sum(
-                [
-                    light_profile.image_2d_from(grid=grid)
-                    if not isinstance(light_profile, LightProfileOperated)
-                    else np.zeros((grid.shape[0],))
-                    for light_profile in self.light_profile_list
-                ]
-            )
-
-        return np.zeros((grid.shape[0],))
 
     @aa.grid_dec.grid_1d_output_structure
     def image_1d_from(self, grid: aa.type.Grid2DLike) -> np.ndarray:
