### 2022-07-06 e11979a
operator now splits operated and unoperated images
