+++ autogalaxy/profiles/light_profiles/light_profiles_snr.py
@@ -30,7 +30,9 @@ class LightProfileSNR:
         """
         self.signal_to_noise_ratio = signal_to_noise_ratio
 
-    def image_2d_from(self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None) -> aa.Array2D:
+    def image_2d_from(
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+    ) -> aa.Array2D:
         """
         Abstract method for obtaining intensity at a grid of Cartesian (y,x) coordinates.
 
