### 2022-07-06 125ba50
operate image works to split operated and not operated images
