+++ autogalaxy/profiles/light_profiles/mock/mock_light_profile.py
@@ -1,4 +1,5 @@
 import numpy as np
+from typing import Optional
 
 import autogalaxy as ag
 
@@ -24,7 +25,7 @@ class MockLightProfile(ag.lp.LightProfile):
         self.value = value
         self.value1 = value1
 
-    def image_2d_from(self, grid):
+    def image_2d_from(self, grid, operated_only: Optional[bool] = None):
 
         if self.image_2d is not None:
             return self.image_2d
