### 2022-07-06 40c080f
added check operated decorator to MOkcLightProfile to fix tests.
