+++ autogalaxy/profiles/light_profiles/mock/mock_light_profile.py
@@ -1,6 +1,9 @@
 import numpy as np
 from typing import Optional
 
+from autogalaxy.profiles.light_profiles.light_profile_decorators import (
+    check_operated_only,
+)
 import autogalaxy as ag
 
 # MockProfiles #
@@ -25,6 +28,8 @@ class MockLightProfile(ag.lp.LightProfile):
         self.value = value
         self.value1 = value1
 
+    #
+    @check_operated_only
     def image_2d_from(self, grid, operated_only: Optional[bool] = None):
 
         if self.image_2d is not None:
