### 2022-07-06 a56af36
operate image gives correct blurred image 2d list
