### 2022-07-06 7e61f4f
merged via psf and convolver methods
