### 2022-07-06 db80e7c
renamed visibiiltiies_via_transformer_from
