### 2022-07-06 bb6b509
added operated only to galaxy_image_2d_dict
