### 2022-07-06 71e6241
added has_light_profile_operated
