### 2022-07-06 5915b85
added has_light_profile_operated to plane
