### 2022-07-06 4a23c3a
checks if operated light profile exists for speed up
