### 2022-07-06 d6ba0fc
test added on fit_imaging
