+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -1,5 +1,4 @@
 from autogalaxy.profiles.light_profiles import light_profiles as lp
-from autogalaxy.profiles.light_profiles import light_profiles_linear as lp_linear
 
 # TODO : Would rather remove `is_operated` and use `isinstance_` when checking in decorator, but currently get import error.
 
@@ -15,7 +14,7 @@ class EllGaussian(lp.EllGaussian, LightProfileOperated):
         return True
 
 
-class EllGaussianLinear(lp_linear.EllGaussian, LightProfileOperated):
+class EllSersic(lp.EllSersic, LightProfileOperated):
     @property
     def is_operated(self):
         return True
