+++ autogalaxy/profiles/light_profiles/light_profiles_linear_operated.py
@@ -2,7 +2,7 @@ from autogalaxy.profiles.light_profiles import light_profiles_operated as lp_ope
 from autogalaxy.profiles.light_profiles import light_profiles_linear as lp_linear
 
 
-class EllGaussianLinear(lp_linear.EllGaussian, lp_operated.LightProfileOperated):
+class EllGaussian(lp_linear.EllGaussian, lp_operated.LightProfileOperated):
     @property
     def is_operated(self):
         return True
