### 2022-07-08 774ff33
use has method over checking length of list
