+++ autogalaxy/galaxy/galaxy.py
@@ -153,7 +153,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
     @property
     def has_profile(self) -> bool:
-        return len(self.mass_profile_list) + len(self.light_profile_list) > 0
+        return self.has_mass_profile or self.has_light_profile
 
     @property
     def light_profile_list(self) -> List[LightProfile]:
@@ -165,7 +165,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
     @property
     def has_light_profile(self) -> bool:
-        return len(self.light_profile_list) > 0
+        return self.has(LightProfile)
 
     @property
     def light_profile_linear_list(self) -> List[LightProfileLinear]:
@@ -177,7 +177,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
     @property
     def has_light_profile_linear(self) -> bool:
-        return len(self.light_profile_linear_list) > 0
+        return self.has(LightProfileLinear)
 
     def radial_projected_shape_slim_from(self, grid: aa.type.Grid1D2DLike) -> int:
         """
@@ -298,7 +298,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
     @property
     def has_mass_profile(self) -> bool:
-        return len(self.mass_profile_list) > 0
+        return self.has(MassProfile)
 
     @aa.grid_dec.grid_2d_to_vector_yx
     @aa.grid_dec.grid_2d_to_structure
