### 2022-07-09 e7980be
simplify if statement in light profile operated
