### 2022-07-09 91b7f6b
simplfy typing of grid
