### 2022-07-09 952b87f
use isistaince in light profile decorator
