+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -1,7 +1,5 @@
 from autogalaxy.profiles.light_profiles import light_profiles as lp
 
-# TODO : Would rather remove `is_operated` and use `isinstance_` when checking in decorator, but currently get import error.
-
 
 class LightProfileOperated:
 
@@ -9,12 +7,10 @@ class LightProfileOperated:
 
 
 class EllGaussian(lp.EllGaussian, LightProfileOperated):
-    @property
-    def is_operated(self):
-        return True
+
+    pass
 
 
 class EllSersic(lp.EllSersic, LightProfileOperated):
-    @property
-    def is_operated(self):
-        return True
+
+    pass
