+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -66,17 +66,12 @@ def test__image_2d_from(sub_grid_2d_7x7, gal_x2_lp):
     assert gal_image.binned[1] == lp_image_1
 
 
-def test__image_2d_from__operated_only_input(sub_grid_2d_7x7):
+def test__image_2d_from__operated_only_input(sub_grid_2d_7x7, lp_0, lp_operated_0):
 
-    light_not_operated = ag.lp.EllSersic(intensity=1.0)
-    light_operated = ag.lp_operated.EllGaussian(intensity=1.0)
+    image_2d_not_operated = lp_0.image_2d_from(grid=sub_grid_2d_7x7)
+    image_2d_operated = lp_operated_0.image_2d_from(grid=sub_grid_2d_7x7)
 
-    image_2d_not_operated = light_not_operated.image_2d_from(grid=sub_grid_2d_7x7)
-    image_2d_operated = light_operated.image_2d_from(grid=sub_grid_2d_7x7)
-
-    galaxy = ag.Galaxy(
-        redshift=0.5, light=light_not_operated, light_operated=light_operated
-    )
+    galaxy = ag.Galaxy(redshift=0.5, light=lp_0, light_operated=lp_operated_0)
 
     image_2d = galaxy.image_2d_from(grid=sub_grid_2d_7x7, operated_only=False)
     assert (image_2d == image_2d_not_operated).all()
@@ -86,6 +81,14 @@ def test__image_2d_from__operated_only_input(sub_grid_2d_7x7):
 
     image_2d = galaxy.image_2d_from(grid=sub_grid_2d_7x7, operated_only=None)
     assert (image_2d == image_2d_not_operated + image_2d_operated).all()
+
+
+def test__image_2d_list_from__operated_only_input(sub_grid_2d_7x7, lp_0, lp_operated_0):
+
+    image_2d_not_operated = lp_0.image_2d_from(grid=sub_grid_2d_7x7)
+    image_2d_operated = lp_operated_0.image_2d_from(grid=sub_grid_2d_7x7)
+
+    galaxy = ag.Galaxy(redshift=0.5, light=lp_0, light_operated=lp_operated_0)
 
     image_2d_list = galaxy.image_2d_list_from(grid=sub_grid_2d_7x7, operated_only=False)
     assert (image_2d_list[0] == image_2d_not_operated).all()
