### 2022-07-09 3f5c142
removed has_light_profile
