### 2022-07-09 d97aa19
removed has_mass_profile
