### 2022-07-09 32a10ac
removed has pixelizaiton and has_regularizaiton
