+++ autogalaxy/imaging/plot/fit_imaging_plotters.py
@@ -1,6 +1,7 @@
 import numpy as np
 from typing import List, Optional
 
+import autoarray as aa
 import autoarray.plot as aplt
 
 from autoarray.fit.plot.fit_imaging_plotters import FitImagingPlotterMeta
@@ -201,7 +202,7 @@ class FitImagingPlotter(Plotter):
             )
             self.figures_2d_of_galaxies(galaxy_index=galaxy_index, model_image=True)
 
-            if self.plane.has_pixelization:
+            if self.plane.has(cls=aa.pix.Pixelization):
                 self.inversion_plotter.figures_2d_of_mapper(
                     mapper_index=0, reconstruction=True
                 )
