+++ autogalaxy/interferometer/fit_interferometer.py
@@ -90,7 +90,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
 
         The image passed to this function is the dataset's image with all light profile images of the plane subtracted.
         """
-        if self.plane.has_pixelization or self.plane.has_light_profile_linear:
+        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has_light_profile_linear:
 
             if self.settings_inversion.use_w_tilde:
                 w_tilde = self.dataset.w_tilde
@@ -120,7 +120,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
         If a inversion is included it is the sum of this sum and the inversion's reconstruction of the image.
         """
 
-        if self.plane.has_pixelization or self.plane.has_light_profile_linear:
+        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has_light_profile_linear:
 
             return self.profile_visibilities + self.inversion.mapped_reconstructed_data
 
