### 2022-07-09 2ec15be
revoed has_linear_light_profile / has_light_profile_operated
