+++ autogalaxy/interferometer/fit_interferometer.py
@@ -10,6 +10,7 @@ from autogalaxy.galaxy.galaxy import Galaxy
 from autogalaxy.hyper.hyper_data import HyperBackgroundNoise
 from autogalaxy.plane.plane import Plane
 from autogalaxy.plane.to_inversion import PlaneToInversion
+from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
 
 
 class FitInterferometer(aa.FitInterferometer, AbstractFit):
@@ -90,7 +91,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
 
         The image passed to this function is the dataset's image with all light profile images of the plane subtracted.
         """
-        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has_light_profile_linear:
+        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has(cls=LightProfileLinear):
 
             if self.settings_inversion.use_w_tilde:
                 w_tilde = self.dataset.w_tilde
@@ -120,7 +121,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
         If a inversion is included it is the sum of this sum and the inversion's reconstruction of the image.
         """
 
-        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has_light_profile_linear:
+        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has(cls=LightProfileLinear):
 
             return self.profile_visibilities + self.inversion.mapped_reconstructed_data
 

[…]
