### 2022-07-09 4a8077b
removed has_hyper_Galaxy
