### 2022-07-09 bc1ae2c
removed is_light_profile
