+++ test_autogalaxy/conftest.py
@@ -215,6 +215,10 @@ def make_lp_0():
 def make_lp_1():
     return fixtures.make_lp_1()
 
+@pytest.fixture(name="lp_linear_0")
+def make_lp_linear_0():
+    return fixtures.make_lp_linear_0()
+
 
 @pytest.fixture(name="lp_operated_0")
 def make_lp_operated_0():
