### 2022-07-09 d9e0ef4
remove light profile linear and operated list
