+++ autogalaxy/plane/to_inversion.py
@@ -30,7 +30,7 @@ class PlaneToInversion:
 
         for galaxy in self.plane.galaxies:
             if galaxy.has(cls=LightProfileLinear):
-                for light_profile_linear in galaxy.light_profile_linear_list:
+                for light_profile_linear in galaxy.has(cls=al.lp_linear.LightProfileLinear):
 
                     lp_linear_func = LightProfileLinearObjFunc(
                         grid=source_grid_slim,
