### 2022-07-09 503a407
fixed tests due to removing light_profile_linear_list
