+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -1,7 +1,6 @@
 from autogalaxy.profiles.light_profiles import light_profiles as lp
 
 
-
 class LightProfileOperated:
 
     pass
@@ -10,6 +9,7 @@ class LightProfileOperated:
 class EllGaussian(lp.EllGaussian, LightProfileOperated):
 
     pass
+
 
 class EllSersic(lp.EllSersic, LightProfileOperated):
 
