### 2022-07-10 3d28b15
removed more functions from plane
