+++ autogalaxy/config/general.ini
@@ -19,6 +19,7 @@ iterations_per_update=5000
 ignore_prior_limits=False
 
 [numba]
+use_numba=True
 nopython=True
 cache=False
 parallel=False
