+++ autogalaxy/__init__.py
@@ -97,4 +97,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2022.05.02.1"
+__version__ = "2022.07.10.1"
