+++ docs/requirements.txt
@@ -1,13 +1,14 @@
+autoconf==2022.06.29.1
+autofit==2022.07.10.1
+autoarray==2022.07.10.1
 astropy>=3.0.0
 colossus==1.2.14
 numpydoc>=1.0.0
 matplotlib>=3.0.3
+numba
 numpy>=1.18.0
 pyprojroot==0.2.0
 scikit-learn==0.21.3
-autoconf==2022.06.29.1
-autofit==2022.05.02.1
-autoarray==2022.05.02.1
 sphinx==4.0.0
 furo
 myst-parser
