+++ docs/installation/conda.rst
@@ -41,7 +41,7 @@ the installation has clean dependencies):
 
 .. code-block:: bash
 
-    pip install autogalaxy==2022.05.02.1
+    pip install autogalaxy==2022.07.10.1
 
 You may get warnings which state something like:
 
