+++ docs/installation/pip.rst
@@ -37,6 +37,26 @@ If there are no errors (but only the warnings above) **PyAutoGalaxy** is install
 
 If there is an error check out the `troubleshooting section <https://pyautogalaxy.readthedocs.io/en/latest/installation/troubleshooting.html>`_.
 
+Numba
+-----
+
+Numba (https://numba.pydata.org)  is an optional library which makes **PyAutoGalaxy** run a lot faster, which we
+strongly recommend users have installed.
+
+You can install numba via the following command:
+
+.. code-block:: bash
+
+    pip install numba
+
+Some users have experienced difficulties installing numba, which is why it is an optional library. If your
+installation is not successful, you can use **PyAutoGalaxy** without it installed for now, to familiarize yourself
+with the software and determine if it is the right software for you.
+
+If you decide that **PyAutoGalaxy** is the right software, then I recommend you commit the time to getting a
+successful numba install working, with more information provided `at this readthedocs page <https://pyautogalaxy.readthedocs.io/en/latest/installation/numba.html>`_
+
+
 Workspace
 ---------
 
