+++ docs/requirements.txt
@@ -1,6 +1,6 @@
-autoconf==2022.07.10.1
-autofit==2022.07.10.1
-autoarray==2022.07.10.1
+autoconf==2022.07.11.1
+autofit==2022.07.11.1
+autoarray==2022.07.11.1
 astropy>=3.0.0
 colossus==1.2.14
 numpydoc>=1.0.0
