+++ docs/installation/pip.rst
@@ -21,7 +21,7 @@ the installation has clean dependencies):
 
 .. code-block:: bash
 
-    pip install autogalaxy==2022.07.10.1
+    pip install autogalaxy==2022.07.11.1
 
 You may get warnings which state something like:
 
