+++ autogalaxy/config/visualize/mat_wrap/Cmap.ini
@@ -1,15 +1,0 @@
-[figure]
-cmap=jet
-norm=linear
-vmin=None
-vmax=None
-linthresh=0.05
-linscale=0.01
-
-[subplot]
-cmap=jet
-norm=linear
-vmin=None
-vmax=None
-linthresh=0.05
-linscale=0.01
