### 2022-07-15 451b70b
implement MoffatLightProfile
