+++ autogalaxy/profiles/light_profiles/light_profiles.py
@@ -243,6 +243,133 @@ class SphGaussian(EllGaussian):
         )
 
 
+class EllMoffat(LightProfile):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        alpha: float = 0.5,
+        beta: float = 2.0,
+    ):
+        """
+        The elliptical Moffat light profile, which is commonly used to model the Point Spread Function of
+        Astronomy observations.
+
+        This form of the MOffat profile is a reparameterizaiton of the original formalism given by
+        https://ui.adsabs.harvard.edu/abs/1969A%26A.....3..455M/abstract. The actual profile itself is identical.
+
+        See `autogalaxy.profiles.light_profiles.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        intensity
+            Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
+            the light profile's image is compared too, which is expected to be electrons per second).
+        alpha
+            Scales the overall size of the Moffat profile and for a PSF typically corresponds to the FWHM / 2.
+        beta
+            Scales the wings at the outskirts of the Moffat profile, where smaller values imply heavier wings and it
+            tends to a Gaussian as beta goes to infinity.
+        """
+
+        super().__init__(
+            centre=centre, elliptical_comps=elliptical_comps, intensity=intensity
+        )
+        self.alpha = alpha
+        self.beta = beta
+
+    def image_2d_via_radii_from(self, grid_radii: np.ndarray) -> np.ndarray:
+        """
+        Returns the 2D image of the Moffat light profile from a grid of coordinates which are the radial distance of
+        each coordinate from the its `centre`.
+
+        Note: sigma is divided by sqrt(q) here.
+
+        Parameters
+        ----------
+        grid_radii
+            The radial distances from the centre of the profile, for each coordinate on the grid.
+        """
+        return np.multiply(
+            self.intensity,
+            np.power(
+                1 + np.square(
+                    np.divide(grid_radii, self.alpha / np.sqrt(self.axis_ratio))
+                ),
+                - self.beta
+            ),
+        )
+
+    @aa.grid_dec.grid_2d_to_structure
+    @check_operated_only
+    @aa.grid_dec.transform
+    @aa.grid_dec.relocate_to_radial_minimum
+    def image_2d_from(
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+    ) -> np.ndarray:
+        """
+        Returns the Moffat light profile's 2D image from a 2D grid of Cartesian (y,x) coordinates.
+
+        If the coordinates have not been transformed to the profile's geometry, this is performed automatically.
+
+        Parameters
+        ----------
+        grid
+            The 2D (y, x) coordinates in the original reference frame of the grid.
+
+        Returns
+        -------
+        image
+            The image of the Moffat evaluated at every (y,x) coordinate on the transformed grid.
+        """
+
+        return self.image_2d_via_radii_from(self.grid_to_eccentric_radii(grid))
+
+
+class SphMoffat(EllMoffat):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        alpha: float = 0.5,
+        beta: float = 2.0,
+    ):
+        """
+        The spherical Moffat light profile:
+
[…]
