### 2022-07-15 e6e5c2d
Moffat linear / operated
