+++ autogalaxy/profiles/light_profiles/light_profiles_linear_operated.py
@@ -5,3 +5,8 @@ from autogalaxy.profiles.light_profiles import light_profiles_linear as lp_linea
 class EllGaussian(lp_linear.EllGaussian, lp_operated.LightProfileOperated):
 
     pass
+
+
+class EllMoffat(lp_linear.EllMoffat, lp_operated.LightProfileOperated):
+
+    pass
