+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -11,6 +11,12 @@ class EllGaussian(lp.EllGaussian, LightProfileOperated):
     pass
 
 
+class EllMoffat(lp.EllMoffat, LightProfileOperated):
+
+    pass
+
+
+
 class EllSersic(lp.EllSersic, LightProfileOperated):
 
     pass
