### 2022-07-15 587cea2
now visualize linear light profiles using resulting plane
