+++ autogalaxy/profiles/light_profiles/light_profiles_operated.py
@@ -16,7 +16,6 @@ class EllMoffat(lp.EllMoffat, LightProfileOperated):
     pass
 
 
-
 class EllSersic(lp.EllSersic, LightProfileOperated):
 
     pass
