### 2022-07-15 9e6df82
fixes to visualize erros
