### 2022-07-15 c1788f5
fixed linear light profile parameter casting
