+++ autogalaxy/config/notation/label_format.ini
@@ -1,5 +1,6 @@
 [format]
 alpha={:.4f}
+beta={:.4f}
 angular_diameter_distance_to_earth={:.4f}
 centre_0={:.4f}
 centre_1={:.4f}
