+++ autogalaxy/config/grids/radial_minimum.ini
@@ -11,6 +11,8 @@ EllPowerLawCored=1e-8
 SphPowerLawCored=1e-8
 EllIsothermalCored=1e-8
 SphIsothermalCored=1e-8
+EllMoffat=1e-8
+SphMoffat=1e-8
 EllGaussian=1e-8
 SphGaussian=1e-8
 EllSersic=1e-6
