### 2022-07-15 1272a71
linea r operarted light profiles do not use blurred_mapping_matrix
