### 2022-07-15 6aeaead
max log likelihood plane uses plane from tracer to get linear light profile intensities
