+++ autogalaxy/imaging/model/result.py
@@ -3,6 +3,7 @@ from typing import List
 import autoarray as aa
 
 from autogalaxy.analysis.result import ResultDataset
+from autogalaxy.plane.plane import Plane
 from autogalaxy.imaging.fit_imaging import FitImaging
 
 
@@ -53,10 +54,28 @@ class ResultImaging(ResultDataset):
             instance=self.instance
         )
 
+        instance = self.analysis.instance_with_associated_hyper_images_from(
+            instance=self.instance
+        )
+
+        plane = self.analysis.plane_via_instance_from(instance=instance)
+
         return self.analysis.fit_imaging_via_plane_from(
-            plane=self.max_log_likelihood_plane,
+            plane=plane,
             hyper_image_sky=hyper_image_sky,
             hyper_background_noise=hyper_background_noise,
+        )
+
+    @property
+    def max_log_likelihood_plane(self) -> Plane:
+        """
+        An instance of a `Plane` corresponding to the maximum log likelihood model inferred by the non-linear search.
+
+        The `Plane` is computed from the `max_log_likelihood_fit`, as this ensures that all linear light profiles
+        are converted to normal light profiles with their `intensity` values updated.
+        """
+        return (
+            self.max_log_likelihood_fit.model_obj_linear_light_profiles_to_light_profiles
         )
 
     @property
