### 2022-07-15 628965c
unit test on linear light profiles of max plane
