### 2022-07-15 a3e7bc6
raise exception if linear light profile in plane plotter
