+++ autogalaxy/profiles/plot/light_profile_plotters.py
@@ -17,6 +17,7 @@ from autogalaxy.plot.mat_wrap.include import Include2D
 from autogalaxy.util import error_util
 from autogalaxy import exc
 
+
 class LightProfilePlotter(Plotter):
     def __init__(
         self,
@@ -62,13 +63,14 @@ class LightProfilePlotter(Plotter):
             Specifies which attributes of the `LightProfile` are extracted and plotted as visuals for 2D plots.
         """
 
-        from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
+        from autogalaxy.profiles.light_profiles.light_profiles_linear import (
+            LightProfileLinear,
+        )
 
         if isinstance(light_profile, LightProfileLinear):
 
             raise exc.raise_linear_light_profile_in_plot(
-                plotter_type=self.__class__.__name__,
-                model_obj="Plane"
+                plotter_type=self.__class__.__name__, model_obj="Plane"
             )
 
         self.light_profile = light_profile
