+++ autogalaxy/plane/plot/plane_plotters.py
@@ -62,13 +62,14 @@ class PlanePlotter(Plotter):
             Specifies which attributes of the `MassProfile` are extracted and plotted as visuals for 2D plots.
         """
 
-        from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
+        from autogalaxy.profiles.light_profiles.light_profiles_linear import (
+            LightProfileLinear,
+        )
 
         if plane.has(cls=LightProfileLinear):
 
             raise exc.raise_linear_light_profile_in_plot(
-                plotter_type=self.__class__.__name__,
-                model_obj="Plane"
+                plotter_type=self.__class__.__name__, model_obj="Plane"
             )
 
         super().__init__(
