### 2022-07-16 Jammy2211/PyAutoGalaxy#31 @alexandergagliano
Import error in test_autogalaxy/conftest.py
When running the unit tests using pytests I get the following error message:

ImportError while loading conftest '/Users/alexgagliano/Documents/AcademicService/JOSS/autogalaxy_modrepo/PyAutoGalaxy/test_autogalaxy/conftest.py'.
test_autogalaxy/conftest.py:8: in <module>
    from autogalaxy import fixtures
../../../../../miniconda3/lib/python3.9/site-packages/autogalaxy/__init__.py:1: in <module>
    from autoarray.dataset import preprocess
../../../../../miniconda3/lib/python3.9/site-packages/autoarray/__init__.py:2: in <module>
    from . import type
../../../../../miniconda3/lib/python3.9/site-packages/autoarray/type.py:28: in <module>
    from autoarray.dataset.interferometer import Visibilities
../../../../../miniconda3/lib/python3.9/site-packages/autoarray/dataset/interferometer.py:13: in <module>
    from autoarray.operators.transformer import TransformerDFT
../../../../../miniconda3/lib/python3.9/site-packages/autoarray/operators/transformer.py:22: in <module>
    import pylops
../../../../../miniconda3/lib/python3.9/site-packages/pylops/__init__.py:2: in <module>
    from .basicoperators import Regression
../../../../../miniconda3/lib/python3.9/site-packages/pylops/basicoperators/__init__.py:17: in <module>
    from .VStack import VStack
../../../../../miniconda3/lib/python3.9/site-packages/pylops/basicoperators/VStack.py:2: in <module>
    from scipy.sparse.linalg.interface import _get_dtype
E   ImportError: cannot import name '_get_dtype' from 'scipy.sparse.linalg.interface' (/Users/alexgagliano/miniconda3/lib/python3.9/site-packages/scipy/sparse/linalg/interface.py)

Python version is 3.9.12 and scipy version is 1.8.1. Should the code be installed with a different version? This is connected to the review at [#4475](https://github.com/openjournals/joss-reviews/issues/4475).
