### 2022-07-16 Jammy2211/PyAutoGalaxy#32 @alexandergagliano
Galaxies attribute in introduction.ipynb on Binder
When uncommenting the line `result = search.fit(model=model, analysis=analysis)` in the main `introduction.ipynb` notebook, I get the error: AttributeError: 'ModelInstance' object has no attribute 'galaxies'. Connected to [#4475](https://github.com/openjournals/joss-reviews/issues/4475).
