### 2022-07-16 Jammy2211/PyAutoGalaxy#33 @alexandergagliano
AttributeError: module 'autogalaxy' has no attribute 'SettingsGalaxy' 
In the file `parametric_to_inversion.ipynb` run in Binder, the command ```
settings_galaxy = ag.SettingsGalaxy(
    positions_threshold=result_1_list[0].positions_threshold_from(
        factor=3.0, minimum_threshold=0.2
    )
)```

fails with the error message: `AttributeError: module 'autogalaxy' has no attribute 'SettingsGalaxy' `. Connected to [#4475](https://github.com/openjournals/joss-reviews/issues/4475).
