### 2022-07-17 Jammy2211/PyAutoGalaxy#32 @Jammy2211
Good spot, this is fixed by changing:

`model = af.Collection(galaxy=galaxy_model)`

to:

`model = af.Collection(galaxies=af.Collection(galaxy=galaxy_model))`

I have updated the notebook locally and will push the fix once my full suite of integration tests have run.
