### 2022-07-17 Jammy2211/PyAutoGalaxy#32 @Jammy2211
Also going to figure out why the animations / visuals are not displaying on the noteobook :/
