### 2022-07-17 Jammy2211/PyAutoGalaxy#33 @Jammy2211
Fixed this by just deleting the line of code that defines `settings_galaxy` altogether, its a weird remnant of another project and not used for anything.

Will push a fixed script soon.
