### 2022-07-17 Jammy2211/PyAutoGalaxy#31 @Jammy2211
This is either an issue with `scipy` or `pylops` versions.

I am on `pylops==1.11.1` and `scipy==1.5.4`.

Could you check if installing pylops 1.11.1 fixes this for you, and if not then try scipy please? :).

I will likely try update the code to support both once I understand the library that is the issue.
