### 2022-07-17 971fab7
simplify light profile tests
