+++ test_autogalaxy/profiles/light_profiles/test_light_profiles.py
@@ -180,40 +180,8 @@ class TestDecorators:
 
 
 class TestGaussian:
-    def test__intensity_as_radius__correct_value(self):
-        gaussian = ag.lp.EllGaussian(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), intensity=1.0, sigma=1.0
-        )
 
-        image = gaussian.image_2d_via_radii_from(grid_radii=1.0)
-
-        assert image == pytest.approx(0.60653, 1e-2)
-
-        gaussian = ag.lp.EllGaussian(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), intensity=2.0, sigma=1.0
-        )
-
-        image = gaussian.image_2d_via_radii_from(grid_radii=1.0)
-
-        assert image == pytest.approx(2.0 * 0.60653, 1e-2)
-
-        gaussian = ag.lp.EllGaussian(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), intensity=1.0, sigma=2.0
-        )
-
-        image = gaussian.image_2d_via_radii_from(grid_radii=1.0)
-
-        assert image == pytest.approx(0.882496, 1e-2)
-
-        gaussian = ag.lp.EllGaussian(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), intensity=1.0, sigma=2.0
-        )
-
-        image = gaussian.image_2d_via_radii_from(grid_radii=3.0)
-
-        assert image == pytest.approx(0.32465, 1e-2)
-
-    def test__image_2d_from__same_values_as_above(self):
+    def test__image_2d_from(self):
         gaussian = ag.lp.EllGaussian(
             centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), intensity=1.0, sigma=1.0
         )
@@ -246,73 +214,6 @@ class TestGaussian:
 
         assert image == pytest.approx(0.3246, 1e-2)
 
-    def test__image_2d_from__change_geometry(self):
-        gaussian = ag.lp.EllGaussian(
-            centre=(1.0, 1.0), elliptical_comps=(0.0, 0.0), intensity=1.0, sigma=1.0
-        )
-        image = gaussian.image_2d_from(grid=np.array([[1.0, 0.0]]))
-        assert image == pytest.approx(0.60653, 1e-2)
-
-        gaussian = ag.lp.EllGaussian(
-            centre=(0.0, 0.0),
-            elliptical_comps=(0.0, 0.333333),
-            intensity=1.0,
-            sigma=1.0,
-        )
-
-        image = gaussian.image_2d_from(grid=np.array([[1.0, 0.0]]))
-
-        assert image == pytest.approx(0.60653, 1e-2)
-
-        gaussian_0 = ag.lp.EllGaussian(
-            centre=(-3.0, -0.0),
-            elliptical_comps=(0.0, 0.333333),
-            intensity=1.0,
-            sigma=1.0,
-        )
-
-        gaussian_1 = ag.lp.EllGaussian(
-            centre=(3.0, 0.0),
-            elliptical_comps=(0.0, 0.333333),
-            intensity=1.0,
-            sigma=1.0,
-        )
-
-        image_0 = gaussian_0.image_2d_from(
-            grid=np.array([[0.0, 0.0], [0.0, 1.0], [0.0, -1.0]])
-        )
-
-        image_1 = gaussian_1.image_2d_from(
-            grid=np.array([[0.0, 0.0], [0.0, 1.0], [0.0, -1.0]])
-        )
-
-        assert image_0 == pytest.approx(image_1, 1e-4)
-
-        gaussian_0 = ag.lp.EllGaussian(
[…]
