+++ test_autogalaxy/profiles/test_geometry_profiles.py
@@ -11,234 +11,231 @@ from autogalaxy.profiles import geometry_profiles
 directory = path.dirname(path.realpath(__file__))
 
 
-class TestEllProfile:
-    def test__cos_and_sin_from_x(self):
+def test__cos_and_sin_from_x():
 
-        elliptical_profile = geometry_profiles.EllProfile(
-            centre=(1.0, 1.0), elliptical_comps=(0.0, 0.0)
-        )
+    elliptical_profile = geometry_profiles.EllProfile(
+        centre=(1.0, 1.0), elliptical_comps=(0.0, 0.0)
+    )
 
-        cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
+    cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
 
-        assert cos_phi == 1.0
-        assert sin_phi == 0.0
+    assert cos_phi == 1.0
+    assert sin_phi == 0.0
 
-        elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
-            centre=(1, 1), axis_ratio=0.1, angle=45.0
-        )
+    elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
+        centre=(1, 1), axis_ratio=0.1, angle=45.0
+    )
 
-        cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
+    cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
 
-        assert cos_phi == pytest.approx(0.707, 1e-3)
-        assert sin_phi == pytest.approx(0.707, 1e-3)
+    assert cos_phi == pytest.approx(0.707, 1e-3)
+    assert sin_phi == pytest.approx(0.707, 1e-3)
 
-        elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
-            centre=(1, 1), axis_ratio=0.1, angle=60.0
-        )
+    elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
+        centre=(1, 1), axis_ratio=0.1, angle=60.0
+    )
 
-        cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
+    cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
 
-        assert cos_phi == pytest.approx(0.5, 1e-3)
-        assert sin_phi == pytest.approx(0.866, 1e-3)
+    assert cos_phi == pytest.approx(0.5, 1e-3)
+    assert sin_phi == pytest.approx(0.866, 1e-3)
 
-        elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
-            centre=(1, 1), axis_ratio=0.1, angle=225.0
-        )
+    elliptical_profile = geometry_profiles.EllProfile.from_axis_ratio_and_phi(
+        centre=(1, 1), axis_ratio=0.1, angle=225.0
+    )
 
-        cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
+    cos_phi, sin_phi = elliptical_profile.cos_and_sin_to_x_axis()
 
-        assert cos_phi == pytest.approx(0.707, 1e-3)
-        assert sin_phi == pytest.approx(0.707, 1e-3)
+    assert cos_phi == pytest.approx(0.707, 1e-3)
+    assert sin_phi == pytest.approx(0.707, 1e-3)
 
-    def test__transform_grid_to_and_from_reference_frame(self):
+def test__grid_to_eccentric_radii():
[…]
