### 2022-07-17 Jammy2211/PyAutoGalaxy#31 @alexandergagliano
My `pylops` version was already `1.11.1`, and after a reinstall I get the same error when trying to even import that package. Installing `scipy==1.5.4` fixes the error, both  in importing `pylops` and in running the test suite as a whole.
