### 2022-07-17 9d2b8a4
cleaned up mass profile tests
