+++ test_autogalaxy/profiles/mass_profiles/test_mass_profiles.py
@@ -28,642 +28,607 @@ def mass_within_radius_of_profile_from_grid_calculation(radius, profile):
     return mass_total
 
 
-class Test1DFromGrid:
-    def test__convergence_1d_from__grid_2d_in__returns_1d_image_via_projected_quantities(
-        self
-    ):
+def test__deflections_2d_via_potential_2d_from():
 
-        grid_2d = ag.Grid2D.uniform(shape_native=(5, 5), pixel_scales=1.0)
+    sis = ag.mp.SphIsothermal(centre=(0.0, 0.0), einstein_radius=2.0)
 
-        sie = ag.mp.EllIsothermal(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), einstein_radius=1.0
-        )
+    grid = ag.Grid2D.uniform(shape_native=(10, 10), pixel_scales=0.05, sub_size=1)
 
-        convergence_1d = sie.convergence_1d_from(grid=grid_2d)
-        convergence_2d = sie.convergence_2d_from(grid=grid_2d)
+    deflections_via_calculation = sis.deflections_yx_2d_from(grid=grid)
 
-        assert convergence_1d[0] == pytest.approx(convergence_2d.native[2, 2], 1.0e-4)
-        assert convergence_1d[1] == pytest.approx(convergence_2d.native[2, 3], 1.0e-4)
-        assert convergence_1d[2] == pytest.approx(convergence_2d.native[2, 4], 1.0e-4)
+    deflections_via_potential = sis.deflections_2d_via_potential_2d_from(grid=grid)
 
-        sie = ag.mp.EllIsothermal(
-            centre=(0.2, 0.2), elliptical_comps=(0.3, 0.3), einstein_radius=1.0
-        )
+    mean_error = np.mean(
+        deflections_via_potential.slim - deflections_via_calculation.slim
+    )
 
-        convergence_1d = sie.convergence_1d_from(grid=grid_2d)
+    assert mean_error < 1e-4
 
-        grid_2d_projected = grid_2d.grid_2d_radial_projected_from(
-            centre=sie.centre, angle=sie.angle + 90.0
-        )
+    sie = ag.mp.EllIsothermal(
+        centre=(0.0, 0.0), elliptical_comps=(0.111111, 0.0), einstein_radius=2.0
+    )
 
-        convergence_projected = sie.convergence_2d_from(grid=grid_2d_projected)
+    grid = ag.Grid2D.uniform(shape_native=(10, 10), pixel_scales=0.05, sub_size=1)
 
-        assert convergence_1d == pytest.approx(convergence_projected, 1.0e-4)
-        assert (convergence_1d.grid_radial == np.array([0.0, 1.0, 2.0])).all()
+    deflections_via_calculation = sie.deflections_yx_2d_from(grid=grid)
 
-    def test__convergence_1d_from__grid_2d_irregular_in__returns_1d_quantities(self):
+    deflections_via_potential = sie.deflections_2d_via_potential_2d_from(grid=grid)
 
-        grid_2d = ag.Grid2DIrregular(grid=[[1.0, 1.0], [2.0, 2.0], [4.0, 4.0]])
+    mean_error = np.mean(
+        deflections_via_potential.slim - deflections_via_calculation.slim
+    )
 
-        sie = ag.mp.EllIsothermal(
-            centre=(0.0, 0.0), elliptical_comps=(0.0, 0.0), einstein_radius=1.0
-        )
+    assert mean_error < 1e-4
 
-        convergence_1d = sie.convergence_1d_from(grid=grid_2d)
-        convergence_2d = sie.convergence_2d_from(grid=grid_2d)
+    sie = ag.mp.EllIsothermal(
[…]
