### 2022-07-17 Jammy2211/PyAutoGalaxy#35 @alexandergagliano
Package fails 3/454 unit tests with pytest
When running pytest, the code catches 3 test failures across 2 files -- `test_autogalaxy/imaging/test_fit_imaging.py` and ` test_autogalaxy/profiles/light_profiles/test_light_profiles.py`. The error messages are:

```
----------------------------------------------------------------------------------- Captured stdout setup ------------------------------------------------------------------------------------
2022-07-17 12:10:33,661 - autoarray.dataset.imaging - INFO - IMAGING - Data masked, contains a total of 9 image-pixels
------------------------------------------------------------------------------------- Captured log setup -------------------------------------------------------------------------------------
INFO     autoarray.dataset.imaging:imaging.py:366 IMAGING - Data masked, contains a total of 9 image-pixels
____________________________________________________________________ TestMoffat.test__intensity_as_radius__correct_value _____________________________________________________________________

self = <test_light_profiles.TestMoffat object at 0x7ff87748a850>

    def test__intensity_as_radius__correct_value(self):
>       moffat = ag.lp.EllMoffat(
            centre=(0.0, 0.0),
            elliptical_comps=(0.0, 0.0),
            intensity=1.0,
            alpha=1.0,
            beta=1.0,
        )
E       AttributeError: module 'autogalaxy.profiles.light_profiles' has no attribute 'EllMoffat'

test_autogalaxy/profiles/light_profiles/test_light_profiles.py:344: AttributeError
```

and

```
____________________________________________________________________ test__plane_linear_light_profiles_to_light_profiles _____________________________________________________________________

masked_imaging_7x7 = <autoarray.dataset.imaging.Imaging object at 0x7ff84c1fd880>

    def test__plane_linear_light_profiles_to_light_profiles(masked_imaging_7x7):

        g0 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))

        g0_linear = ag.Galaxy(
            redshift=0.5, light_profile=ag.lp_linear.EllSersic(sersic_index=1.0)
        )

        g1_linear = ag.Galaxy(
            redshift=1.0, light_profile=ag.lp_linear.EllSersic(sersic_index=4.0)
        )

        plane = ag.Plane(galaxies=[g0, g0_linear, g1_linear])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.galaxies[0].light_profile.intensity == pytest.approx(1.0, 1.0e-4)

>       plane = fit.plane_linear_light_profiles_to_light_profiles
E       AttributeError: 'FitImaging' object has no attribute 'plane_linear_light_profiles_to_light_profiles'

test_autogalaxy/imaging/test_fit_imaging.py:587: AttributeError
```

and

```
_________________________________________________________________________________ test__fit_figure_of_merit __________________________________________________________________________________

masked_imaging_7x7 = <autoarray.dataset.imaging.Imaging object at 0x7ff84bbe2730>

    def test__fit_figure_of_merit(masked_imaging_7x7):

        g0 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
        g1 = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))

        plane = ag.Plane(redshift=0.5, galaxies=[g0, g1])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.log_likelihood == pytest.approx(-75938.05, 1e-4)
        assert fit.figure_of_merit == pytest.approx(-75938.05, 1.0e-4)

        pix = ag.pix.Rectangular(shape=(3, 3))
        reg = ag.reg.Constant(coefficient=1.0)

        galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)

        plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5), galaxy_pix])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.log_evidence == pytest.approx(-22.9005, 1e-4)
        assert fit.figure_of_merit == pytest.approx(-22.9005, 1.0e-4)

        galaxy_light = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))

        plane = ag.Plane(redshift=0.5, galaxies=[galaxy_light, galaxy_pix])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.log_evidence == pytest.approx(-6840.5851, 1e-4)
        assert fit.figure_of_merit == pytest.approx(-6840.5851, 1.0e-4)

        g0_linear_light = ag.Galaxy(
            redshift=0.5, light_profile=ag.lp_linear.EllSersic(sersic_index=1.0)
        )

        g1_linear_light = ag.Galaxy(
            redshift=0.5, light_profile=ag.lp_linear.EllSersic(sersic_index=4.0)
        )

        plane = ag.Plane(redshift=0.5, galaxies=[g0_linear_light, g1_linear_light])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.log_likelihood == pytest.approx(-14.52327, 1e-4)
        assert fit.figure_of_merit == pytest.approx(-14.52327, 1.0e-4)

        g0_operated_light = ag.Galaxy(
            redshift=0.5, light_profile=ag.lp_operated.EllSersic(intensity=1.0)
        )
        g1_operated_light = ag.Galaxy(
            redshift=0.5, light_profile=ag.lp_operated.EllSersic(intensity=1.0)
        )

        plane = ag.Plane(redshift=0.5, galaxies=[g0_operated_light, g1_operated_light])

        fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)

        assert fit.log_likelihood == pytest.approx(-342374.9618, 1e-4)
        assert fit.figure_of_merit == pytest.approx(-342374.9618, 1.0e-4)

        g0_linear_operated_light = ag.Galaxy(
>           redshift=0.5, light_profile=ag.lp_linear_operated.EllSersic(sersic_index=1.0)
        )
E       AttributeError: module 'autogalaxy.profiles.light_profiles.light_profiles_linear_operated' has no attribute 'EllSersic'

test_autogalaxy/imaging/test_fit_imaging.py:208: AttributeError
```

Related to [#4475](https://github.com/openjournals/joss-reviews/issues/4475).
