### 2022-07-17 3322bff
cleaned up dark tests
