+++ test_autogalaxy/profiles/mass_profiles/test_total_mass_profiles.py
@@ -7,7 +7,7 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 class TestPointMass:
-    def test__deflections_from(self):
+    def test__deflections_yx_2d_from(self):
 
         # The radial coordinate at (1.0, 1.0) is sqrt(2)
         # This is decomposed into (y,x) angles of sin(45) = cos(45) = sqrt(2) / 2.0
@@ -50,130 +50,9 @@ class TestPointMass:
         assert deflections[0, 0] == pytest.approx(0.5, 1e-3)
         assert deflections[0, 1] == pytest.approx(0.5, 1e-3)
 
-    def test__deflections__change_geometry(self):
-
-        point_mass_0 = ag.mp.PointMass(centre=(0.0, 0.0))
-        point_mass_1 = ag.mp.PointMass(centre=(1.0, 1.0))
-        deflections_0 = point_mass_0.deflections_yx_2d_from(grid=np.array([[1.0, 1.0]]))
-        deflections_1 = point_mass_1.deflections_yx_2d_from(grid=np.array([[0.0, 0.0]]))
-        assert deflections_0[0, 0] == pytest.approx(-deflections_1[0, 0], 1e-5)
-        assert deflections_0[0, 1] == pytest.approx(-deflections_1[0, 1], 1e-5)
-
-        point_mass_0 = ag.mp.PointMass(centre=(0.0, 0.0))
-        point_mass_1 = ag.mp.PointMass(centre=(0.0, 0.0))
-        deflections_0 = point_mass_0.deflections_yx_2d_from(grid=np.array([[1.0, 0.0]]))
-        deflections_1 = point_mass_1.deflections_yx_2d_from(grid=np.array([[0.0, 1.0]]))
-        assert deflections_0[0, 0] == pytest.approx(deflections_1[0, 1], 1e-5)
-        assert deflections_0[0, 1] == pytest.approx(deflections_1[0, 0], 1e-5)
-
-    def test__multiple_coordinates_in__multiple_coordinates_out(self):
-
-        point_mass = ag.mp.PointMass(centre=(1.0, 2.0), einstein_radius=1.0)
-
-        deflections = point_mass.deflections_yx_2d_from(
-            grid=np.array([[2.0, 3.0], [2.0, 3.0], [2.0, 3.0]])
-        )
-
-        assert deflections[0, 0] == pytest.approx(0.5, 1e-3)
-        assert deflections[0, 1] == pytest.approx(0.5, 1e-3)
-        assert deflections[1, 0] == pytest.approx(0.5, 1e-3)
-        assert deflections[1, 1] == pytest.approx(0.5, 1e-3)
-        assert deflections[2, 0] == pytest.approx(0.5, 1e-3)
-        assert deflections[2, 1] == pytest.approx(0.5, 1e-3)
-
-        point_mass = ag.mp.PointMass(centre=(0.0, 0.0), einstein_radius=1.0)
-
-        deflections = point_mass.deflections_yx_2d_from(
-            grid=np.array([[1.0, 1.0], [2.0, 2.0], [1.0, 1.0], [2.0, 2.0]])
-        )
-
-        assert deflections[0, 0] == pytest.approx(0.5, 1e-3)
-        assert deflections[0, 1] == pytest.approx(0.5, 1e-3)
-
-        assert deflections[1, 0] == pytest.approx(0.25, 1e-3)
-        assert deflections[1, 1] == pytest.approx(0.25, 1e-3)
-
-        assert deflections[2, 0] == pytest.approx(0.5, 1e-3)
[…]
