### 2022-07-17 f22d578
stellar profile test cleaned
