+++ test_autogalaxy/profiles/mass_profiles/test_stellar_mass_profiles.py
@@ -7,43 +7,6 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 class TestEllGaussian:
-    def test__convergence_2d_from(self):
-        gaussian = ag.mp.EllGaussian(
-            centre=(0.0, 0.0),
-            elliptical_comps=(0.0, 0.0),
-            intensity=1.0,
-            sigma=1.0,
-            mass_to_light_ratio=1.0,
-        )
-
-        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
-
-        assert convergence == pytest.approx(0.60653, 1e-2)
-
-        gaussian = ag.mp.EllGaussian(
-            centre=(0.0, 0.0),
-            elliptical_comps=(0.0, 0.0),
-            intensity=1.0,
-            sigma=1.0,
-            mass_to_light_ratio=2.0,
-        )
-
-        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
-
-        assert convergence == pytest.approx(2.0 * 0.60653, 1e-2)
-
-        gaussian = ag.mp.EllGaussian(
-            centre=(0.0, 0.0),
-            elliptical_comps=(0.0, 0.333333),
-            intensity=2.0,
-            sigma=3.0,
-            mass_to_light_ratio=4.0,
-        )
-
-        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
-
-        assert convergence == pytest.approx(7.88965, 1e-2)
-
     def test__deflections_2d_via_analytic_from(self):
 
         gaussian = ag.mp.EllGaussian(
@@ -187,6 +150,43 @@ class TestEllGaussian:
 
         assert (deflections == deflections_via_integral).all()
 
+    def test__convergence_2d_from(self):
+        gaussian = ag.mp.EllGaussian(
+            centre=(0.0, 0.0),
+            elliptical_comps=(0.0, 0.0),
+            intensity=1.0,
+            sigma=1.0,
+            mass_to_light_ratio=1.0,
+        )
+
+        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
+
+        assert convergence == pytest.approx(0.60653, 1e-2)
+
+        gaussian = ag.mp.EllGaussian(
+            centre=(0.0, 0.0),
+            elliptical_comps=(0.0, 0.0),
+            intensity=1.0,
+            sigma=1.0,
+            mass_to_light_ratio=2.0,
+        )
+
+        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
+
+        assert convergence == pytest.approx(2.0 * 0.60653, 1e-2)
+
+        gaussian = ag.mp.EllGaussian(
+            centre=(0.0, 0.0),
+            elliptical_comps=(0.0, 0.333333),
+            intensity=2.0,
+            sigma=3.0,
+            mass_to_light_ratio=4.0,
+        )
+
+        convergence = gaussian.convergence_2d_from(grid=np.array([[0.0, 1.0]]))
+
+        assert convergence == pytest.approx(7.88965, 1e-2)
+
     def test__intensity_and_convergence_match_for_mass_light_ratio_1(self):
 
         gaussian_light_profile = ag.lp.EllGaussian(
@@ -208,8 +208,6 @@ class TestEllGaussian:
         convergence = gaussian_mass_profile.convergence_2d_from(
             grid=np.array([[1.0, 0.0]])
         )
-
-        print(intensity, convergence)
 
         assert (intensity == convergence).all()
 
[…]
