+++ test_autogalaxy/profiles/mass_profiles/test_mass_sheets.py
@@ -8,80 +8,6 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 class TestMassSheet:
-    def test__convergence_2d_from(self):
-
-        mass_sheet = ag.mp.MassSheet(centre=(0.0, 0.0), kappa=1.0)
-
-        convergence = mass_sheet.convergence_2d_from(grid=np.array([[1.0, 0.0]]))
-
-        assert convergence[0] == pytest.approx(1.0, 1e-3)
-
-        convergence = mass_sheet.convergence_2d_from(
-            grid=np.array([[1.0, 0.0], [3.0, 3.0], [5.0, -9.0]])
-        )
-
-        assert convergence[0] == pytest.approx(1.0, 1e-3)
-        assert convergence[1] == pytest.approx(1.0, 1e-3)
-        assert convergence[2] == pytest.approx(1.0, 1e-3)
-
-        mass_sheet = ag.mp.MassSheet(centre=(0.0, 0.0), kappa=-3.0)
-
-        convergence = mass_sheet.convergence_2d_from(
-            grid=np.array([[1.0, 0.0], [3.0, 3.0], [5.0, -9.0]])
-        )
-
-        assert convergence[0] == pytest.approx(-3.0, 1e-3)
-        assert convergence[1] == pytest.approx(-3.0, 1e-3)
-        assert convergence[2] == pytest.approx(-3.0, 1e-3)
-
-        mass_sheet = ag.mp.MassSheet(centre=(0.0, 0.0), kappa=1.0)
-
-        convergence = mass_sheet.convergence_2d_from(
-            grid=ag.Grid2D.manual_native(
-                [[[1.0, 0.0], [1.0, 0.0]], [[1.0, 0.0], [1.0, 0.0]]],
-                sub_size=2,
-                pixel_scales=(1.0, 1.0),
-            )
-        )
-
-        assert convergence[0] == pytest.approx(1.0, 1e-3)
-        assert convergence[1] == pytest.approx(1.0, 1e-3)
-        assert convergence[2] == pytest.approx(1.0, 1e-3)
-        assert convergence[3] == pytest.approx(1.0, 1e-3)
-
-        convergence = mass_sheet.convergence_2d_from(grid=np.array([[1.0, 0.0]]))
-
-        assert convergence[0] == pytest.approx(1.0, 1e-3)
-
-    def test__potential_from(self):
-        mass_sheet = ag.mp.MassSheet(centre=(0.0, 0.0), kappa=1.0)
-
-        potential = mass_sheet.potential_2d_from(
-            grid=np.array([[1.0, 0.0], [3.0, 3.0], [5.0, -9.0]])
-        )
-
-        assert potential[0] == pytest.approx(0.0, 1e-3)
-        assert potential[1] == pytest.approx(0.0, 1e-3)
-        assert potential[2] == pytest.approx(0.0, 1e-3)
-
-        mass_sheet = ag.mp.MassSheet(centre=(0.0, 0.0), kappa=1.0)
-
-        potential = mass_sheet.potential_2d_from(
-            grid=ag.Grid2D.manual_native(
-                [[[1.0, 0.0], [1.0, 0.0]], [[1.0, 0.0], [1.0, 0.0]]],
-                sub_size=2,
-                pixel_scales=(1.0, 1.0),
-            )
-        )
-
-        assert potential[0] == pytest.approx(0.0, 1e-3)
-        assert potential[1] == pytest.approx(0.0, 1e-3)
-        assert potential[2] == pytest.approx(0.0, 1e-3)
-        assert potential[3] == pytest.approx(0.0, 1e-3)
-
[…]
