+++ test_autogalaxy/profiles/mass_profiles/test_stellar_mass_profiles.py
@@ -148,7 +148,7 @@ class TestEllGaussian:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_integral).all()
+        assert deflections == pytest.approx(deflections_via_integral, 1.0e-4)
 
     def test__convergence_2d_from(self):
         gaussian = ag.mp.EllGaussian(
@@ -383,7 +383,7 @@ class TestSersic:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_integral).all()
+        assert deflections == pytest.approx(deflections_via_integral, 1.0e-4)
 
         gaussian = ag.mp.SphSersic()
 
@@ -392,7 +392,7 @@ class TestSersic:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_integral).all()
+        assert deflections == pytest.approx(deflections_via_integral, 1.0e-4)
 
         elliptical = ag.mp.EllSersic(
             centre=(0.0, 0.0),
@@ -584,10 +584,11 @@ class TestSersic:
             mass_to_light_ratio=1.0,
         )
 
-        assert (
-            elliptical.convergence_2d_from(grid=grid)
-            == spherical.convergence_2d_from(grid=grid)
-        ).all()
+        ell_convergence_2d = elliptical.convergence_2d_from(grid=grid)
+        sph_convergence_2d = spherical.convergence_2d_from(grid=grid)
+
+        assert ell_convergence_2d == pytest.approx(sph_convergence_2d, 1.0e-4)
+
 
 
 class TestExponential:
@@ -600,7 +601,7 @@ class TestExponential:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_cse).all()
+        assert deflections == pytest.approx(deflections_via_cse, 1.0e-4)
 
         gaussian = ag.mp.SphExponential()
 
@@ -609,7 +610,7 @@ class TestExponential:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_cse).all()
+        assert deflections == pytest.approx(deflections_via_cse, 1.0e-4)
 
     def test__deflections_2d_via_integral_from(self):
         exponential = ag.mp.EllExponential(
@@ -812,10 +813,11 @@ class TestExponential:
             mass_to_light_ratio=1.0,
         )
 
-        assert (
-            elliptical.convergence_2d_from(grid=grid)
-            == spherical.convergence_2d_from(grid=grid)
-        ).all()
+        ell_convergence_2d = elliptical.convergence_2d_from(grid=grid)
+        sph_convergence_2d = spherical.convergence_2d_from(grid=grid)
+
+        assert ell_convergence_2d == pytest.approx(sph_convergence_2d, 1.0e-4)
+
 
 
 class TestDevVaucouleurs:
@@ -828,7 +830,7 @@ class TestDevVaucouleurs:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_cse).all()
+        assert deflections == pytest.approx(deflections_via_cse, 1.0e-4)
 
         gaussian = ag.mp.SphDevVaucouleurs()
 
@@ -837,7 +839,7 @@ class TestDevVaucouleurs:
             grid=np.array([[1.0, 0.0]])
         )
 
-        assert (deflections == deflections_via_cse).all()
+        assert deflections == pytest.approx(deflections_via_cse, 1.0e-4)
 
     def test__deflections_via_integral_from(self):
         dev = ag.mp.EllDevVaucouleurs(
[…]
