### 2022-07-17 Jammy2211/PyAutoGalaxy#35 @Jammy2211
A `git pull` should fix this.
