+++ autogalaxy/imaging/fit_imaging.py
@@ -1,5 +1,5 @@
 import numpy as np
-from typing import Dict, Optional
+from typing import Dict, List, Optional
 
 from autoconf import conf
 from autoconf import cached_property
@@ -7,12 +7,15 @@ from autoconf import cached_property
 import autoarray as aa
 
 from autogalaxy.abstract_fit import AbstractFit
+from autogalaxy.analysis.preloads import Preloads
 from autogalaxy.galaxy.galaxy import Galaxy
 from autogalaxy.hyper.hyper_data import HyperImageSky
 from autogalaxy.hyper.hyper_data import HyperBackgroundNoise
 from autogalaxy.plane.plane import Plane
 from autogalaxy.plane.to_inversion import PlaneToInversion
 from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
+
+from autogalaxy import exc
 
 
 class FitImaging(aa.FitImaging, AbstractFit):
@@ -25,16 +28,57 @@ class FitImaging(aa.FitImaging, AbstractFit):
         use_hyper_scaling: bool = True,
         settings_pixelization: aa.SettingsPixelization = aa.SettingsPixelization(),
         settings_inversion: aa.SettingsInversion = aa.SettingsInversion(),
-        preloads: aa.Preloads = aa.Preloads(),
+        preloads: aa.Preloads = Preloads(),
         profiling_dict: Optional[Dict] = None,
     ):
-        """ An lens fitter, which contains the plane's used to perform the fit and functions to manipulate \
-        the lens dataset's hyper_galaxies.
+        """
+        Fits an imaging dataset using a `Plane` object.
+
+        The fit performs the following steps:
+
+        1) Compute the sum of all images of galaxy light profiles in the `Plane`.
+
+        2) Blur this with the imaging PSF to created the `blurred_image`.
+
+        3) Subtract this image from the `data` to create the `profile_subtracted_image`.
+
+        4) If the `Plane` has any linear algebra objects (e.g. linear light profiles, a pixelization / regulariation)
+        fit the `profile_subtracted_image` with these objects via an inversion.
+
+        5) Compute the `model_data` as the sum of the `blurred_image` and `reconstructed_data` of the inversion (if
+        an inversion is not performed the `model_data` is only the `blurred_image`.
+
+        6) Subtract the `model_data` from the data and compute the residuals, chi-squared and likelihood via the
+        noise-map (if an inversion is performed the `log_evidence`, including additional terms describing the linear
+        algebra solution, is computed).
+
+        When performing a `model-fit`via an `AnalysisImaging` object the `figure_of_merit` of this `FitImaging` object
+        is called and returned in the `log_likelihood_function`.
 
         Parameters
         -----------
+        dataset
+            The imaging dataset which is fitted by the galaxies in the plane.
         plane
-            The plane of galaxies whose model images are used to fit the imaging data.
+            The plane of galaxies whose light profile images are used to fit the imaging data.
+        hyper_image_sky
+            If included, accounts for the background sky in the fit.
+        hyper_background_noise
+            If included, adds a noise-scaling term to the background to account for an inaacurate background sky model.
+        use_hyper_scaling
+            If set to False, the hyper scaling functions (e.g. the `hyper_image_sky` / `hyper_background_noise`) are
+            omitted irrespective of their inputs.
+        settings_pixelization
+            Settings controlling how a pixelization is fitted for example if a border is used when creating the
+            pixelization.
+        settings_inversion
+            Settings controlling how an inversion is fitted for example which linear algebra formalism is used.
+        preloads
+            Contains preloaded calculations (e.g. linear algebra matrices) which can skip certain calculations in
+            the fit.
+        profiling_dict
[…]
