### 2022-07-18 Jammy2211/PyAutoGalaxy#35 @alexandergagliano
@Jammy2211 I've done a fresh install and a fresh pull of this repo and now am getting five errors (they look related to the original 3). I should mention that this is after pip installing the package and then running the code tests from this repo. If the pypi release and the repo are at different versions, that might be the cause?

```
E       AttributeError: module 'autogalaxy.profiles.light_profiles.light_profiles_linear_operated' has no attribute 'EllSersic'

imaging/test_fit_imaging.py:208: AttributeError
```

```
E       AttributeError: 'FitImaging' object has no attribute 'plane_linear_light_profiles_to_light_profiles'

imaging/test_fit_imaging.py:587: AttributeError
```

```
E       AssertionError: assert not True
E        +  where True = isinstance(EllSersic\ncentre: (0.0, 0.0)\nelliptical_comps: (0.0, 0.0)\nintensity: 1.0\neffective_radius: 0.6\nsersic_index: 4.0, <class 'autogalaxy.profiles.light_profiles.light_profiles_linear.LightProfileLinear'>)
E        +    where EllSersic\ncentre: (0.0, 0.0)\nelliptical_comps: (0.0, 0.0)\nintensity: 1.0\neffective_radius: 0.6\nsersic_index: 4.0 = Redshift: 0.5\nLight Profiles:\nEllSersic\ncentre: (0.0, 0.0)\nelliptical_comps: (0.0, 0.0)\nintensity: 1.0\neffective_radius: 0.6\nsersic_index: 4.0.bulge
E        +    and   <class 'autogalaxy.profiles.light_profiles.light_profiles_linear.LightProfileLinear'> = <module 'autogalaxy.profiles.light_profiles.light_profiles_linear' from '/Users/alexgagliano/miniconda3/lib/python3.9/site-packages/autogalaxy/profiles/light_profiles/light_profiles_linear.py'>.LightProfileLinear
E        +      where <module 'autogalaxy.profiles.light_profiles.light_profiles_linear' from '/Users/alexgagliano/miniconda3/lib/python3.9/site-packages/autogalaxy/profiles/light_profiles/light_profiles_linear.py'> = ag.lp_linear

imaging/model/test_result_imaging.py:58: AssertionError
```

```
E       AttributeError: module 'autogalaxy.profiles.light_profiles' has no attribute 'EllMoffat'

profiles/light_profiles/test_light_profiles.py:229: AttributeError
```

```
E       assert 9.0 < Array2D(2.89827085)
E        +  where Array2D(2.89827085) = <autoarray.dataset.imaging.Imaging object at 0x7fe5beee66a0>.signal_to_noise_max

profiles/light_profiles/test_snr_light_profiles.py:40: AssertionError
```
