### 2022-07-18 Jammy2211/PyAutoGalaxy#35 @Jammy2211
Ok yes, the issue is that the pip install and master branch are out of sync (as one would expect).

If you are to run the tests you should also run the PyAutoGalaxy source code cloned from the repo (and will possibly need the parent projects cloned as well):

https://pyautogalaxy.readthedocs.io/en/latest/installation/source.html
