### 2022-07-18 Jammy2211/PyAutoGalaxy#35 @alexandergagliano
A few errors in PyAutoArray after following the instructions in "Build All Projects" [here](https://pyautogalaxy.readthedocs.io/en/latest/installation/source.html). It looks like these are all coming from similar functions:

===================================================================================== short test summary info ======================================================================================
FAILED test_autoarray/inversion/inversion/test_factory.py::test__inversion_imaging__via_regularizations - AttributeError: module 'autoarray.util.nn.nn_py' has no attribute 'natural_interpolatio...
FAILED test_autoarray/inversion/linear_eqn/test_leq_util.py::TestCurvatureMatrixImaging::test__curvature_matrix_via_preload_imaging - assert False
FAILED test_autoarray/inversion/mappers/test_voronoi.py::test__pix_indexes_for_sub_slim_index__matches_util - AttributeError: module 'autoarray.util.nn.nn_py' has no attribute 'natural_interpol...
FAILED test_autoarray/structures/grids/test_grid_2d_pixelizations.py::TestGrid2DVoronoi::test__interpolated_array_from - AttributeError: module 'autoarray.util.nn.nn_py' has no attribute 'natur...
FAILED test_autoarray/util/test_nn.py::test__returning_weights_correct - AttributeError: module 'autoarray.util.nn.nn_py' has no attribute 'natural_interpolation_weights'
FAILED test_autoarray/util/test_nn.py::test__nn_interpolation_correct - AttributeError: module 'autoarray.util.nn.nn_py' has no attribute 'natural_interpolation'
====================================================================== 6 failed, 835 passed, 15 warnings in 154.93s (0:02:34) ======================================================================
