### 2022-07-18 Jammy2211/PyAutoGalaxy#35 @alexandergagliano
PyAutoArray testing re-done from scratch in a new condo environment and fails the same tests.
