### 2022-07-18 Jammy2211/PyAutoGalaxy#35 @Jammy2211
Yeah these tests require an optional C library, I hadnt really considered we put some tests that depend on it.

Will look into fixing them by removing the dependency.
