+++ autogalaxy/config/visualize/plots.ini
@@ -40,9 +40,9 @@ chi_squared_map=False
 [fit_imaging]
 
 [galaxies]
-image=True
-convergence=True
-potential=True
+image=False
+convergence=False
+potential=False
 
 [inversion]
 all_at_end_png=False
