### 2022-07-18 1579603
blurred mapping matrix to operated mapping matrix
