### 2022-07-21 f3f7068
Plane now computes a regulrization list, placing None when no regularization in galaxy
