+++ autogalaxy/util/__init__.py
@@ -10,9 +10,15 @@ from autoarray.fit import fit_util as fit
 from autoarray.inversion.pixelizations import pixelization_util as pixelization
 from autoarray.inversion.mappers import mapper_util as mapper
 from autoarray.inversion.regularization import regularization_util as regularization
-from autoarray.inversion.linear_eqn import leq_util as leq
 from autoarray.inversion.inversion import inversion_util as inversion
+from autoarray.inversion.inversion.imaging import (
+    inversion_imaging_util as inversion_imaging,
+)
+from autoarray.inversion.inversion.interferometer import (
+    inversion_interferometer_util as inversion_interferometer,
+)
 from autoarray.operators import transformer_util as transformer
+from autoarray.util import misc_util as misc
 from autogalaxy.analysis import model_util as model
 
 from autogalaxy.util import error_util as error
