### 2022-07-21 89ffc58
updated plane to use regularization_list and fixed tests
