+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -414,12 +414,8 @@ def test__image_2d_from__does_not_include_linear_light_profiles(sub_grid_2d_7x7,
 
 def test__only_pixelization_raises_error():
     with pytest.raises(exc.GalaxyException):
-        ag.Galaxy(redshift=0.5, pixelization=object())
+        ag.Galaxy(redshift=0.5, pixelization=ag.pix.Rectangular())
 
-
-def test__only_regularization_raises_error():
-    with pytest.raises(exc.GalaxyException):
-        ag.Galaxy(redshift=0.5, regularization=object())
 
 
 def test__light_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric_results():
