+++ autogalaxy/interferometer/fit_interferometer.py
@@ -131,7 +131,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
         return self.visibilities - self.profile_visibilities
 
     @cached_property
-    def inversion(self) -> Optional[aa.Inversion]:
+    def inversion(self) -> Optional[aa.AbstractInversion]:
         """
         If the plane has linear objects which are used to fit the data (e.g. a linear light profile / pixelization)
         this function returns a linear inversion, where the flux values of these objects (e.g. the `intensity`
