### 2022-07-22 9ae25c1
added basis class but now need to split LightProfile into AbstractLightProfile
