### 2022-07-22 4e7cb8d
fixed tests now LinearObjFuncList uses light of light profiles
