+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -417,7 +417,6 @@ def test__only_pixelization_raises_error():
         ag.Galaxy(redshift=0.5, pixelization=ag.pix.Rectangular())
 
 
-
 def test__light_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric_results():
     lp_0 = ag.lp.EllSersic(centre=(0.0, 0.0), intensity=1.0)
 
