+++ autogalaxy/abstract_fit.py
@@ -92,15 +92,21 @@ class AbstractFit:
         if self.inversion is None:
             return None
 
-        linear_obj_func_list = self.inversion.cls_list_from(cls=aa.LinearObjFuncList)
+        linear_obj_func_list = self.inversion.cls_list_from(
+            cls=aa.AbstractLinearObjFuncList
+        )
 
         linear_light_profile_intensity_dict = {}
 
         for linear_obj_func in linear_obj_func_list:
 
-            linear_light_profile_intensity_dict[linear_obj_func.light_profile] = float(
-                self.inversion.reconstruction_dict[linear_obj_func]
-            )
+            reconstruction = self.inversion.reconstruction_dict[linear_obj_func]
+
+            for i, light_profile in enumerate(linear_obj_func.light_profile_list):
+
+                linear_light_profile_intensity_dict[light_profile] = float(
+                    reconstruction[i]
+                )
 
         return linear_light_profile_intensity_dict
 
