+++ test_autogalaxy/plane/test_to_inversion.py
@@ -40,9 +40,9 @@ def test__lp_linear_func_galaxy_dict_from(lp_0):
     assert lp_linear_func_galaxy_dict[lp_linear_func_list[1]] == galaxy_0
     assert lp_linear_func_galaxy_dict[lp_linear_func_list[2]] == galaxy_1
 
-    assert lp_linear_func_list[0].light_profile == lp_linear_0
-    assert lp_linear_func_list[1].light_profile == lp_linear_1
-    assert lp_linear_func_list[2].light_profile == lp_linear_2
+    assert lp_linear_func_list[0].light_profile_list[0] == lp_linear_0
+    assert lp_linear_func_list[1].light_profile_list[0] == lp_linear_1
+    assert lp_linear_func_list[2].light_profile_list[0] == lp_linear_2
 
 
 def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
