+++ test_autogalaxy/imaging/test_simulate_and_fit_imaging.py
@@ -94,7 +94,9 @@ def test__simulate_imaging_data_and_fit__known_likelihood():
     )
 
     galaxy_0 = ag.Galaxy(
-        redshift=0.5, light=ag.lp.EllSersic(centre=(0.1, 0.1), intensity=0.1)
+        redshift=0.5,
+        bulge=ag.lp.EllSersic(centre=(0.1, 0.1), intensity=0.1),
+        disk=ag.lp.EllSersic(centre=(0.2, 0.2), intensity=0.2),
     )
     galaxy_1 = ag.Galaxy(
         redshift=1.0,
@@ -104,9 +106,9 @@ def test__simulate_imaging_data_and_fit__known_likelihood():
 
     plane = ag.Plane(galaxies=[galaxy_0, galaxy_1])
 
-    imaging = ag.SimulatorImaging(exposure_time=300.0, psf=psf, noise_seed=1)
+    simulator = ag.SimulatorImaging(exposure_time=300.0, psf=psf, noise_seed=1)
 
-    imaging = imaging.via_plane_from(plane=plane, grid=grid)
+    imaging = simulator.via_plane_from(plane=plane, grid=grid)
 
     mask = ag.Mask2D.circular(
         shape_native=imaging.image.shape_native, pixel_scales=0.2, radius=2.0
@@ -116,7 +118,28 @@ def test__simulate_imaging_data_and_fit__known_likelihood():
 
     fit = ag.FitImaging(dataset=masked_imaging, plane=plane)
 
-    assert fit.figure_of_merit == pytest.approx(621.5594864161, 1.0e-2)
+    assert fit.figure_of_merit == pytest.approx(532.19918562, 1.0e-2)
+
+    # Check that using a Basis gives the same result.
+
+    basis = ag.lp_basis.Basis(
+        light_profile_list=[
+            ag.lp.EllSersic(centre=(0.1, 0.1), intensity=0.1),
+            ag.lp.EllSersic(centre=(0.2, 0.2), intensity=0.2),
+        ]
+    )
+
+    galaxy_0 = ag.Galaxy(redshift=0.5, bulge=basis)
+
+    plane = ag.Plane(galaxies=[galaxy_0, galaxy_1])
+
+    imaging = simulator.via_plane_from(plane=plane, grid=grid)
+
+    masked_imaging = imaging.apply_mask(mask=mask)
+
+    fit = ag.FitImaging(dataset=masked_imaging, plane=plane)
+
+    assert fit.figure_of_merit == pytest.approx(532.19918562, 1.0e-2)
 
 
 def test__simulate_imaging_data_and_fit__linear_light_profiles_agree_with_standard_light_profiles():
