+++ autogalaxy/__init__.py
@@ -77,6 +77,7 @@ from .profiles import (
     light_and_mass_profiles as lmp,
     scaling_relations as sr,
 )
+from .profiles.light_profiles import basis as lp_basis
 from .profiles.light_profiles.light_profiles_linear import LightProfileLinearObjFunc
 from .profiles.light_profiles import light_profiles_init as lp_init
 from .profiles.light_profiles import light_profiles_linear as lp_linear
