### 2022-07-23 3a32248
baiss tests on fit interfereomter
