+++ test_autogalaxy/imaging/test_fit_imaging.py
@@ -151,7 +151,21 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_likelihood == pytest.approx(-75938.05, 1e-4)
+    assert fit.figure_of_merit == pytest.approx(-75938.05, 1.0e-4)
+
+    basis = ag.lp_basis.Basis(
+        light_profile_list=[
+            ag.lp.EllSersic(intensity=1.0),
+            ag.lp.EllSersic(intensity=1.0),
+        ]
+    )
+
+    g0 = ag.Galaxy(redshift=0.5, light_profile=basis)
+
+    plane = ag.Plane(redshift=0.5, galaxies=[g0])
+
+    fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
+
     assert fit.figure_of_merit == pytest.approx(-75938.05, 1.0e-4)
 
     pix = ag.pix.Rectangular(shape=(3, 3))
@@ -163,7 +177,6 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_evidence == pytest.approx(-22.9005, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-22.9005, 1.0e-4)
 
     galaxy_light = ag.Galaxy(redshift=0.5, light_profile=ag.lp.EllSersic(intensity=1.0))
@@ -172,7 +185,6 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_evidence == pytest.approx(-6840.5851, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-6840.5851, 1.0e-4)
 
     g0_linear_light = ag.Galaxy(
@@ -187,7 +199,6 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_likelihood == pytest.approx(-14.52327, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-14.52327, 1.0e-4)
 
     g0_operated_light = ag.Galaxy(
@@ -201,7 +212,6 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_likelihood == pytest.approx(-342374.9618, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-342374.9618, 1.0e-4)
 
     g0_linear_operated_light = ag.Galaxy(
@@ -217,14 +227,12 @@ def test__fit_figure_of_merit(masked_imaging_7x7):
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_likelihood == pytest.approx(-14.7237273, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-14.7237273, 1.0e-4)
 
     plane = ag.Plane(redshift=0.5, galaxies=[g0_linear_light, galaxy_pix])
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7, plane=plane)
 
-    assert fit.log_evidence == pytest.approx(-22.87827302, 1e-4)
     assert fit.figure_of_merit == pytest.approx(-22.87827302, 1.0e-4)
 
 
[…]
