### 2022-07-23 6d7ae0c
tests added to basis.py package
