+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -1,6 +1,7 @@
-import autogalaxy as ag
 import numpy as np
 import pytest
+
+import autogalaxy as ag
 
 from autogalaxy import exc
 
