### 2022-07-23 a9d4eb0
linear light profile list func now supports multiple pixels
