+++ autogalaxy/profiles/light_profiles/mock/mock_light_profile.py
@@ -1,6 +1,8 @@
 import numpy as np
 from typing import Optional
 
+import autoarray as aa
+
 from autogalaxy.profiles.light_profiles.light_profile_decorators import (
     check_operated_only,
 )
@@ -26,6 +28,7 @@ class MockLightProfile(ag.lp.LightProfile):
         self.value = value
         self.value1 = value1
 
+    @aa.grid_dec.grid_2d_to_structure
     @check_operated_only
     def image_2d_from(self, grid, operated_only: Optional[bool] = None):
 
