### 2022-07-23 cb390e1
lp_linear_func_list_galaxy_dict now receives input class
