### 2022-07-23 7145d9d
fixed test where basis was including normal light profiles in linear light profile func list object
