### 2022-07-23 51e573b
moved grids to constructor of PlaneToInversion
