+++ test_autogalaxy/plane/test_to_inversion.py
@@ -10,8 +10,8 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane)
 
-    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict_from(
-        source_grid_slim=1, source_blurring_grid_slim=2, convolver=3
+    lp_linear_func_galaxy_dict = (
+        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
     )
 
     assert lp_linear_func_galaxy_dict == {}
@@ -30,8 +30,8 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane)
 
-    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict_from(
-        source_grid_slim=1, source_blurring_grid_slim=2, convolver=3
+    lp_linear_func_galaxy_dict = (
+        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
     )
 
     lp_linear_func_list = list(lp_linear_func_galaxy_dict.keys())
@@ -52,8 +52,8 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane)
 
-    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict_from(
-        source_grid_slim=1, source_blurring_grid_slim=2, convolver=3
+    lp_linear_func_galaxy_dict = (
+        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
     )
 
     lp_linear_func_list = list(lp_linear_func_galaxy_dict.keys())
@@ -76,10 +76,9 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
 
     sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
-        grid=sub_grid_2d_7x7
     )
 
     assert (sparse_grid == np.array([[1.0, 1.0]])).all()
@@ -98,10 +97,9 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
 
     sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
-        grid=sub_grid_2d_7x7
     )
 
     assert (sparse_grid == np.array([[2.0, 2.0]])).all()
@@ -112,10 +110,9 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_no_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
 
     sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
-        grid=sub_grid_2d_7x7
[…]
