+++ test_autogalaxy/plane/test_to_inversion.py
@@ -76,10 +76,11 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
-
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
+    plane_to_inversion = ag.PlaneToInversion(
+        plane=plane, grid_pixelized=sub_grid_2d_7x7
     )
+
+    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
 
     assert (sparse_grid == np.array([[1.0, 1.0]])).all()
 
@@ -97,10 +98,11 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
-
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
+    plane_to_inversion = ag.PlaneToInversion(
+        plane=plane, grid_pixelized=sub_grid_2d_7x7
     )
+
+    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
 
     assert (sparse_grid == np.array([[2.0, 2.0]])).all()
 
@@ -110,10 +112,11 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_no_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
-
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from(
+    plane_to_inversion = ag.PlaneToInversion(
+        plane=plane, grid_pixelized=sub_grid_2d_7x7
     )
+
+    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
 
     assert sparse_grid is None
 
@@ -131,10 +134,11 @@ def test__mapper_galaxy_dict_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_no_pix, galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
-
-    mapper_galaxy_dict = plane_to_inversion.mapper_galaxy_dict_from(
+    plane_to_inversion = ag.PlaneToInversion(
+        plane=plane, grid_pixelized=sub_grid_2d_7x7
     )
+
+    mapper_galaxy_dict = plane_to_inversion.mapper_galaxy_dict_from()
 
     mapper_list = list(mapper_galaxy_dict.keys())
 
@@ -154,10 +158,11 @@ def test__mapper_galaxy_dict_from(sub_grid_2d_7x7):
         galaxies=[galaxy_no_pix, galaxy_pix, galaxy_no_pix, galaxy_pix_2], redshift=0.5
     )
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane, source_grid_pixelized_slim=sub_grid_2d_7x7)
-
-    mapper_galaxy_dict = plane_to_inversion.mapper_galaxy_dict_from(
+    plane_to_inversion = ag.PlaneToInversion(
[…]

[…]
