### 2022-07-23 a99ee68
moved all data into PlaneToInversion
