+++ autogalaxy/plane/to_inversion.py
@@ -2,8 +2,7 @@ from typing import Dict, List, Optional, Type, Union
 
 import autoarray as aa
 
-from autoarray.inversion.inversion.factory import inversion_imaging_unpacked_from
-from autoarray.inversion.inversion.factory import inversion_interferometer_unpacked_from
+from autoarray.inversion.inversion.factory import inversion_unpacked_from
 from autogalaxy.profiles.light_profiles.light_profiles_linear import (
     LightProfileLinearObjFuncList,
 )
@@ -11,24 +10,28 @@ from autogalaxy.profiles.light_profiles.basis import Basis
 from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
 from autogalaxy.galaxy.galaxy import Galaxy
 
-from autogalaxy import exc
-
 
 class PlaneToInversion:
     def __init__(
         self,
         plane: "Plane",
-        grid: Optional[aa.type.Grid2DLike] = None,
-        blurring_grid: Optional[aa.type.Grid2DLike] = None,
-        convolver: Optional[aa.Convolver] = None,
-        grid_pixelized: Optional[aa.type.Grid2DLike] = None,
+        dataset: Optional[Union[aa.Imaging, aa.Interferometer]] = None,
+        data: Optional[Union[aa.Array2D, aa.Visibilities]] = None,
+        noise_map: Optional[Union[aa.Array2D, aa.VisibilitiesNoiseMap]] = None,
+        w_tilde: Optional[Union[aa.WTildeImaging, aa.WTildeInterferometer]] = None,
+        settings_pixelization=aa.SettingsPixelization(),
+        settings_inversion: aa.SettingsInversion = aa.SettingsInversion(),
+        preloads=aa.Preloads(),
     ):
 
         self.plane = plane
-        self.grid = grid
-        self.blurring_grid = blurring_grid
-        self.convolver = convolver
-        self.grid_pixelized = grid_pixelized
+        self.dataset = dataset
+        self.data = data
+        self.noise_map = noise_map
+        self.w_tilde = w_tilde
+        self.settings_pixelization = settings_pixelization
+        self.settings_inversion = settings_inversion
+        self.preloads = preloads
 
     def cls_light_profile_func_list_galaxy_dict_from(
         self, cls: Type
@@ -56,9 +59,9 @@ class PlaneToInversion:
                     if len(light_profile_list) > 0:
 
                         lp_linear_func = LightProfileLinearObjFuncList(
-                            grid=self.grid,
-                            blurring_grid=self.blurring_grid,
-                            convolver=self.convolver,
+                            grid=self.dataset.grid,
+                            blurring_grid=self.dataset.blurring_grid,
+                            convolver=self.dataset.convolver,
                             light_profile_list=light_profile_list,
                         )
 
@@ -83,18 +86,16 @@ class PlaneToInversion:
             **lp_basis_func_list_galaxy_dict,
         }
 
-    def sparse_image_plane_grid_list_from(
-        self, settings_pixelization=aa.SettingsPixelization()
-    ) -> Optional[List[aa.type.Grid2DLike]]:
+    def sparse_image_plane_grid_list_from(self,) -> Optional[List[aa.type.Grid2DLike]]:
 
         if not self.plane.has(cls=aa.pix.Pixelization):
             return None
 
         return [
             pixelization.data_pixelization_grid_from(
-                data_grid_slim=self.grid_pixelized,
+                data_grid_slim=self.dataset.grid_pixelized,
                 hyper_image=hyper_galaxy_image,
-                settings=settings_pixelization,
+                settings=self.settings_pixelization,
[…]
