+++ test_autogalaxy/plane/test_to_inversion.py
@@ -4,11 +4,11 @@ import pytest
 import autogalaxy as ag
 
 
-def test__lp_linear_func_list_galaxy_dict_from(lp_0):
+def test__lp_linear_func_list_galaxy_dict_from(lp_0, masked_imaging_7x7):
 
     plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5)], redshift=None)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
     lp_linear_func_galaxy_dict = (
         plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
@@ -28,7 +28,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
 
     plane = ag.Plane(galaxies=[galaxy_0, galaxy_1], redshift=None)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
     lp_linear_func_galaxy_dict = (
         plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
@@ -50,7 +50,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
 
     plane = ag.Plane(galaxies=[galaxy_0, galaxy_1], redshift=None)
 
-    plane_to_inversion = ag.PlaneToInversion(plane=plane)
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
     lp_linear_func_galaxy_dict = (
         plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
@@ -66,7 +66,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0):
     assert lp_linear_func_list[1].light_profile_list[1] == lp_linear_1
 
 
-def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
+def test__sparse_image_plane_grid_list_from(masked_imaging_7x7):
 
     galaxy_pix = ag.Galaxy(
         redshift=0.5,
@@ -76,9 +76,7 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(
-        plane=plane, grid_pixelized=sub_grid_2d_7x7
-    )
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
     sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
 
@@ -98,9 +96,7 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(
-        plane=plane, grid_pixelized=sub_grid_2d_7x7
-    )
+    plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
     sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
 
@@ -112,21 +108,19 @@ def test__sparse_image_plane_grid_list_from(sub_grid_2d_7x7):
 
     plane = ag.Plane(galaxies=[galaxy_no_pix], redshift=0.5)
 
-    plane_to_inversion = ag.PlaneToInversion(
[…]
