### 2022-07-23 6a6dfa9
PlaneToInversion things made to properties
