+++ test_autogalaxy/plane/test_to_inversion.py
@@ -4,15 +4,13 @@ import pytest
 import autogalaxy as ag
 
 
-def test__lp_linear_func_list_galaxy_dict_from(lp_0, masked_imaging_7x7):
+def test__lp_linear_func_list_galaxy_dict(lp_0, masked_imaging_7x7):
 
     plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5)], redshift=None)
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
-    lp_linear_func_galaxy_dict = (
-        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
-    )
+    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict
 
     assert lp_linear_func_galaxy_dict == {}
 
@@ -30,9 +28,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0, masked_imaging_7x7):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
-    lp_linear_func_galaxy_dict = (
-        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
-    )
+    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict
 
     lp_linear_func_list = list(lp_linear_func_galaxy_dict.keys())
 
@@ -52,9 +48,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0, masked_imaging_7x7):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
-    lp_linear_func_galaxy_dict = (
-        plane_to_inversion.lp_linear_func_list_galaxy_dict_from()
-    )
+    lp_linear_func_galaxy_dict = plane_to_inversion.lp_linear_func_list_galaxy_dict
 
     lp_linear_func_list = list(lp_linear_func_galaxy_dict.keys())
 
@@ -66,7 +60,7 @@ def test__lp_linear_func_list_galaxy_dict_from(lp_0, masked_imaging_7x7):
     assert lp_linear_func_list[1].light_profile_list[1] == lp_linear_1
 
 
-def test__sparse_image_plane_grid_list_from(masked_imaging_7x7):
+def test__sparse_image_plane_grid_list(masked_imaging_7x7):
 
     galaxy_pix = ag.Galaxy(
         redshift=0.5,
@@ -78,7 +72,7 @@ def test__sparse_image_plane_grid_list_from(masked_imaging_7x7):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
+    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list
 
     assert (sparse_grid == np.array([[1.0, 1.0]])).all()
 
@@ -98,7 +92,7 @@ def test__sparse_image_plane_grid_list_from(masked_imaging_7x7):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
-    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list_from()
+    sparse_grid = plane_to_inversion.sparse_image_plane_grid_list
 
     assert (sparse_grid == np.array([[2.0, 2.0]])).all()
 
@@ -110,12 +104,12 @@ def test__sparse_image_plane_grid_list_from(masked_imaging_7x7):
 
     plane_to_inversion = ag.PlaneToInversion(plane=plane, dataset=masked_imaging_7x7)
 
[…]
