+++ autogalaxy/interferometer/fit_interferometer.py
@@ -155,7 +155,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
         """
         if self.perform_inversion:
 
-            return self.plane_to_inversion.inversion_from()
+            return self.plane_to_inversion.inversion
 
     @property
     def model_data(self) -> aa.Visibilities:
