+++ autogalaxy/imaging/fit_imaging.py
@@ -173,7 +173,7 @@ class FitImaging(aa.FitImaging, AbstractFit):
 
         if self.perform_inversion:
 
-            return self.plane_to_inversion.inversion_from()
+            return self.plane_to_inversion.inversion
 
     @property
     def model_data(self) -> aa.Array2D:
