+++ autogalaxy/plane/to_inversion.py
@@ -1,5 +1,7 @@
 from typing import Dict, List, Optional, Type, Union
 
+from autoconf import cached_property
+
 import autoarray as aa
 
 from autoarray.inversion.inversion.factory import inversion_unpacked_from
@@ -19,6 +21,9 @@ class PlaneToInversion:
         data: Optional[Union[aa.Array2D, aa.Visibilities]] = None,
         noise_map: Optional[Union[aa.Array2D, aa.VisibilitiesNoiseMap]] = None,
         w_tilde: Optional[Union[aa.WTildeImaging, aa.WTildeInterferometer]] = None,
+        grid: Optional[aa.type.Grid2DLike] = None,
+        blurring_grid: Optional[aa.type.Grid2DLike] = None,
+        grid_pixelized: Optional[aa.type.Grid2DLike] = None,
         settings_pixelization=aa.SettingsPixelization(),
         settings_inversion: aa.SettingsInversion = aa.SettingsInversion(),
         preloads=aa.Preloads(),
@@ -29,8 +34,31 @@ class PlaneToInversion:
         self.data = data
         self.noise_map = noise_map
         self.w_tilde = w_tilde
+
+        if grid is not None:
+            self.grid = grid
+        elif dataset is not None:
+            self.grid = dataset.grid
+        else:
+            self.grid = None
+
+        if blurring_grid is not None:
+            self.blurring_grid = blurring_grid
+        elif dataset is not None:
+            self.blurring_grid = dataset.blurring_grid
+        else:
+            self.blurring_grid = None
+
+        if grid_pixelized is not None:
+            self.grid_pixelized = grid_pixelized
+        elif dataset is not None:
+            self.grid_pixelized = dataset.grid_pixelized
+        else:
+            self.grid_pixelized = None
+
         self.settings_pixelization = settings_pixelization
         self.settings_inversion = settings_inversion
+
         self.preloads = preloads
 
     def cls_light_profile_func_list_galaxy_dict_from(
@@ -59,8 +87,8 @@ class PlaneToInversion:
                     if len(light_profile_list) > 0:
 
                         lp_linear_func = LightProfileLinearObjFuncList(
-                            grid=self.dataset.grid,
-                            blurring_grid=self.dataset.blurring_grid,
+                            grid=self.grid,
+                            blurring_grid=self.blurring_grid,
                             convolver=self.dataset.convolver,
                             light_profile_list=light_profile_list,
                         )
@@ -69,7 +97,8 @@ class PlaneToInversion:
 
         return lp_linear_func_galaxy_dict
 
-    def lp_linear_func_list_galaxy_dict_from(
+    @cached_property
+    def lp_linear_func_list_galaxy_dict(
         self,
     ) -> Dict[LightProfileLinearObjFuncList, Galaxy]:
 
@@ -86,14 +115,15 @@ class PlaneToInversion:
             **lp_basis_func_list_galaxy_dict,
         }
 
-    def sparse_image_plane_grid_list_from(self,) -> Optional[List[aa.type.Grid2DLike]]:
+    @cached_property
+    def sparse_image_plane_grid_list(self,) -> Optional[List[aa.type.Grid2DLike]]:
 
         if not self.plane.has(cls=aa.pix.Pixelization):
             return None
 
         return [
             pixelization.data_pixelization_grid_from(
-                data_grid_slim=self.dataset.grid_pixelized,
+                data_grid_slim=self.grid_pixelized,
                 hyper_image=hyper_galaxy_image,
                 settings=self.settings_pixelization,
             )
@@ -109,26 +139,25 @@ class PlaneToInversion:
[…]
