### 2022-07-23 60ae358
light_profile= in galaxy replaced with bulge= for test galaxies
