+++ autogalaxy/imaging/fit_imaging.py
@@ -172,7 +172,6 @@ class FitImaging(aa.FitImaging, AbstractFit):
         """
 
         if self.perform_inversion:
-
             return self.plane_to_inversion.inversion
 
     @property

[…]
