### 2022-07-23 e74c45e
fixed bugs in fit imaging where inversion was not performed and model data didnt have basis added to it
