+++ autogalaxy/imaging/fit_imaging.py
@@ -185,9 +185,7 @@ class FitImaging(aa.FitImaging, AbstractFit):
         If a inversion is included it is the sum of this image and the inversion's reconstruction of the image.
         """
 
-        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has(
-            cls=LightProfileLinear
-        ):
+        if self.perform_inversion:
 
             return self.blurred_image + self.inversion.mapped_reconstructed_data
 
