+++ autogalaxy/interferometer/fit_interferometer.py
@@ -168,9 +168,7 @@ class FitInterferometer(aa.FitInterferometer, AbstractFit):
         If a inversion is included it is the sum of these visibilities and the inversion's reconstructed visibilities.
         """
 
-        if self.plane.has(cls=aa.pix.Pixelization) or self.plane.has(
-            cls=LightProfileLinear
-        ):
+        if self.perform_inversion:
 
             return self.profile_visibilities + self.inversion.mapped_reconstructed_data
 
