### 2022-07-23 4c8afab
test light_profile_linear_intensity_dict with Basis
