### 2022-07-23 78b3076
can now regularization Basis object
