### 2022-07-24 c79438e
added unit tests to set_upper_limit_of_pixelization_pixels_prior
