+++ test_autogalaxy/analysis/test_model_util.py
@@ -74,6 +74,57 @@ def test__pixelization_from_model():
 #     assert galaxies.has_model(cls=ag.pix.Pixelization) is True
 
 
+def test__set_upper_limit_of_pixelization_pixels_prior():
+
+    pixelization = af.Model(ag.pix.DelaunayBrightnessImage)
+
+    pixelization.pixels = af.UniformPrior(lower_limit=0.0, upper_limit=10.0)
+
+    galaxies = af.Collection(
+        source=ag.Galaxy(
+            redshift=0.5, pixelization=pixelization, regularization=ag.reg.Constant
+        )
+    )
+
+    model = af.Collection(galaxies=galaxies)
+
+    ag.util.model.set_upper_limit_of_pixelization_pixels_prior(
+        model=model, pixels_in_mask=12
+    )
+
+    assert model.galaxies.source.pixelization.pixels.upper_limit == pytest.approx(
+        10, 1.0e-4
+    )
+
+    ag.util.model.set_upper_limit_of_pixelization_pixels_prior(
+        model=model, pixels_in_mask=8
+    )
+
+    assert model.galaxies.source.pixelization.pixels.upper_limit == pytest.approx(
+        8, 1.0e-4
+    )
+
+    pixelization = af.Model(ag.pix.VoronoiBrightnessImage)
+
+    pixelization.pixels = af.UniformPrior(lower_limit=0.0, upper_limit=100.0)
+
+    galaxies = af.Collection(
+        source=ag.Galaxy(
+            redshift=0.5, pixelization=pixelization, regularization=ag.reg.Constant
+        )
+    )
+
+    model = af.Collection(galaxies=galaxies)
+
+    ag.util.model.set_upper_limit_of_pixelization_pixels_prior(
+        model=model, pixels_in_mask=6
+    )
+
+    assert model.galaxies.source.pixelization.pixels.upper_limit == pytest.approx(
+        6, 1.0e-4
+    )
+
+
 def test__hyper_model_noise_from():
     model = af.Collection(
         galaxies=af.Collection(
