### 2022-07-24 b2feecb
model pixelization pixels check done in modify_before_fit
