### 2022-07-24 7abca6c
now also update the lower limit
