+++ autogalaxy/analysis/analysis.py
@@ -1,4 +1,3 @@
-import copy
 import json
 import logging
 import numpy as np
