### 2022-07-24 318cfc5
pixelization being chekced mustg have pixels
