+++ autogalaxy/analysis/model_util.py
@@ -136,17 +136,19 @@ def set_upper_limit_of_pixelization_pixels_prior(
 
                                 lower_limit = pixelization.pixels.lower_limit
 
-                                log_str = "MODIFY BEFORE FIT -  A pixelization's pixel UniformPrior upper limit" \
-                                          "was greater than the number of pixels in the mask. It has been " \
-                                          "reduced to the number of pixels in the mask.\,"
+                                log_str = (
+                                    "MODIFY BEFORE FIT -  A pixelization's pixel UniformPrior upper limit"
+                                    "was greater than the number of pixels in the mask. It has been "
+                                    "reduced to the number of pixels in the mask.\,"
+                                )
 
                                 if lower_limit > pixels_in_mask:
 
                                     lower_limit = pixels_in_mask - 10
 
                                     logger.info(
-                                        log_str +
-                                        "MODIFY BEFORE FIT - The pixelization's pixel UniformPrior lower_limit was "
+                                        log_str
+                                        + "MODIFY BEFORE FIT - The pixelization's pixel UniformPrior lower_limit was "
                                         "also above the number of pixels in the mask, and has been reduced"
                                         "to the number of pixels in the mask minus 10."
                                     )
