+++ test_autogalaxy/analysis/test_model_util.py
@@ -44,36 +44,6 @@ def test__pixelization_from_model():
     assert type(pixelization) == type(ag.pix.Rectangular)
 
 
-# def test__has_pixelization():
-#     galaxies = af.Collection(galaxy=af.Model(ag.Galaxy, redshift=0.5))
-#
-#     assert galaxies.has_model(cls=ag.pix.Pixelization) is False
-#
-#     galaxies = af.Collection(
-#         galaxy=af.Model(
-#             ag.Galaxy,
-#             redshift=0.5,
-#             pixelization=ag.pix.Rectangular(),
-#             regularization=ag.reg.Constant(),
-#         )
-#     )
-#
-#     assert galaxies.has_instance(cls=ag.pix.Pixelization) is True
-#     assert galaxies.has_model(cls=ag.pix.Pixelization) is False
-#
-#     galaxies = af.Collection(
-#         galaxy=af.Model(
-#             ag.Galaxy,
-#             redshift=0.5,
-#             pixelization=ag.pix.Rectangular,
-#             regularization=ag.reg.Constant,
-#         )
-#     )
-#
-#     assert galaxies.has_instance(cls=ag.pix.Pixelization) is False
-#     assert galaxies.has_model(cls=ag.pix.Pixelization) is True
-
-
 def test__set_upper_limit_of_pixelization_pixels_prior():
 
     pixelization = af.Model(ag.pix.DelaunayBrightnessImage)
@@ -115,7 +85,7 @@ def test__set_upper_limit_of_pixelization_pixels_prior():
     )
 
     assert model.galaxies.source.pixelization.pixels.lower_limit == pytest.approx(
-        2, 1.0e-4
+        -7, 1.0e-4
     )
     assert model.galaxies.source.pixelization.pixels.upper_limit == pytest.approx(
         3, 1.0e-4
