### 2022-07-24 5c44265
did as much pixelizatio -> mesh as possible, but will finish once pixelization / mesh separation is clear
