+++ autogalaxy/__init__.py
@@ -5,7 +5,7 @@ from autoarray.dataset.interferometer import Interferometer
 from autoarray.dataset.interferometer import SettingsInterferometer
 from autoarray.instruments import acs
 from autoarray.instruments import euclid
-from autoarray.inversion import mesh
+from autoarray.inversion.pixelization import mesh
 from autoarray.inversion import regularization as reg
 from autoarray.inversion.inversion.settings import SettingsInversion
 from autoarray.inversion.inversion.factory import inversion_from as Inversion
@@ -15,8 +15,8 @@ from autoarray.inversion.inversion.factory import (
 from autoarray.inversion.inversion.factory import (
     inversion_interferometer_unpacked_from as InversionInterferometer,
 )
-from autoarray.inversion.mappers.factory import mapper_from as Mapper
-from autoarray.inversion.mesh.settings import SettingsPixelization
+from autoarray.inversion.pixelization.mappers.factory import mapper_from as Mapper
+from autoarray.inversion.pixelization.settings import SettingsPixelization
 from autoarray.mask.mask_1d import Mask1D
 from autoarray.mask.mask_2d import Mask2D
 from autoarray.operators.convolver import Convolver
