### 2022-07-24 22080c9
removed pixelization only exception from galaxy
