+++ test_autogalaxy/galaxy/test_galaxy.py
@@ -413,11 +413,6 @@ def test__image_2d_from__does_not_include_linear_light_profiles(sub_grid_2d_7x7,
     assert (image == lp_image).all()
 
 
-def test__only_pixelization_raises_error():
-    with pytest.raises(exc.GalaxyException):
-        ag.Galaxy(redshift=0.5, pixelization=ag.mesh.Rectangular())
-
-
 def test__light_profile_2d_quantity_from_grid__symmetric_profiles_give_symmetric_results():
     lp_0 = ag.lp.EllSersic(centre=(0.0, 0.0), intensity=1.0)
 
