### 2022-07-24 ffbdebe
fixed plane_to_inversion.sparse_image_plane_grid_list test
