### 2022-07-24 fe96b52
fixed mapper_galaxy_dcit test
