+++ autogalaxy/abstract_fit.py
@@ -35,9 +35,7 @@ class AbstractFit:
         """
         # TODO : When we add regularization to basis need to change this to reflect mappers.
 
-        return len(
-            list(filter(None, self.model_obj.cls_list_from(cls=aa.mesh.Mesh)))
-        )
+        return len(list(filter(None, self.model_obj.cls_list_from(cls=aa.mesh.Mesh))))
 
     @property
     def perform_inversion(self) -> bool:
