+++ autogalaxy/plot/mat_wrap/include.py
@@ -33,9 +33,9 @@ class Include2D(aplt.Include2D):
         critical_curves=None,
         caustics=None,
         multiple_images=None,
-        mapper_source_pixelization_grid: Optional[bool] = None,
+        mapper_source_mesh_grid: Optional[bool] = None,
         mapper_source_grid_slim: Optional[bool] = None,
-        mapper_data_pixelization_grid=None,
+        mapper_data_mesh_grid=None,
     ):
 
         super().__init__(
@@ -43,9 +43,9 @@ class Include2D(aplt.Include2D):
             mask=mask,
             border=border,
             grid=grid,
-            mapper_source_pixelization_grid=mapper_source_pixelization_grid,
+            mapper_source_mesh_grid=mapper_source_mesh_grid,
             mapper_source_grid_slim=mapper_source_grid_slim,
-            mapper_data_pixelization_grid=mapper_data_pixelization_grid,
+            mapper_data_mesh_grid=mapper_data_mesh_grid,
         )
 
         self._positions = positions
