+++ test_autogalaxy/plot/mat_wrap/config/visualize/include.ini
@@ -8,8 +8,8 @@ caustics=False
 multiple_images=False
 light_profile_centres=True
 mass_profile_centres=True
-mapper_data_pixelization_grid=True
-mapper_source_pixelization_grid=True
+mapper_data_mesh_grid=True
+mapper_source_mesh_grid=True
 mapper_source_grid_slim=False
 parallel_overscan=True
 serial_prescan=True
