+++ test_autogalaxy/interferometer/test_fit_interferometer.py
@@ -67,10 +67,12 @@ def test__fit_figure_of_merit(interferometer_7):
     assert fit.perform_inversion is False
     assert fit.figure_of_merit == pytest.approx(-2398107.3849, 1.0e-4)
 
-    pix = ag.mesh.Rectangular(shape=(3, 3))
-    reg = ag.reg.Constant(coefficient=0.01)
+    pixelization = ag.Pixelization(
+        mesh=ag.mesh.Rectangular(shape=(3, 3)),
+        regularization=ag.reg.Constant(coefficient=0.01),
+    )
 
-    g0 = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
+    g0 = ag.Galaxy(redshift=0.5, pixelization=pixelization)
 
     plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5), g0])
 
@@ -85,9 +87,12 @@ def test__fit_figure_of_merit(interferometer_7):
 
     galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.EllSersic(intensity=1.0))
 
-    pix = ag.mesh.Rectangular(shape=(3, 3))
-    reg = ag.reg.Constant(coefficient=1.0)
-    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
+    pixelization = ag.Pixelization(
+        mesh=ag.mesh.Rectangular(shape=(3, 3)),
+        regularization=ag.reg.Constant(coefficient=1.0),
+    )
+
+    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pixelization)
 
     plane = ag.Plane(redshift=0.5, galaxies=[galaxy_light, galaxy_pix])
 
@@ -168,10 +173,12 @@ def test__fit_figure_of_merit__include_hyper_methods(interferometer_7):
 
     assert fit.noise_map == pytest.approx(interferometer_7.noise_map, 1.0e-4)
 
-    pix = ag.mesh.Rectangular(shape=(3, 3))
-    reg = ag.reg.Constant(coefficient=0.01)
+    pixelization = ag.Pixelization(
+        mesh=ag.mesh.Rectangular(shape=(3, 3)),
+        regularization=ag.reg.Constant(coefficient=0.01),
+    )
 
-    g0 = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
+    g0 = ag.Galaxy(redshift=0.5, pixelization=pixelization)
 
     plane = ag.Plane(galaxies=[ag.Galaxy(redshift=0.5), g0])
 
@@ -188,9 +195,12 @@ def test__fit_figure_of_merit__include_hyper_methods(interferometer_7):
 
     galaxy_light = ag.Galaxy(redshift=0.5, bulge=ag.lp.EllSersic(intensity=1.0))
 
-    pix = ag.mesh.Rectangular(shape=(3, 3))
-    reg = ag.reg.Constant(coefficient=1.0)
-    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pix, regularization=reg)
+    pixelization = ag.Pixelization(
+        mesh=ag.mesh.Rectangular(shape=(3, 3)),
+        regularization=ag.reg.Constant(coefficient=1.0),
+    )
+
+    galaxy_pix = ag.Galaxy(redshift=0.5, pixelization=pixelization)
 
     plane = ag.Plane(redshift=0.5, galaxies=[galaxy_light, galaxy_pix])
 
@@ -210,10 +220,12 @@ def test___fit_figure_of_merit__different_settings(
     interferometer_7, interferometer_7_lop
 ):
 
-    pix = ag.mesh.Rectangular(shape=(3, 3))
[…]

[…]
