+++ test_autogalaxy/analysis/test_model_util.py
@@ -4,44 +4,46 @@ import autofit as af
 import autogalaxy as ag
 
 
-def test__pixelization_from_model():
+def test__mesh_list_from_model():
 
     galaxies = af.Collection(galaxy=af.Model(ag.Galaxy, redshift=0.5))
 
-    pixelization = ag.util.model.pixelization_from(
-        model=af.Collection(galaxies=galaxies)
-    )
+    mesh_list = ag.util.model.mesh_list_from(model=af.Collection(galaxies=galaxies))
 
-    assert pixelization is None
+    assert mesh_list == []
+
+    pixelization_0 = ag.Pixelization(mesh=ag.mesh.Rectangular())
+
+    pixelization_1 = ag.Pixelization(mesh=ag.mesh.VoronoiMagnification())
 
     galaxies = af.Collection(
         galaxy=ag.Galaxy(
-            redshift=0.5,
-            pixelization=ag.mesh.Rectangular(),
-            regularization=ag.reg.Constant(),
+            redshift=0.5, pixelization_0=pixelization_0, pixelization_1=pixelization_1
         )
     )
 
-    pixelization = ag.util.model.pixelization_from(
-        model=af.Collection(galaxies=galaxies)
-    )
+    mesh_list = ag.util.model.mesh_list_from(model=af.Collection(galaxies=galaxies))
 
-    assert isinstance(pixelization, ag.mesh.Rectangular)
+    assert isinstance(mesh_list[0], ag.mesh.Rectangular)
+    assert isinstance(mesh_list[1], ag.mesh.VoronoiMagnification)
+
+    pixelization_0 = af.Model(ag.Pixelization, mesh=ag.mesh.Rectangular)
+
+    pixelization_1 = af.Model(ag.Pixelization, mesh=ag.mesh.VoronoiMagnification)
 
     galaxies = af.Collection(
         galaxy=af.Model(
             ag.Galaxy,
             redshift=0.5,
-            pixelization=ag.mesh.Rectangular,
-            regularization=ag.reg.Constant,
+            pixelization_0=pixelization_0,
+            pixelization_1=pixelization_1,
         )
     )
 
-    pixelization = ag.util.model.pixelization_from(
-        model=af.Collection(galaxies=galaxies)
-    )
+    mesh_list = ag.util.model.mesh_list_from(model=af.Collection(galaxies=galaxies))
 
-    assert type(pixelization) == type(ag.mesh.Rectangular)
+    assert isinstance(mesh_list[0], ag.mesh.Rectangular)
+    assert isinstance(mesh_list[1], ag.mesh.VoronoiMagnification)
 
 
 def test__set_upper_limit_of_pixelization_pixels_prior():
