### 2022-07-25 a79c7b4
fixed more model_util tests 0_0
