+++ autogalaxy/imaging/plot/fit_imaging_plotters.py
@@ -202,7 +202,7 @@ class FitImagingPlotter(Plotter):
             )
             self.figures_2d_of_galaxies(galaxy_index=galaxy_index, model_image=True)
 
-            if self.plane.has(cls=aa.mesh.Mesh):
+            if self.plane.has(cls=aa.Pixelization):
                 self.inversion_plotter.figures_2d_of_mapper(
                     mapper_index=0, reconstruction=True
                 )
