### 2022-07-25 3c70389
more of a horror show in model_util
