+++ test_autogalaxy/analysis/test_analysis.py
@@ -76,15 +76,12 @@ def test__modify_before_fit__kmeans_pixelization_upper_limit_ajusted_based_on_ma
     masked_imaging_7x7
 ):
 
-    pixelization = af.Model(ag.mesh.DelaunayBrightnessImage)
+    mesh = af.Model(ag.mesh.DelaunayBrightnessImage)
+    mesh.pixels = af.UniformPrior(lower_limit=0.0, upper_limit=100.0)
 
-    pixelization.pixels = af.UniformPrior(lower_limit=0.0, upper_limit=100.0)
+    pixelization = af.Model(ag.Pixelization, mesh=mesh)
 
-    galaxies = af.Collection(
-        source=ag.Galaxy(
-            redshift=0.5, pixelization=pixelization, regularization=ag.reg.Constant
-        )
-    )
+    galaxies = af.Collection(source=ag.Galaxy(redshift=0.5, pixelization=pixelization))
 
     model = af.Collection(galaxies=galaxies)
 
@@ -92,6 +89,6 @@ def test__modify_before_fit__kmeans_pixelization_upper_limit_ajusted_based_on_ma
 
     analysis.modify_before_fit(paths=af.DirectoryPaths(), model=model)
 
-    assert model.galaxies.source.pixelization.pixels.upper_limit == pytest.approx(
+    assert model.galaxies.source.pixelization.mesh.pixels.upper_limit == pytest.approx(
         9, 1.0e-4
     )
