+++ autogalaxy/analysis/model_util.py
@@ -141,52 +141,54 @@ def set_upper_limit_of_pixelization_pixels_prior(
 
     for galaxy in model.galaxies:
 
-        key_list = galaxy_key_list_dict[galaxy]
+       if galaxy in galaxy_key_list_dict:
 
-        for key in key_list:
+            key_list = galaxy_key_list_dict[galaxy]
 
-            pixelization = getattr(galaxy, key)
+            for key in key_list:
 
-            if pixelization is not None:
+                pixelization = getattr(galaxy, key)
 
-                mesh = pixelization.mesh
+                if pixelization is not None:
 
-                if hasattr(mesh, "pixels"):
+                    mesh = pixelization.mesh
 
-                    if hasattr(mesh.pixels, "upper_limit"):
+                    if hasattr(mesh, "pixels"):
 
-                        if pixels_in_mask < mesh.pixels.upper_limit:
+                        if hasattr(mesh.pixels, "upper_limit"):
 
-                            if (
-                                mesh.cls is aa.mesh.DelaunayBrightnessImage
-                                or aa.mesh.VoronoiBrightnessImage
-                                or aa.mesh.VoronoiNNBrightnessImage
-                            ):
+                            if pixels_in_mask < mesh.pixels.upper_limit:
 
-                                lower_limit = mesh.pixels.lower_limit
+                                if (
+                                    mesh.cls is aa.mesh.DelaunayBrightnessImage
+                                    or aa.mesh.VoronoiBrightnessImage
+                                    or aa.mesh.VoronoiNNBrightnessImage
+                                ):
 
-                                log_str = (
-                                    "MODIFY BEFORE FIT -  A pixelization mesh's pixel UniformPrior upper limit"
-                                    "was greater than the number of pixels in the mask. It has been "
-                                    "reduced to the number of pixels in the mask.\,"
-                                )
+                                    lower_limit = mesh.pixels.lower_limit
 
-                                if lower_limit > pixels_in_mask:
-
-                                    lower_limit = pixels_in_mask - 10
-
-                                    logger.info(
-                                        log_str
-                                        + "MODIFY BEFORE FIT - The pixelization's mesh's pixel UniformPrior lower_limit was "
-                                        "also above the number of pixels in the mask, and has been reduced"
-                                        "to the number of pixels in the mask minus 10."
+                                    log_str = (
+                                        "MODIFY BEFORE FIT -  A pixelization mesh's pixel UniformPrior upper limit"
+                                        "was greater than the number of pixels in the mask. It has been "
+                                        "reduced to the number of pixels in the mask.\,"
                                     )
-                                else:
-                                    logger.info(log_str)
 
-                                mesh.pixels = af.UniformPrior(
-                                    lower_limit=lower_limit, upper_limit=pixels_in_mask
-                                )
+                                    if lower_limit > pixels_in_mask:
+
+                                        lower_limit = pixels_in_mask - 10
+
+                                        logger.info(
+                                            log_str
+                                            + "MODIFY BEFORE FIT - The pixelization's mesh's pixel UniformPrior lower_limit was "
+                                            "also above the number of pixels in the mask, and has been reduced"
+                                            "to the number of pixels in the mask minus 10."
+                                        )
+                                    else:
+                                        logger.info(log_str)
+
+                                    mesh.pixels = af.UniformPrior(
+                                        lower_limit=lower_limit, upper_limit=pixels_in_mask
+                                    )
 
 
 def clean_model_of_hyper_images(model):
