+++ test_autogalaxy/interferometer/test_simulate_and_fit_interferometer.py
@@ -117,11 +117,13 @@ def test__simulate_interferometer_data_and_fit__known_likelihood():
         light=ag.lp.EllSersic(centre=(0.1, 0.1), intensity=0.1),
         mass=ag.mp.EllIsothermal(centre=(0.1, 0.1), einstein_radius=1.8),
     )
-    galaxy_1 = ag.Galaxy(
-        redshift=1.0,
-        pixelization=ag.mesh.Rectangular(shape=(16, 16)),
+
+    pixelization = ag.Pixelization(
+        mesh=ag.mesh.Rectangular(shape=(16, 16)),
         regularization=ag.reg.Constant(coefficient=(1.0)),
     )
+
+    galaxy_1 = ag.Galaxy(redshift=1.0, pixelization=pixelization)
 
     plane = ag.Plane(galaxies=[galaxy_0, galaxy_1])
 
