+++ test_autogalaxy/analysis/test_result.py
@@ -48,10 +48,7 @@ def test__results_include_pixelization__available_as_property(analysis_imaging_7
 
     pixelization = ag.m.MockPixelization(mapper=1)
 
-    source = ag.Galaxy(
-        redshift=1.0,
-        pixelization=pixelization
-    )
+    source = ag.Galaxy(redshift=1.0, pixelization=pixelization)
 
     max_log_likelihood_plane = ag.Plane(galaxies=[source])
 
@@ -61,5 +58,7 @@ def test__results_include_pixelization__available_as_property(analysis_imaging_7
         samples=samples, analysis=analysis_imaging_7x7, model=None
     )
 
-    assert isinstance(result.cls_list_from(cls=ag.Pixelization)[0], ag.m.MockPixelization)
+    assert isinstance(
+        result.cls_list_from(cls=ag.Pixelization)[0], ag.m.MockPixelization
+    )
     assert result.cls_list_from(cls=ag.Pixelization)[0].mapper == 1
