### 2022-07-25 Jammy2211/PyAutoGalaxy#35 @Jammy2211
Pushed what I think is a fix, feel free to reopn this issue if theres still failing tests.
