+++ autogalaxy/imaging/plot/fit_imaging_plotters.py
@@ -203,8 +203,8 @@ class FitImagingPlotter(Plotter):
             self.figures_2d_of_galaxies(galaxy_index=galaxy_index, model_image=True)
 
             if self.plane.has(cls=aa.Pixelization):
-                self.inversion_plotter.figures_2d_of_mapper(
-                    mapper_index=0, reconstruction=True
+                self.inversion_plotter.figures_2d_of_pixelization(
+                    pixelization_index=0, reconstruction=True
                 )
 
             self.mat_plot_2d.output.subplot_to_figure(
