+++ autogalaxy/quantity/dataset_quantity.py
@@ -30,22 +30,22 @@ class SettingsQuantity(AbstractSettingsDataset):
         grid_class : ag.Grid2D
             The type of grid used to create the image from the `Galaxy` and `Plane`. The options are `Grid2D`,
             and `Grid2DIterate` (see the `Grid2D` documentation for a description of these options).
-        grid_pixelized_class : ag.Grid2D
+        grid_pixelizaiton_class : ag.Grid2D
             The type of grid used to create the grid that maps the `LEq` source pixels to the data's image-pixels.
             The options are `Grid2D` and `Grid2DIterate`.
             (see the `Grid2D` documentation for a description of these options).
         sub_size
-            If the grid and / or grid_pixelized use a `Grid2D`, this sets the sub-size used by the `Grid2D`.
+            If the grid and / or grid_pixelizaiton use a `Grid2D`, this sets the sub-size used by the `Grid2D`.
         fractional_accuracy
-            If the grid and / or grid_pixelized use a `Grid2DIterate`, this sets the fractional accuracy it
+            If the grid and / or grid_pixelizaiton use a `Grid2DIterate`, this sets the fractional accuracy it
             uses when evaluating functions, where the fraction accuracy is the ratio of the values computed using
             two grids at a higher and lower sub-grid size.
         relative_accuracy
-            If the grid and / or grid_pixelized use a `Grid2DIterate`, this sets the relative accuracy it
+            If the grid and / or grid_pixelizaiton use a `Grid2DIterate`, this sets the relative accuracy it
             uses when evaluating functions, where the relative accuracy is the absolute difference of the values
             computed using two grids at a higher and lower sub-grid size.
         sub_steps : [int]
-            If the grid and / or grid_pixelized use a `Grid2DIterate`, this sets the steps the sub-size is increased by
+            If the grid and / or grid_pixelizaiton use a `Grid2DIterate`, this sets the steps the sub-size is increased by
             to meet the fractional accuracy when evaluating functions.
         signal_to_noise_limit
             If input, the dataset's noise-map is rescaled such that no pixel has a signal-to-noise above the

[…]
