### 2022-07-27 Jammy2211/PyAutoGalaxy#35 @alexandergagliano
Confirming that, yes, the fix works and all tests pass.
