### 2022-07-29 cfe4f02
fix model_util test
