+++ test_autogalaxy/analysis/test_model_util.py
@@ -135,18 +135,12 @@ def test__set_upper_limit_of_pixelization_pixels_prior():
 def test__hyper_model_noise_from():
 
     pixelization = af.Model(
-        ag.Pixelization,
-        mesh=ag.mesh.Rectangular,
-                regularization=ag.reg.Constant,
+        ag.Pixelization, mesh=ag.mesh.Rectangular, regularization=ag.reg.Constant
     )
 
     model = af.Collection(
         galaxies=af.Collection(
-            galaxy=af.Model(
-                ag.Galaxy,
-                redshift=0.5,
-                pixelization=pixelization
-            ),
+            galaxy=af.Model(ag.Galaxy, redshift=0.5, pixelization=pixelization),
             galaxy_1=af.Model(ag.Galaxy, redshift=1.0, bulge=ag.lp.EllSersic),
         )
     )
