+++ autogalaxy/plane/plane.py
@@ -1,6 +1,6 @@
 import json
 import numpy as np
-from typing import Dict, List, Optional, Type
+from typing import Dict, List, Optional, Tuple, Type
 
 import autoarray as aa
 
@@ -71,7 +71,7 @@ class Plane(OperateImageGalaxies, OperateDeflections, Dictable):
     def galaxy_redshifts(self) -> List[float]:
         return [galaxy.redshift for galaxy in self.galaxies]
 
-    def has(self, cls: Type) -> bool:
+    def has(self, cls: Tuple[Type]) -> bool:
         if self.galaxies is not None:
             return any(list(map(lambda galaxy: galaxy.has(cls=cls), self.galaxies)))
         return False
@@ -332,7 +332,7 @@ class Plane(OperateImageGalaxies, OperateDeflections, Dictable):
         -------
             A bool which is True if an inversion is performed.
         """
-        if self.has(cls=aa.Pixelization) or self.has(cls=LightProfileLinear):
+        if self.has(cls=(aa.Pixelization, LightProfileLinear)):
             return True
         elif self.has(cls=Basis):
             basis_list = self.cls_list_from(cls=Basis)
