### 2022-08-12 8049970
fixed tests by introducing PixelizationGalaxy
