+++ autogalaxy/__init__.py
@@ -1,52 +1,48 @@
-from autoarray.dataset import preprocess
-from autoarray.dataset.imaging import SettingsImaging
-from autoarray.dataset.imaging import Imaging
-from autoarray.dataset.interferometer import Interferometer
-from autoarray.dataset.interferometer import SettingsInterferometer
-from autoarray.instruments import acs
-from autoarray.instruments import euclid
-from autoarray.inversion.pixelization import mesh
-from autoarray.inversion import regularization as reg
-from autoarray.inversion.pixelization.mappers.abstract import AbstractMapper
-from autoarray.inversion.inversion.settings import SettingsInversion
-from autoarray.inversion.inversion.factory import inversion_from as Inversion
-from autoarray.inversion.inversion.factory import (
-    inversion_imaging_unpacked_from as InversionImaging,
-)
-from autoarray.inversion.inversion.factory import (
-    inversion_interferometer_unpacked_from as InversionInterferometer,
-)
-from autoarray.inversion.pixelization.pixelization import Pixelization
-from autoarray.inversion.pixelization.mappers.mapper_grids import MapperGrids
-from autoarray.inversion.pixelization.mappers.factory import mapper_from as Mapper
-from autoarray.inversion.pixelization.settings import SettingsPixelization
-from autoarray.mask.mask_1d import Mask1D
-from autoarray.mask.mask_2d import Mask2D
-from autoarray.operators.convolver import Convolver
-from autoarray.operators.convolver import Convolver
-from autoarray.operators.transformer import TransformerDFT
-from autoarray.operators.transformer import TransformerNUFFT
-from autoarray.layout.layout import Layout2D
-from autoarray.structures.arrays.uniform_1d import Array1D
-from autoarray.structures.arrays.uniform_2d import Array2D
-from autoarray.structures.values import ValuesIrregular
-from autoarray.structures.header import Header
-from autoarray.structures.grids.uniform_1d import Grid1D
-from autoarray.structures.grids.uniform_2d import Grid2D
-from autoarray.structures.grids.sparse_2d import Grid2DSparse
-from autoarray.structures.grids.iterate_2d import Grid2DIterate
-from autoarray.structures.grids.irregular_2d import Grid2DIrregular
-from autoarray.structures.grids.irregular_2d import Grid2DIrregularUniform
-from autoarray.structures.mesh.rectangular_2d import Mesh2DRectangular
-from autoarray.structures.mesh.voronoi_2d import Mesh2DVoronoi
-from autoarray.structures.mesh.delaunay_2d import Mesh2DDelaunay
-from autoarray.structures.vectors.uniform import VectorYX2D
-from autoarray.structures.vectors.irregular import VectorYX2DIrregular
-from autoarray.layout.region import Region1D
-from autoarray.layout.region import Region2D
-from autoarray.structures.arrays.kernel_2d import Kernel2D
-from autoarray.structures.visibilities import Visibilities
-from autoarray.structures.visibilities import VisibilitiesNoiseMap
+from autoarray.dataset import preprocess  # noqa
+from autoarray.dataset.imaging import SettingsImaging  # noqa
+from autoarray.dataset.imaging import Imaging  # noqa
+from autoarray.dataset.interferometer import Interferometer  # noqa
+from autoarray.dataset.interferometer import SettingsInterferometer  # noqa
+from autoarray.instruments import acs  # noqa
+from autoarray.instruments import euclid  # noqa
+from autoarray.inversion.pixelization import mesh  # noqa
+from autoarray.inversion import regularization as reg  # noqa
[…]
