### 2022-08-12 a29aee3
remove redundant config fixtures
