+++ test_autogalaxy/analysis/test_visualizer.py
@@ -15,11 +15,6 @@ def make_visualizer_plotter_setup():
     return path.join("{}".format(directory), "files")
 
 
-@pytest.fixture(autouse=True)
-def push_config(plot_path):
-    conf.instance.push(path.join(directory, "config"), output_path=plot_path)
-
-
 def test__visualizes_plane__uses_configs(
     masked_imaging_7x7, plane_7x7, include_2d_all, plot_path, plot_patch
 ):
