### 2022-08-17 1157e2c
split API for pixels and parameters in linear objects
