+++ setup.py
@@ -15,7 +15,15 @@ version = environ.get("VERSION", "1.0.dev0")
 requirements.extend(
     [f"autoconf=={version}", f"autoarray=={version}", f"autofit=={version}"]
 )
-print(requirements)
+
+
+def config_packages(directory):
+    paths = [directory.replace("/", ".")]
+    for (path, directories, filenames) in os.walk(directory):
+        for directory in directories:
+            paths.append(f'{path}/{directory}'.replace("/", "."))
+    return paths
+
 
 setup(
     name="autogalaxy",
@@ -45,7 +53,7 @@ setup(
         "Programming Language :: Python :: 3.9",
     ],
     keywords="cli",
-    packages=find_packages(exclude=["docs", "test_autogalaxy", "test_autogalaxy*"]),
+    packages=find_packages(exclude=["docs", "test_autogalaxy", "test_autogalaxy*"]) + config_packages('autogalaxy/config'),
     install_requires=requirements,
     setup_requires=["pytest-runner"],
     tests_require=["pytest"],
