+++ optional_requirements.txt
@@ -1,0 +1,1 @@
+pylops>=1.10.0
