### 2022-08-22 f85144f
add numba to optional requirements
