### 2022-09-14 c13ee2c
sigma residual map in configs
