+++ autogalaxy/galaxy/galaxy.py
@@ -22,7 +22,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
     """
 
     def __init__(
-            self, redshift: float, hyper_galaxy: Optional["HyperGalaxy"] = None, **kwargs
+        self, redshift: float, hyper_galaxy: Optional["HyperGalaxy"] = None, **kwargs
     ):
         """
         Class representing a galaxy, which is composed of attributes used for fitting hyper_galaxies (e.g. light profiles, \
@@ -180,7 +180,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
     @aa.grid_dec.grid_2d_to_structure
     def image_2d_from(
-            self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
     ) -> Union[np.ndarray, aa.Array2D]:
         """
         Returns the summed 2D image of the galaxy's light profiles from a 2D grid of Cartesian (y,x) coordinates.
@@ -207,15 +207,15 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
             and returned.
         """
         if (
-                len(self.cls_list_from(cls=LightProfile, cls_filtered=LightProfileLinear))
-                > 0
+            len(self.cls_list_from(cls=LightProfile, cls_filtered=LightProfileLinear))
+            > 0
         ):
             return sum(self.image_2d_list_from(grid=grid, operated_only=operated_only))
 
         return np.zeros((grid.shape[0],))
 
     def image_2d_list_from(
-            self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
+        self, grid: aa.type.Grid2DLike, operated_only: Optional[bool] = None
     ) -> List[aa.Array2D]:
         """
         Returns a list of the 2D images of the galaxy's light profiles from a 2D grid of Cartesian (y,x) coordinates.
@@ -271,7 +271,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
             image_1d_list = []
 
             for light_profile in self.cls_list_from(
-                    cls=LightProfile, cls_filtered=LightProfileLinear
+                cls=LightProfile, cls_filtered=LightProfileLinear
             ):
 
                 grid_radial = self.grid_radial_from(
@@ -569,10 +569,10 @@ class HyperGalaxy:
     _ids = count()
 
     def __init__(
-            self,
-            contribution_factor: float = 0.0,
-            noise_factor: float = 0.0,
-            noise_power: float = 1.0,
+        self,
+        contribution_factor: float = 0.0,
+        noise_factor: float = 0.0,
+        noise_power: float = 1.0,
     ):
         """
         If a `Galaxy` is given a *HyperGalaxy* as an attribute, the noise-map in \
@@ -609,9 +609,9 @@ class HyperGalaxy:
     def __eq__(self, other):
         if isinstance(other, HyperGalaxy):
             return (
-                    self.contribution_factor == other.contribution_factor
-                    and self.noise_factor == other.noise_factor
-                    and self.noise_power == other.noise_power
+                self.contribution_factor == other.contribution_factor
+                and self.noise_factor == other.noise_factor
+                and self.noise_power == other.noise_power
             )
         return False
 
@@ -647,10 +647,10 @@ class HyperGalaxy:
             raise
 
     def hyper_noise_map_via_hyper_images_from(
-            self,
[…]
