+++ autogalaxy/config/visualize/plots.ini
@@ -31,10 +31,10 @@ subtracted_images_of_galaxies=True
 model_images_of_galaxies=True
 data=False
 noise_map=False
-signal_to_noise_map=False
 model_data=False
 residual_map=False
 normalized_residual_map=False
+sigma_residual_map=False
 chi_squared_map=False
 
 [fit_imaging]

[…]
