+++ autogalaxy/config/visualize/plots.ini
@@ -34,7 +34,6 @@ noise_map=False
 model_data=False
 residual_map=False
 normalized_residual_map=False
-sigma_residual_map=False
 chi_squared_map=False
 
 [fit_imaging]
