+++ test_autogalaxy/plot/mat_wrap/config/visualize/plots.ini
@@ -19,7 +19,6 @@ noise_map=False
 model_data=False
 residual_map=False
 normalized_residual_map=False
-sigma_residual_map=False
 chi_squared_map=False
 subtracted_images_of_galaxies=False
 model_images_of_galaxies=False
