+++ test_autogalaxy/plot/mat_wrap/config/visualize/plots.ini
@@ -16,6 +16,7 @@ subplot_fit=True
 subplots_of_planes_fits=True
 data=False
 noise_map=False
+signal_to_noise_map=False
 model_data=False
 residual_map=False
 normalized_residual_map=False
