+++ autogalaxy/config/visualize/plots.ini
@@ -31,6 +31,7 @@ subtracted_images_of_galaxies=True
 model_images_of_galaxies=True
 data=False
 noise_map=False
+signal_to_noise_map=False
 model_data=False
 residual_map=False
 normalized_residual_map=False
