### 2022-09-27 #4475 @Jammy2211
I have respond to all comments and fixed them on the project `master` branches.

A `pip` release will follow in the next week or so, we are doing a big switch to automated build servers so trying not to rock the boat with a release until everything is fully tested.
