+++ docs/requirements.txt
@@ -2,7 +2,7 @@ autoconf==2022.07.11.1
 autofit==2022.07.11.1
 autoarray==2022.07.11.1
 astropy>=3.0.0
-colossus==1.2.14
+colossus==1.3.1
 numpydoc>=1.0.0
 matplotlib>=3.0.3
 numba
