+++ test_autogalaxy/config/non_linear/nest/DynestyStatic.ini
@@ -7,8 +7,7 @@ sample=auto
 bootstrap=None
 enlarge=None
 update_interval=None
-vol_dec=0.5
-vol_check=2.0
+first_update=None
 walks=5
 slices=5
 fmove=0.9
