### 2022-09-28 049f90d
simplifiy docs requirements
