+++ docs/requirements.txt
@@ -3,12 +3,7 @@ autofit==2022.07.11.1
 autoarray==2022.07.11.1
 astropy>=3.0.0
 colossus==1.3.1
-numpydoc>=1.0.0
-matplotlib>=3.0.3
 numba
-numpy>=1.18.0
-pyprojroot==0.2.0
-scikit-learn==0.21.3
 sphinx==4.0.0
 furo
 myst-parser
