+++ optional_requirements.txt
@@ -1,3 +1,7 @@
-pynufft==2020.2.7
 numba
+llvmlite
 pylops>=1.10.0
+pynufft==2020.2.7
+ultranest==3.2.0
+jax==0.3.1
+jaxlib==0.3.0
