+++ docs/installation/overview.rst
@@ -3,7 +3,7 @@
 Overview
 ========
 
-**PyAutoGalaxy** requires Python 3.6+ and support the Linux, MacOS and Windows operating systems.
+**PyAutoGalaxy** requires Python 3.8+ and support the Linux, MacOS and Windows operating systems.
 
 **PyAutoGalaxy** can be installed via the Python distribution `Anaconda <https://www.anaconda.com/>`_ or using
 `Pypi <https://pypi.org/>`_ to ``pip install`` **PyAutoGalaxy** into your Python distribution.
