+++ docs/conf.py
@@ -15,13 +15,37 @@ import datetime
 
 from pyprojroot import here
 
-workspace_path = str(here())
+# /home/docs/checkouts/readthedocs.org/user_builds/pyautolens/checkouts/latest/docs
+
+clone_path = str(here())
 
 import os
 import sys
-import autogalaxy
 
 sys.path.insert(0, os.path.abspath("."))
+
+def clone_repo(name:str, url:str):
+
+    clone = f"git clone {url}/{name}"
+    os.system(clone)
+    os.system(f"pip install -r {name}/requirements.txt")
+    os.system(f"rm -rf {name}/docs")
+    sys.path.insert(
+        0,
+        os.path.abspath(f"{clone_path}/{name}"),
+    )
+
+clone_repo(name="PyAutoFit", url="https://github.com/rhayes777")
+clone_repo(name="PyAutoArray", url="https://github.com/Jammy2211")
+
+clone_path = os.path.split(clone_path)[0]
+
+sys.path.insert(
+    0,
+    os.path.abspath(clone_path),
+)
+
+import autogalaxy
 
 # -- Project information -----------------------------------------------------
 
@@ -31,7 +55,7 @@ copyright = "2022, James Nightingale, Richard Hayes"
 author = "James Nightingale, Richard Hayes"
 
 # The full version, including alpha/beta/rc tags
-release = "2022.05.02.1"
+release = autogalaxy.__version__
 master_doc = "index"
 
 
