+++ docs/requirements.txt
@@ -1,10 +1,8 @@
 autoconf==2022.07.11.1
-autofit==2022.07.11.1
-autoarray==2022.07.11.1
 astropy>=3.0.0
 colossus==1.3.1
 numba
-sphinx==4.0.0
+sphinx==5.2.3
 furo
 myst-parser
 sphinx_copybutton
