+++ test_autogalaxy/config/general.ini
@@ -37,7 +37,6 @@ hyper_noise_limit=1.0e8
 stochastic_outputs=False
 
 [test]
-test_mode=False
 check_figure_of_merit_sanity=False
 check_preloads=False
 
