+++ docs/requirements.txt
@@ -2,6 +2,7 @@ autoconf==2022.07.11.1
 astropy>=3.0.0
 colossus==1.3.1
 numba
+pyprojroot==0.2.0
 sphinx==5.2.3
 furo
 myst-parser
