+++ optional_requirements.txt
@@ -1,7 +1,8 @@
 numba
 llvmlite
-pylops>=1.10.0
+pylops>=1.10.0,<=1.18.3
 pynufft==2020.2.7
 ultranest==3.2.0
+zeus-mcmc==2.4.1
 jax==0.3.1
 jaxlib==0.3.0
