+++ autogalaxy/profiles/mass_profiles/dark_mass_profiles.py
@@ -337,8 +337,10 @@ class AbstractEllNFWGeneralized(MassProfile, DarkProfile, MassProfileMGE):
                 "string. Must be {local, profile}"
             )
 
-        cosmic_average_density = cosmology.cosmic_average_density_solar_mass_per_kpc3_from(
-            redshift=redshift_calc
+        cosmic_average_density = (
+            cosmology.cosmic_average_density_solar_mass_per_kpc3_from(
+                redshift=redshift_calc
+            )
         )
 
         rho_scale_radius = self.rho_at_scale_radius_solar_mass_per_kpc3(
@@ -417,8 +419,10 @@ class AbstractEllNFWGeneralized(MassProfile, DarkProfile, MassProfileMGE):
                 "string. Must be {local, profile}"
             )
 
-        cosmic_average_density = cosmology.cosmic_average_density_solar_mass_per_kpc3_from(
-            redshift=redshift_calc
+        cosmic_average_density = (
+            cosmology.cosmic_average_density_solar_mass_per_kpc3_from(
+                redshift=redshift_calc
+            )
         )
 
         radius_at_200 = self.radius_at_200(
@@ -436,7 +440,7 @@ class AbstractEllNFWGeneralized(MassProfile, DarkProfile, MassProfileMGE):
             200.0
             * ((4.0 / 3.0) * np.pi)
             * cosmic_average_density
-            * (radius_at_200_kpc ** 3.0)
+            * (radius_at_200_kpc**3.0)
         )
 
     @property
@@ -566,19 +570,19 @@ class EllNFWGeneralized(AbstractEllNFWGeneralized):
         surface_density_integral,
     ):
 
-        eta_u = np.sqrt((u * ((x ** 2) + (y ** 2 / (1 - (1 - axis_ratio ** 2) * u)))))
+        eta_u = np.sqrt((u * ((x**2) + (y**2 / (1 - (1 - axis_ratio**2) * u)))))
         bin_size = (maximum_log_eta - minimum_log_eta) / (tabulate_bins - 1)
         i = 1 + int((np.log10(eta_u) - minimum_log_eta) / bin_size)
         r1 = 10.0 ** (minimum_log_eta + (i - 1) * bin_size)
-        r2 = r1 * 10.0 ** bin_size
+        r2 = r1 * 10.0**bin_size
         kap = surface_density_integral[i] + (
             surface_density_integral[i + 1] - surface_density_integral[i]
         ) * (eta_u - r1) / (r2 - r1)
-        return kap / (1.0 - (1.0 - axis_ratio ** 2) * u) ** (npow + 0.5)
+        return kap / (1.0 - (1.0 - axis_ratio**2) * u) ** (npow + 0.5)
 
     def convergence_func(self, grid_radius: float) -> float:
         def integral_y(y, eta):
-            return (y + eta) ** (self.inner_slope - 4) * (1 - np.sqrt(1 - y ** 2))
+            return (y + eta) ** (self.inner_slope - 4) * (1 - np.sqrt(1 - y**2))
 
         grid_radius = (1.0 / self.scale_radius) * grid_radius
 
@@ -623,7 +627,7 @@ class EllNFWGeneralized(AbstractEllNFWGeneralized):
         @jit_integrand
         def deflection_integrand(x, kappa_radius, scale_radius, inner_slope):
             return (x + kappa_radius / scale_radius) ** (inner_slope - 3) * (
[…]
