+++ autogalaxy/operate/image.py
@@ -361,10 +361,10 @@ class OperateImageList(OperateImage):
 
 class OperateImageGalaxies(OperateImageList):
     """
-    Packages methods which using a list of galaxies returns a dictionary of their 2D images using the function 
+    Packages methods which using a list of galaxies returns a dictionary of their 2D images using the function
     `galaxy_image_2d_dict_from` (e.g. a `Plane`, a `Tracer` in the library **PyAutoLens**).
 
-    The majority of methods apply data operators to the dictionary of 2D images which perform tasks such as a 2D 
+    The majority of methods apply data operators to the dictionary of 2D images which perform tasks such as a 2D
     convolution of Fourier transform. This retains the keys of the dictionary to maintain information on the galaxies.
 
     The methods in `OperateImageGalaxies` are inherited by light objects to provide a concise API.
