+++ test_autogalaxy/config/general.ini
@@ -31,6 +31,9 @@ remove_projected_centre=False
 [structures]
 use_dataset_grids=True
 
+[pixelization]
+voronoi_nn_max_interpolation_neighbors=10
+
 [hyper]
 hyper_minimum_percent=0.01
 hyper_noise_limit=1.0e8
