### 2022-10-24 65868e2
refactored imaging module plus tests into package
