+++ test_autogalaxy/config/general.ini
@@ -32,7 +32,7 @@ remove_projected_centre=False
 use_dataset_grids=True
 
 [pixelization]
-voronoi_nn_max_interpolation_neighbors=10
+voronoi_nn_max_interpolation_neighbors=100
 
 [hyper]
 hyper_minimum_percent=0.01
