### 2022-10-24 f6842f0
split fit imaging into two classes, one abstract, for correlated fit
