+++ autogalaxy/imaging/fit/fit_imaging.py
@@ -1,24 +1,22 @@
 import numpy as np
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
 from autoconf import conf
 from autoconf import cached_property
 
 import autoarray as aa
 
-from autogalaxy.abstract_fit import AbstractFit
+from autogalaxy.abstract_fit import AbstractFitInversion
 from autogalaxy.analysis.preloads import Preloads
+from autogalaxy.imaging.fit.abstract import AbstractFitImaging
 from autogalaxy.galaxy.galaxy import Galaxy
 from autogalaxy.hyper.hyper_data import HyperImageSky
 from autogalaxy.hyper.hyper_data import HyperBackgroundNoise
 from autogalaxy.plane.plane import Plane
 from autogalaxy.plane.to_inversion import PlaneToInversion
-from autogalaxy.profiles.light_profiles.light_profiles_linear import LightProfileLinear
-
-from autogalaxy import exc
 
 
-class FitImaging(aa.FitImaging, AbstractFit):
+class FitImaging(AbstractFitImaging, AbstractFitInversion):
     def __init__(
         self,
         dataset: aa.Imaging,
@@ -81,12 +79,15 @@ class FitImaging(aa.FitImaging, AbstractFit):
             decorator take to run.
         """
 
-        super().__init__(dataset=dataset, profiling_dict=profiling_dict)
-        AbstractFit.__init__(
+        super().__init__(
+            dataset=dataset,
+            plane=plane,
+            preloads=preloads,
+            profiling_dict=profiling_dict,
+        )
+        AbstractFitInversion.__init__(
             self=self, model_obj=plane, settings_inversion=settings_inversion
         )
-
-        self.plane = plane
 
         self.hyper_image_sky = hyper_image_sky
         self.hyper_background_noise = hyper_background_noise
@@ -95,8 +96,6 @@ class FitImaging(aa.FitImaging, AbstractFit):
 
         self.settings_pixelization = settings_pixelization
         self.settings_inversion = settings_inversion
-
-        self.preloads = preloads
 
     @property
     def data(self) -> aa.Array2D:
@@ -127,25 +126,6 @@ class FitImaging(aa.FitImaging, AbstractFit):
             )
 
         return self.dataset.noise_map
-
-    @property
-    def blurred_image(self) -> aa.Array2D:
-        """
-        Returns the image of the light profiles of all galaxies in the fit's plane, convolved with the
-        imaging dataset's PSF.
-        """
-        return self.plane.blurred_image_2d_from(
-            grid=self.dataset.grid,
-            convolver=self.dataset.convolver,
-            blurring_grid=self.dataset.blurring_grid,
-        )
-
-    @property
-    def profile_subtracted_image(self) -> aa.Array2D:
-        """
-        Returns the dataset's image data with all blurred light profile images in the fit's plane subtracted.
-        """
-        return self.image - self.blurred_image
 
     @property
     def plane_to_inversion(self) -> PlaneToInversion:
@@ -192,14 +172,6 @@ class FitImaging(aa.FitImaging, AbstractFit):
         return self.blurred_image
 
     @property
-    def galaxies(self) -> List[Galaxy]:
-        return self.plane.galaxies
-
-    @property
-    def grid(self) -> aa.type.Grid2DLike:
-        return self.imaging.grid
-
-    @property
     def galaxy_model_image_dict(self) -> Dict[Galaxy, np.ndarray]:
         """
         A dictionary which associates every galaxy in the plane with its `model_image`.
@@ -225,73 +197,6 @@ class FitImaging(aa.FitImaging, AbstractFit):
         )
 
[…]
