+++ autogalaxy/quantity/dataset_quantity.py
@@ -47,9 +47,6 @@ class SettingsQuantity(AbstractSettingsDataset):
         sub_steps : [int]
             If the grid and / or grid_pixelization use a `Grid2DIterate`, this sets the steps the sub-size is increased by
             to meet the fractional accuracy when evaluating functions.
-        signal_to_noise_limit
-            If input, the dataset's noise-map is rescaled such that no pixel has a signal-to-noise above the
-            signa to noise limit.
         psf_shape_2d
             The shape of the PSF used for convolving model image generated using analytic light profiles. A smaller
             shape will trim the PSF relative to the input image PSF, giving a faster analysis run-time.
