### 2022-10-26 3353577
raise exception of inversion performe with noise covariane
