### 2022-10-30 f623f5c
symmetric residual maps in imaging / interferometer
