+++ autogalaxy/imaging/plot/fit_imaging_plotters.py
@@ -21,6 +21,7 @@ class FitImagingPlotter(Plotter):
         mat_plot_2d: MatPlot2D = MatPlot2D(),
         visuals_2d: Visuals2D = Visuals2D(),
         include_2d: Include2D = Include2D(),
+        residuals_symmetric_cmap: bool = True
     ):
         """
         Plots the attributes of `FitImaging` objects using the matplotlib method `imshow()` and many other matplotlib
@@ -44,6 +45,9 @@ class FitImagingPlotter(Plotter):
             Contains visuals that can be overlaid on the plot.
         include_2d
             Specifies which attributes of the `FitImaging` are extracted and plotted as visuals.
+        residuals_symmetric_cmap
+            If true, the `residual_map` and `normalized_residual_map` are plotted with a symmetric color map such
+            that `abs(vmin) = abs(vmax)`.
         """
         super().__init__(
             mat_plot_2d=mat_plot_2d, include_2d=include_2d, visuals_2d=visuals_2d
@@ -57,6 +61,7 @@ class FitImagingPlotter(Plotter):
             mat_plot_2d=self.mat_plot_2d,
             include_2d=self.include_2d,
             visuals_2d=self.visuals_2d,
+            residuals_symmetric_cmap=residuals_symmetric_cmap
         )
 
         self.figures_2d = self._fit_imaging_meta_plotter.figures_2d
