+++ requirements.txt
@@ -1,2 +1,2 @@
 colossus==1.3.1
-astropy>=3.0.0
+astropy>=5.0
