+++ docs/requirements.txt
@@ -1,5 +1,5 @@
 autoconf==2022.07.11.1
-astropy>=3.0.0
+astropy>=5.0
 colossus==1.3.1
 numba
 pyprojroot==0.2.0
