### 2022-10-31 31c1903
check preload config option now works
