+++ autogalaxy/interferometer/plot/fit_interferometer_plotters.py
@@ -25,7 +25,7 @@ class FitInterferometerPlotter(Plotter):
         mat_plot_2d: MatPlot2D = MatPlot2D(),
         visuals_2d: Visuals2D = Visuals2D(),
         include_2d: Include2D = Include2D(),
-        residuals_symmetric_cmap: bool = True
+        residuals_symmetric_cmap: bool = True,
     ):
         """
         Plots the attributes of `FitInterferometer` objects using the matplotlib method `imshow()` and many
@@ -80,7 +80,7 @@ class FitInterferometerPlotter(Plotter):
             mat_plot_2d=self.mat_plot_2d,
             include_2d=self.include_2d,
             visuals_2d=self.visuals_2d,
-            residuals_symmetric_cmap=residuals_symmetric_cmap
+            residuals_symmetric_cmap=residuals_symmetric_cmap,
         )
 
         self.figures_2d = self._fit_interferometer_meta_plotter.figures_2d
