+++ autogalaxy/analysis/analysis.py
@@ -214,10 +214,13 @@ class AnalysisDataset(Analysis):
             self.preloads = self.preloads_cls.setup_all_via_fits(
                 fit_0=fit_0, fit_1=fit_1
             )
-            try:
-                self.preloads.check_via_fit(fit=fit_0)
-            except (aa.exc.InversionException, exc.InversionException):
-                pass
+
+            if conf.instance["general"]["test"]["check_preloads"]:
+
+                try:
+                    self.preloads.check_via_fit(fit=fit_0)
+                except (aa.exc.InversionException, exc.InversionException):
+                    pass
 
         self.preloads.output_info_to_summary(file_path=paths.profile_path)
 
