+++ autogalaxy/plot/__init__.py
@@ -54,7 +54,7 @@ from autoarray.dataset.plot.interferometer_plotters import InterferometerPlotter
 from autoarray.plot.multi_plotters import MultiFigurePlotter
 from autoarray.plot.multi_plotters import MultiYX1DPlotter
 
-from autoarray.plot.mat_wrap.mat_plot import AutoLabels
+from autoarray.plot.auto_labels import AutoLabels
 
 from autogalaxy.plot.wrap import (
     HalfLightRadiusAXVLine,
@@ -67,9 +67,13 @@ from autogalaxy.plot.wrap import (
     MultipleImagesScatter,
 )
 
-from autogalaxy.plot.mat_wrap.mat_plot import MatPlot1D, MatPlot2D
-from autogalaxy.plot.mat_wrap.include import Include1D, Include2D
-from autogalaxy.plot.mat_wrap.visuals import Visuals1D, Visuals2D
+
+from autogalaxy.plot.mat_plot.one_d import MatPlot1D
+from autogalaxy.plot.mat_plot.two_d import MatPlot2D
+from autogalaxy.plot.include.one_d import Include1D
+from autogalaxy.plot.include.two_d import Include2D
+from autogalaxy.plot.visuals.one_d import Visuals1D
+from autogalaxy.plot.visuals.two_d import Visuals2D
 
 from autogalaxy.profiles.plot.light_profile_plotters import LightProfilePlotter
 from autogalaxy.profiles.plot.light_profile_plotters import LightProfilePDFPlotter
