+++ autogalaxy/plot/mat_plot/two_d.py
@@ -4,135 +4,6 @@ import autoarray.plot as aplt
 from autogalaxy.plot import wrap as w
 
 
-class MatPlot1D(aplt.MatPlot1D):
-    def __init__(
-        self,
-        units: Optional[aplt.Units] = None,
-        figure: Optional[aplt.Figure] = None,
-        axis: Optional[aplt.Axis] = None,
-        cmap: Optional[aplt.Cmap] = None,
-        colorbar: Optional[aplt.Colorbar] = None,
-        colorbar_tickparams: Optional[aplt.ColorbarTickParams] = None,
-        tickparams: Optional[aplt.TickParams] = None,
-        yticks: Optional[aplt.YTicks] = None,
-        xticks: Optional[aplt.XTicks] = None,
-        title: Optional[aplt.Title] = None,
-        ylabel: Optional[aplt.YLabel] = None,
-        xlabel: Optional[aplt.XLabel] = None,
-        text: Optional[aplt.Text] = None,
-        legend: Optional[aplt.Legend] = None,
-        output: Optional[aplt.Output] = None,
-        yx_plot: Optional[aplt.YXPlot] = None,
-        fill_between: Optional[aplt.FillBetween] = None,
-        half_light_radius_axvline: Optional[w.HalfLightRadiusAXVLine] = None,
-        einstein_radius_axvline: Optional[w.EinsteinRadiusAXVLine] = None,
-        model_fluxes_yx_scatter: Optional[w.ModelFluxesYXScatter] = None,
-    ):
-        """
-        Visualizes 1D data structures as a y versus x plot using Matplotlib.
-
-        The `Plotter` is passed objects from the `wrap_base` package which wrap matplotlib plot functions and customize
-        the appearance of the plots of the data structure. If the values of these matplotlib wrapper objects are not
-        manually specified, they assume the default values provided in the `config.visualize.mat_*` `.ini` config files.
-
-        Parameters
-        ----------
-        units
-            The units of the figure used to plot the data structure which sets the y and x ticks and labels.
-        figure
-            Opens the matplotlib figure before plotting via `plt.figure` and closes it once plotting is complete
-            via `plt.close`.
-        axis
-            Sets the extent of the figure axis via `plt.axis` and allows for a manual axis range.
-        cmap
-            Customizes the colormap of the plot and its normalization via matplotlib `colors` objects such
-            as `colors.Normalize` and `colors.LogNorm`.
-        colorbar
-            Plots the colorbar of the plot via `plt.colorbar` and customizes its tick labels and values using method
-            like `cb.set_yticklabels`.
-        colorbar_tickparams
-            Customizes the yticks of the colorbar plotted via `plt.colorbar`.
-        tickparams
-            Customizes the appearances of the y and x ticks on the plot, (e.g. the fontsize), using `plt.tick_params`.
-        yticks
-            Sets the yticks of the plot, including scaling them to new units depending on the `Units` object, via
-            `plt.yticks`.
-        xticks
-            Sets the xticks of the plot, including scaling them to new units depending on the `Units` object, via
-            `plt.xticks`.
-        title
-            Sets the figure title and customizes its appearance using `plt.title`.
-        ylabel
-            Sets the figure ylabel and customizes its appearance using `plt.ylabel`.
-        xlabel
-            Sets the figure xlabel and customizes its appearance using `plt.xlabel`.
-        legend
-            Sets whether the plot inclues a legend and customizes its appearance and labels using `plt.legend`.
-        output
-            Sets if the figure is displayed on the user's screen or output to `.png` using `plt.show` and `plt.savefig`
-        yx_plot
[…]
