+++ autogalaxy/profiles/light_profiles/base/dev_vaucouleurs.py
@@ -1,0 +1,81 @@
+import numpy as np
+from scipy.integrate import quad
+from typing import Optional, Tuple, Union
+
+import autoarray as aa
+
+from autogalaxy.profiles.light_profiles.light_profile_decorators import (
+    check_operated_only,
+)
+from autogalaxy.operate.image import OperateImage
+from autogalaxy.profiles.geometry_profiles import EllProfile
+
+
+
+class EllDevVaucouleurs(EllSersic):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        effective_radius: float = 0.6,
+    ):
+        """
+        The elliptical Dev Vaucouleurs light profile.
+
+        This is a specific case of the elliptical Sersic profile where `sersic_index=4.0`.
+
+        See `autogalaxy.profiles.light_profiles.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        intensity
+            Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
+            the light profile's image is compared too, which is expected to be electrons per second).
+        effective_radius
+            The circular radius containing half the light of this profile.
+        """
+        super().__init__(
+            centre=centre,
+            elliptical_comps=elliptical_comps,
+            intensity=intensity,
+            effective_radius=effective_radius,
+            sersic_index=4.0,
+        )
+
+
+class SphDevVaucouleurs(EllDevVaucouleurs):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        effective_radius: float = 0.6,
+    ):
+        """
+        The spherical Dev Vaucouleurs light profile.
+
+        This is a specific case of the elliptical Sersic profile where `sersic_index=4.0`.
+
+        See `autogalaxy.profiles.light_profiles.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        intensity
+            Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
+            the light profile's image is compared too, which is expected to be electrons per second).
+        effective_radius
+            The circular radius containing half the light of this profile.
+        """
+        super().__init__(
+            centre=centre,
+            elliptical_comps=(0.0, 0.0),
+            intensity=intensity,
+            effective_radius=effective_radius,
+        )
