### 2022-11-12 6204b2f
repackage light profile tests
