### 2022-11-12 0504a7d
repackage total and stellar mass profiles
