+++ autogalaxy/profiles/mass_profiles/total/power_law_cored.py
@@ -1,0 +1,257 @@
+import copy
+import numpy as np
+from scipy.integrate import quad
+from scipy import special
+from typing import Tuple
+
+import autoarray as aa
+
+from autogalaxy.profiles.abstract.abstract import MassProfile
+
+
+class EllPowerLawCored(MassProfile):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        einstein_radius: float = 1.0,
+        slope: float = 2.0,
+        core_radius: float = 0.01,
+    ):
+        """
+        Represents a cored elliptical power-law density distribution
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        einstein_radius
+            The arc-second Einstein radius.
+        slope
+            The density slope of the power-law (lower value -> shallower profile, higher value -> steeper profile).
+        core_radius
+            The arc-second radius of the inner core.
+        """
+        super().__init__(centre=centre, elliptical_comps=elliptical_comps)
+
+        self.einstein_radius = einstein_radius
+        self.slope = slope
+        self.core_radius = core_radius
+
+    @property
+    def einstein_radius_rescaled(self):
+        """Rescale the einstein radius by slope and axis_ratio, to reduce its degeneracy with other mass-profiles
+        parameters"""
+        return ((3 - self.slope) / (1 + self.axis_ratio)) * self.einstein_radius ** (
+            self.slope - 1
+        )
+
+    @aa.grid_dec.grid_2d_to_structure
+    @aa.grid_dec.transform
+    @aa.grid_dec.relocate_to_radial_minimum
+    def convergence_2d_from(self, grid: aa.type.Grid2DLike):
+        """
+        Calculate the projected convergence on a grid of (y,x) arc-second coordinates.
+
+        The `grid_2d_to_structure` decorator reshapes the ndarrays the convergence is outputted on. See \
+        *aa.grid_2d_to_structure* for a description of the output.
+
+        Parameters
+        ----------
+        grid
+            The grid of (y,x) arc-second coordinates the convergence is computed on.
+
+        """
+
+        covnergence_grid = np.zeros(grid.shape[0])
+
+        grid_eta = self.grid_to_elliptical_radii(grid)
+
+        for i in range(grid.shape[0]):
+            covnergence_grid[i] = self.convergence_func(grid_eta[i])
+
+        return covnergence_grid
+
+    @aa.grid_dec.grid_2d_to_structure
+    @aa.grid_dec.transform
+    @aa.grid_dec.relocate_to_radial_minimum
+    def potential_2d_from(self, grid: aa.type.Grid2DLike):
+        """
+        Calculate the potential on a grid of (y,x) arc-second coordinates.
+
+        Parameters
+        ----------
+        grid
+            The grid of (y,x) arc-second coordinates the deflection angles are computed on.
+
+        """
+
+        potential_grid = np.zeros(grid.shape[0])
+
+        for i in range(grid.shape[0]):
+
+            potential_grid[i] = quad(
+                self.potential_func,
+                a=0.0,
+                b=1.0,
+                args=(
+                    grid[i, 0],
+                    grid[i, 1],
+                    self.axis_ratio,
+                    self.slope,
+                    self.core_radius,
+                ),
+            )[0]
+
+        return self.einstein_radius_rescaled * self.axis_ratio * potential_grid
+
[…]
