+++ test_autogalaxy/profiles/mass_profiles/dark/test_gnfw.py
@@ -1,0 +1,307 @@
+import numpy as np
+import pytest
+
+import autogalaxy as ag
+
+grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
+
+
+def test__deflections_2d_via_integral_from():
+
+    gnfw = ag.mp.SphNFWGeneralized(
+        centre=(0.0, 0.0), kappa_s=1.0, inner_slope=0.5, scale_radius=8.0
+    )
+
+    deflections = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(0.43501, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.37701, 1e-3)
+
+    gnfw = ag.mp.SphNFWGeneralized(
+        centre=(0.3, 0.2), kappa_s=2.5, inner_slope=1.5, scale_radius=4.0
+    )
+
+    deflections = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(-9.31254, 1e-3)
+    assert deflections[0, 1] == pytest.approx(-3.10418, 1e-3)
+
+    gnfw = ag.mp.EllNFWGeneralized(
+        centre=(0.0, 0.0),
+        kappa_s=1.0,
+        elliptical_comps=ag.convert.elliptical_comps_from(
+            axis_ratio=0.3, angle=100.0
+        ),
+        inner_slope=0.5,
+        scale_radius=8.0,
+    )
+    deflections = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+    assert deflections[0, 0] == pytest.approx(0.26604, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.58988, 1e-3)
+
+    gnfw = ag.mp.EllNFWGeneralized(
+        centre=(0.3, 0.2),
+        kappa_s=2.5,
+        elliptical_comps=ag.convert.elliptical_comps_from(
+            axis_ratio=0.5, angle=100.0
+        ),
+        inner_slope=1.5,
+        scale_radius=4.0,
+    )
+    deflections = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+    assert deflections[0, 0] == pytest.approx(-5.99032, 1e-3)
+    assert deflections[0, 1] == pytest.approx(-4.02541, 1e-3)
+
+def test__deflections_2d_via_mge_from():
+
+    gnfw = ag.mp.SphNFWGeneralized(
+        centre=(0.0, 0.0), kappa_s=1.0, inner_slope=0.5, scale_radius=8.0
+    )
+
+    deflections_via_integral = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+    deflections_via_mge = gnfw.deflections_2d_via_mge_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
+
+    assert deflections_via_integral == pytest.approx(deflections_via_mge, 1.0e-3)
+
+    gnfw = ag.mp.SphNFWGeneralized(
+        centre=(0.3, 0.2), kappa_s=2.5, inner_slope=1.5, scale_radius=4.0
+    )
+
+    deflections_via_integral = gnfw.deflections_2d_via_integral_from(
+        grid=np.array([[0.1875, 0.1625]])
+    )
[…]
