### 2022-11-12 d6b840f
stellar mass tests repackaged
