+++ test_autogalaxy/profiles/mass_profiles/stellar/test_sersic.py
@@ -1,0 +1,356 @@
+import numpy as np
+import pytest
+
+import autogalaxy as ag
+
+grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
+
+
+def test__deflections_via_integral_from():
+
+    sersic = ag.mp.EllSersic(
+        centre=(-0.4, -0.2),
+        elliptical_comps=(-0.07142, -0.085116),
+        intensity=5.0,
+        effective_radius=0.2,
+        sersic_index=2.0,
+        mass_to_light_ratio=1.0,
+    )
+
+    deflections = sersic.deflections_2d_via_integral_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(1.1446, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.79374, 1e-3)
+
+    sersic = ag.mp.EllSersic(
+        centre=(-0.4, -0.2),
+        elliptical_comps=(-0.07142, -0.085116),
+        intensity=10.0,
+        effective_radius=0.2,
+        sersic_index=3.0,
+        mass_to_light_ratio=1.0,
+    )
+
+    deflections = sersic.deflections_2d_via_integral_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(2.6134, 1e-3)
+    assert deflections[0, 1] == pytest.approx(1.80719, 1e-3)
+
+
+def test__deflections_2d_via_mge_from():
+
+    sersic = ag.mp.EllSersic(
+        centre=(-0.4, -0.2),
+        elliptical_comps=(-0.07142, -0.085116),
+        intensity=5.0,
+        effective_radius=0.2,
+        sersic_index=2.0,
+        mass_to_light_ratio=1.0,
+    )
+
+    deflections_via_integral = sersic.deflections_2d_via_integral_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+    deflections_via_mge = sersic.deflections_2d_via_mge_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
+    assert deflections_via_integral == pytest.approx(deflections_via_mge, 1.0e-3)
+
+    sersic = ag.mp.EllSersic(
+        centre=(-0.4, -0.2),
+        elliptical_comps=(-0.07142, -0.085116),
+        intensity=10.0,
+        effective_radius=0.2,
+        sersic_index=3.0,
+        mass_to_light_ratio=1.0,
+    )
+
+    deflections_via_integral = sersic.deflections_2d_via_integral_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+    deflections_via_mge = sersic.deflections_2d_via_mge_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
+    assert deflections_via_integral == pytest.approx(deflections_via_mge, 1.0e-3)
+
+
+def test__deflections_2d_via_cse_from():
+
+    sersic = ag.mp.EllSersic(
+        centre=(-0.4, -0.2),
+        elliptical_comps=(-0.07142, -0.085116),
+        intensity=5.0,
+        effective_radius=0.2,
+        sersic_index=2.0,
+        mass_to_light_ratio=1.0,
+    )
+
+    deflections_via_integral = sersic.deflections_2d_via_integral_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+    deflections_via_cse = sersic.deflections_2d_via_cse_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
[…]
