+++ test_autogalaxy/profiles/mass_profiles/dark/test_nfw_scatter.py
@@ -5,6 +5,7 @@ import autogalaxy as ag
 
 grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
+
 def test__scatter_is_nonzero():
 
     nfw = ag.mp.SphNFWMCRScatterLudlow(
