+++ test_autogalaxy/profiles/mass_profiles/sheets/test_input_deflections.py
@@ -1,0 +1,250 @@
+import autogalaxy as ag
+import numpy as np
+import pytest
+from autogalaxy import exc
+
+
+grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
+
+
+def test__deflections_yx_2d_from__grid_coordinates_overlap_image_grid_of_deflections():
+
+    deflections_y = ag.Array2D.manual_native(
+        [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]],
+        pixel_scales=0.1,
+        origin=(0.0, 0.0),
+    )
+    deflections_x = ag.Array2D.manual_native(
+        [[9.0, 8.0, 7.0], [6.0, 5.0, 4.0], [3.0, 2.0, 1.0]],
+        pixel_scales=0.1,
+        origin=(0.0, 0.0),
+    )
+
+    image_plane_grid = ag.Grid2D.uniform(
+        shape_native=deflections_y.shape_native,
+        pixel_scales=deflections_y.pixel_scales,
+    )
+
+    input_deflections = ag.mp.InputDeflections(
+        deflections_y=deflections_y,
+        deflections_x=deflections_x,
+        image_plane_grid=image_plane_grid,
+    )
+
+    grid = ag.Grid2D.uniform(
+        shape_native=deflections_y.shape_native,
+        pixel_scales=deflections_y.pixel_scales,
+    )
+
+    deflections = input_deflections.deflections_yx_2d_from(grid=grid)
+
+    assert deflections[:, 0] == pytest.approx(deflections_y, 1.0e-4)
+    assert deflections[:, 1] == pytest.approx(deflections_x, 1.0e-4)
+
+    grid = ag.Grid2D.manual_slim(
+        grid=np.array([[0.1, 0.0], [0.0, 0.0], [-0.1, -0.1]]),
+        shape_native=deflections_y.shape_native,
+        pixel_scales=deflections_y.pixel_scales,
+    )
+
+    deflections = input_deflections.deflections_yx_2d_from(grid=grid)
+
+    assert deflections[:, 0] == pytest.approx([2.0, 5.0, 7.0], 1.0e-4)
+    assert deflections[:, 1] == pytest.approx([8.0, 5.0, 3.0], 1.0e-4)
+
+    # input_deflections = ag.mp.InputDeflections(
+    #     deflections_y=deflections_y,
+    #     deflections_x=deflections_x,
+    #     image_plane_grid=image_plane_grid,
+    #     normalization_scale=2.0,
+    # )
+    #
+    # deflections = input_deflections.deflections_yx_2d_from(grid=grid)
+    #
+    # assert deflections[:, 0] == pytest.approx([4.0, 10.0, 14.0], 1.0e-4)
+    # assert deflections[:, 1] == pytest.approx([16.0, 10.0, 6.0], 1.0e-4)
+
+
+def test__deflections_yx_2d_from__grid_coordinates_dont_overlap_image_grid_of_deflections__uses_interpolation():
+
+    deflections_y = ag.Array2D.manual_native(
+        [[1.0, 2.0, 3.0], [4.0, 5.0, 6.0], [7.0, 8.0, 9.0]],
+        pixel_scales=0.1,
+        origin=(0.0, 0.0),
+    )
+    deflections_x = ag.Array2D.manual_native(
+        [[9.0, 8.0, 7.0], [6.0, 5.0, 4.0], [3.0, 2.0, 1.0]],
+        pixel_scales=0.1,
+        origin=(0.0, 0.0),
+    )
+
+    image_plane_grid = ag.Grid2D.uniform(
[…]
