+++ test_autogalaxy/profiles/mass_profiles/total/test_power_law_cored.py
@@ -1,0 +1,178 @@
+import numpy as np
+import pytest
+
+import autogalaxy as ag
+
+grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
+
+
+def test__deflections_yx_2d_from():
+
+    power_law = ag.mp.SphPowerLawCored(
+        centre=(-0.7, 0.5), einstein_radius=1.0, slope=1.8, core_radius=0.2
+    )
+
+    deflections = power_law.deflections_yx_2d_from(grid=np.array([[0.1875, 0.1625]]))
+
+    assert deflections[0, 0] == pytest.approx(0.80677, 1e-3)
+    assert deflections[0, 1] == pytest.approx(-0.30680, 1e-3)
+
+    power_law = ag.mp.SphPowerLawCored(
+        centre=(0.2, -0.2), einstein_radius=0.5, slope=2.4, core_radius=0.5
+    )
+
+    deflections = power_law.deflections_yx_2d_from(grid=np.array([[0.1875, 0.1625]]))
+
+    assert deflections[0, 0] == pytest.approx(-0.00321, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.09316, 1e-3)
+
+    cored_power_law = ag.mp.EllPowerLawCored(
+        centre=(-0.7, 0.5),
+        elliptical_comps=(0.152828, -0.088235),
+        einstein_radius=1.3,
+        slope=1.8,
+        core_radius=0.2,
+    )
+
+    deflections = cored_power_law.deflections_yx_2d_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(0.9869, 1e-3)
+    assert deflections[0, 1] == pytest.approx(-0.54882, 1e-3)
+
+    cored_power_law = ag.mp.EllPowerLawCored(
+        centre=(0.2, -0.2),
+        elliptical_comps=(-0.216506, -0.125),
+        einstein_radius=0.5,
+        slope=2.4,
+        core_radius=0.5,
+    )
+
+    deflections = cored_power_law.deflections_yx_2d_from(
+        grid=np.array([[0.1625, 0.1625]])
+    )
+    assert deflections[0, 0] == pytest.approx(0.01111, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.11403, 1e-3)
+
+    elliptical = ag.mp.EllPowerLawCored(
+        centre=(1.1, 1.1),
+        elliptical_comps=(0.0, 0.0),
+        einstein_radius=3.0,
+        slope=2.2,
+        core_radius=0.1,
+    )
+    spherical = ag.mp.SphPowerLawCored(
+        centre=(1.1, 1.1), einstein_radius=3.0, slope=2.2, core_radius=0.1
+    )
+
+    assert elliptical.deflections_yx_2d_from(grid=grid) == pytest.approx(
+        spherical.deflections_yx_2d_from(grid=grid), 1e-4
+    )
+
+
+def test__convergence_2d_from():
+
+    cored_power_law = ag.mp.SphPowerLawCored(
+        centre=(1, 1), einstein_radius=1.0, slope=2.2, core_radius=0.1
+    )
+
+    convergence = cored_power_law.convergence_func(grid_radius=1.0)
+
+    assert convergence == pytest.approx(0.39762, 1e-4)
+
+    cored_power_law = ag.mp.EllPowerLawCored(
+        centre=(0.0, 0.0),
+        elliptical_comps=(0.0, 0.333333),
+        einstein_radius=1.0,
+        slope=2.3,
+        core_radius=0.2,
+    )
+
[…]
