+++ test_autogalaxy/profiles/mass_profiles/total/test_power_law_broken.py
@@ -1,0 +1,203 @@
+import numpy as np
+import pytest
+
+import autogalaxy as ag
+
+grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
+
+
+def test__deflections_yx_2d_from():
+
+    broken_power_law = ag.mp.SphPowerLawBroken(
+        centre=(0, 0),
+        einstein_radius=1.0,
+        inner_slope=1.5,
+        outer_slope=2.5,
+        break_radius=0.1,
+    )
+
+    deflections = broken_power_law.deflections_yx_2d_from(grid=np.array([[0.5, 1.0]]))
+
+    assert deflections[0, 0] == pytest.approx(0.404076, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.808152, 1e-3)
+
+    deflections = broken_power_law.deflections_yx_2d_from(
+        grid=np.array([[0.5, 1.0], [0.5, 1.0]])
+    )
+
+    assert deflections[0, 0] == pytest.approx(0.404076, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.808152, 1e-3)
+    assert deflections[1, 0] == pytest.approx(0.404076, 1e-3)
+    assert deflections[1, 1] == pytest.approx(0.808152, 1e-3)
+
+    broken_power_law = ag.mp.EllPowerLawBroken(
+        centre=(0, 0),
+        elliptical_comps=(0.096225, 0.055555),
+        einstein_radius=1.0,
+        inner_slope=1.5,
+        outer_slope=2.5,
+        break_radius=0.1,
+    )
+
+    deflections = broken_power_law.deflections_yx_2d_from(grid=np.array([[0.5, 1.0]]))
+
+    assert deflections[0, 0] == pytest.approx(0.40392, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.811619, 1e-3)
+
+    broken_power_law = ag.mp.EllPowerLawBroken(
+        centre=(0, 0),
+        elliptical_comps=(-0.07142, -0.085116),
+        einstein_radius=1.0,
+        inner_slope=1.5,
+        outer_slope=2.5,
+        break_radius=0.1,
+    )
+
+    deflections = broken_power_law.deflections_yx_2d_from(grid=np.array([[0.5, 1.0]]))
+
+    assert deflections[0, 0] == pytest.approx(0.4005338, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.8067221, 1e-3)
+
+    broken_power_law = ag.mp.EllPowerLawBroken(
+        centre=(0, 0),
+        elliptical_comps=(0.109423, 0.019294),
+        einstein_radius=1.0,
+        inner_slope=1.5,
+        outer_slope=2.5,
+        break_radius=0.1,
+    )
+
+    deflections = broken_power_law.deflections_yx_2d_from(grid=np.array([[0.5, 1.0]]))
+
+    assert deflections[0, 0] == pytest.approx(0.399651, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.813372, 1e-3)
+
+    broken_power_law = ag.mp.EllPowerLawBroken(
+        centre=(0, 0),
+        elliptical_comps=(-0.216506, -0.125),
+        einstein_radius=1.0,
+        inner_slope=1.5,
+        outer_slope=2.5,
+        break_radius=0.1,
+    )
+
+    deflections = broken_power_law.deflections_yx_2d_from(grid=np.array([[0.5, 1.0]]))
+
+    assert deflections[0, 0] == pytest.approx(0.402629, 1e-3)
+    assert deflections[0, 1] == pytest.approx(0.798795, 1e-3)
+
+
[…]

[…]
