+++ test_autogalaxy/profiles/mass_profiles/abstract/test_abstract.py
@@ -1,0 +1,634 @@
+import math
+import numpy as np
+import pytest
+
+import autogalaxy as ag
+
+from autogalaxy import exc
+
+
+def mass_within_radius_of_profile_from_grid_calculation(radius, profile):
+
+    mass_total = 0.0
+
+    xs = np.linspace(-radius * 1.5, radius * 1.5, 40)
+    ys = np.linspace(-radius * 1.5, radius * 1.5, 40)
+
+    edge = xs[1] - xs[0]
+    area = edge**2
+
+    for x in xs:
+        for y in ys:
+
+            eta = profile.grid_to_elliptical_radii(grid=np.array([[x, y]]))
+
+            if eta < radius:
+                mass_total += profile.convergence_func(eta) * area
+
+    return mass_total
+
+
+def test__deflections_2d_via_potential_2d_from():
+
+    sis = ag.mp.SphIsothermal(centre=(0.0, 0.0), einstein_radius=2.0)
+
+    grid = ag.Grid2D.uniform(shape_native=(10, 10), pixel_scales=0.05, sub_size=1)
+
+    deflections_via_calculation = sis.deflections_yx_2d_from(grid=grid)
+
+    deflections_via_potential = sis.deflections_2d_via_potential_2d_from(grid=grid)
+
+    mean_error = np.mean(
+        deflections_via_potential.slim - deflections_via_calculation.slim
+    )
+
+    assert mean_error < 1e-4
+
+    sie = ag.mp.EllIsothermal(
+        centre=(0.0, 0.0), elliptical_comps=(0.111111, 0.0), einstein_radius=2.0
+    )
+
+    grid = ag.Grid2D.uniform(shape_native=(10, 10), pixel_scales=0.05, sub_size=1)
+
+    deflections_via_calculation = sie.deflections_yx_2d_from(grid=grid)
+
+    deflections_via_potential = sie.deflections_2d_via_potential_2d_from(grid=grid)
+
+    mean_error = np.mean(
+        deflections_via_potential.slim - deflections_via_calculation.slim
+    )
+
+    assert mean_error < 1e-4
+
+    sie = ag.mp.EllIsothermal(
+        centre=(0.0, 0.0), elliptical_comps=(0.0, -0.111111), einstein_radius=2.0
+    )
+
+    grid = ag.Grid2D.uniform(shape_native=(10, 10), pixel_scales=0.05, sub_size=1)
+
+    deflections_via_calculation = sie.deflections_yx_2d_from(grid=grid)
+
+    deflections_via_potential = sie.deflections_2d_via_potential_2d_from(grid=grid)
+
+    mean_error = np.mean(
+        deflections_via_potential.slim - deflections_via_calculation.slim
+    )
+
+    assert mean_error < 1e-4
+
+
+def test__mass_angular_within_circle_from():
+
+    sis = ag.mp.SphIsothermal(einstein_radius=2.0)
+
+    mass = sis.mass_angular_within_circle_from(radius=2.0)
+    assert math.pi * sis.einstein_radius * 2.0 == pytest.approx(mass, 1e-3)
+
+    sis = ag.mp.SphIsothermal(einstein_radius=4.0)
+
+    mass = sis.mass_angular_within_circle_from(radius=4.0)
+    assert math.pi * sis.einstein_radius * 4.0 == pytest.approx(mass, 1e-3)
+
+    sis = ag.mp.SphIsothermal(einstein_radius=2.0)
+
+    mass_grid = mass_within_radius_of_profile_from_grid_calculation(
+        radius=1.0, profile=sis
+    )
+
+    mass = sis.mass_angular_within_circle_from(radius=1.0)
+
+    assert mass_grid == pytest.approx(mass, 0.02)
+
+
[…]
