+++ test_autogalaxy/profiles/light_profiles/test_light_profile_decorators.py
@@ -5,7 +5,7 @@ from typing import Optional
 import autoarray as aa
 import autogalaxy as ag
 
-from autogalaxy.profiles.light_profiles.light_profile_decorators import (
+from autogalaxy.profiles.light_profiles.decorators import (
     check_operated_only,
 )
 
