+++ autogalaxy/plane/plot/plane_plotters.py
@@ -62,7 +62,7 @@ class PlanePlotter(Plotter):
             Specifies which attributes of the `MassProfile` are extracted and plotted as visuals for 2D plots.
         """
 
-        from autogalaxy.profiles.light_profiles.light_profiles_linear import (
+        from autogalaxy.profiles.light_profiles.linear import (
             LightProfileLinear,
         )
 
