+++ autogalaxy/profiles/mass_profiles/abstract/mge.py
@@ -1,17 +1,5 @@
-from abc import ABC, abstractmethod
 import numpy as np
-from scipy.integrate import quad
-from scipy.optimize import root_scalar
-from scipy.special import wofz, comb
-from scipy.linalg import lstsq
-from typing import Callable, List, Optional, Tuple
-
-import autoarray as aa
-
-from autogalaxy.profiles.geometry_profiles import EllProfile
-from autogalaxy.operate.deflections import OperateDeflections
-
-from autogalaxy import exc
+from scipy.special import comb
 
 
 def w_f_approx(z):
