+++ autogalaxy/profiles/mass_profiles/dark/nfw_truncated_mcr_scatter.py
@@ -1,30 +1,8 @@
-from astropy import cosmology as cosmo
-from astropy import units
-from colossus.cosmology import cosmology as col_cosmology
-from colossus.halo.concentration import concentration as col_concentration
-import copy
-import inspect
-import numpy as np
-from scipy import LowLevelCallable
-from scipy import special
-from scipy.integrate import quad
-from scipy.optimize import fsolve
-import warnings
 from typing import Tuple
 
-import autoarray as aa
+from autogalaxy.profiles.mass_profiles.dark.nfw_truncated import SphNFWTruncated
 
-from autogalaxy.profiles.mass_profiles.dark.abstract import DarkProfile
-from autogalaxy.profiles.mass_profiles import MassProfile
-from autogalaxy.cosmology.lensing import LensingCosmology
-from autogalaxy.cosmology.wrap import Planck15
-
-from autogalaxy.profiles.mass_profiles.mass_profiles import (
-    MassProfileMGE,
-    MassProfileCSE,
-)
-
-from autogalaxy import exc
+from autogalaxy.profiles.mass_profiles.dark import mcr_util
 
 
 class SphNFWTruncatedMCRScatterLudlow(SphNFWTruncated):
@@ -42,7 +20,11 @@ class SphNFWTruncatedMCRScatterLudlow(SphNFWTruncated):
         self.redshift_object = redshift_object
         self.redshift_source = redshift_source
 
-        kappa_s, scale_radius, radius_at_200 = kappa_s_and_scale_radius_for_ludlow(
+        (
+            kappa_s,
+            scale_radius,
+            radius_at_200,
+        ) = mcr_util.kappa_s_and_scale_radius_for_ludlow(
             mass_at_200=mass_at_200,
             scatter_sigma=scatter_sigma,
             redshift_object=redshift_object,
