+++ autogalaxy/profiles/mass_profiles/total/__init__.py
@@ -1,0 +1,6 @@
+from .isothermal import EllIsothermal, SphIsothermal
+from .isothermal_cored import EllIsothermalCored, SphIsothermalCored
+from .point import PointMass
+from .power_law import EllPowerLaw, SphPowerLaw
+from .power_law_broken import EllPowerLawBroken, SphPowerLawBroken
+from .power_law_cored import EllPowerLawCored, SphPowerLawCored
