+++ autogalaxy/profiles/mass_profiles/total/power_law_cored.py
@@ -1,7 +1,6 @@
 import copy
 import numpy as np
 from scipy.integrate import quad
-from scipy import special
 from typing import Tuple
 
 import autoarray as aa
@@ -43,8 +42,10 @@ class EllPowerLawCored(MassProfile):
 
     @property
     def einstein_radius_rescaled(self):
-        """Rescale the einstein radius by slope and axis_ratio, to reduce its degeneracy with other mass-profiles
-        parameters"""
+        """
+        Rescale the einstein radius by slope and axis_ratio, to reduce its degeneracy with other mass-profiles
+        parameters.
+        """
         return ((3 - self.slope) / (1 + self.axis_ratio)) * self.einstein_radius ** (
             self.slope - 1
         )
