+++ autogalaxy/profiles/light_profiles/snr/__init__.py
@@ -1,14 +1,11 @@
 from .abstract import LightProfileSNR
 from .gaussian import EllGaussian, SphGaussian
-from .moffat import EllMoffat, SphMoffat
 from .sersic import (
     EllSersic,
     SphSersic,
 )
 from .exponential import EllExponential, SphExponential
 from .dev_vaucouleurs import EllDevVaucouleurs, SphDevVaucouleurs
-from .sersic_core import EllSersicCore, SphSersicCore
-from .exponential_core import EllExponentialCore, SphExponentialCore
 from .chameleon import EllChameleon, SphChameleon
 from .eff import (
     EllEff,
