+++ autogalaxy/profiles/light_profiles/operated/moffat.py
@@ -1,5 +1,7 @@
 from autogalaxy.profiles.light_profiles import base as lp
 
+from autogalaxy.profiles.light_profiles.operated.abstract import LightProfileOperated
+
 
 class EllMoffat(lp.EllMoffat, LightProfileOperated):
 
