+++ autogalaxy/analysis/model_util.py
@@ -9,7 +9,7 @@ import autoarray as aa
 from autoarray.inversion.pixelization.mesh.abstract import AbstractMesh
 from autoarray.inversion.regularization.abstract import AbstractRegularization
 
-from autogalaxy.profiles.light_profiles.light_profiles import LightProfile
+from autogalaxy.profiles.light_profiles.base.abstract import LightProfile
 from autogalaxy.profiles.mass_profiles.abstract.abstract import MassProfile
 
 from autogalaxy import exc
