+++ test_autogalaxy/profiles/test_dict.py
@@ -12,7 +12,7 @@ def make_ell_sersic():
 @pytest.fixture(name="ell_sersic_dict")
 def make_ell_sersic_dict():
     return {
-        "type": "autogalaxy.profiles.mass_profiles.stellar_mass_profiles.EllSersic",
+        "type": "autogalaxy.profiles.mass_profiles.stellar.sersic.EllSersic",
         "centre": (0.0, 0.0),
         "elliptical_comps": (0.0, 0.0),
         "intensity": 0.1,
