+++ autogalaxy/profiles/light_profiles/operated/gaussian.py
@@ -1,5 +1,7 @@
 from autogalaxy.profiles.light_profiles import base as lp
 
+from autogalaxy.profiles.light_profiles.operated.abstract import LightProfileOperated
+
 
 class EllGaussian(lp.EllGaussian, LightProfileOperated):
 
