+++ autogalaxy/profiles/mass/dark/nfw_mcr.py
@@ -1,0 +1,90 @@
+from typing import Tuple
+
+from autogalaxy.profiles.mass.dark.nfw import SphNFW
+from autogalaxy.profiles.mass.dark.nfw import EllNFW
+from autogalaxy.profiles.mass.dark.nfw_mcr_scatter import (
+    SphNFWMCRScatterLudlow,
+)
+
+from autogalaxy.profiles.mass.dark import mcr_util
+
+
+class SphNFWMCRDuffy(SphNFW):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        mass_at_200: float = 1e9,
+        redshift_object: float = 0.5,
+        redshift_source: float = 1.0,
+    ):
+
+        self.mass_at_200 = mass_at_200
+        self.redshift_object = redshift_object
+        self.redshift_source = redshift_source
+
+        (
+            kappa_s,
+            scale_radius,
+            radius_at_200,
+        ) = mcr_util.kappa_s_and_scale_radius_for_duffy(
+            mass_at_200=mass_at_200,
+            redshift_object=redshift_object,
+            redshift_source=redshift_source,
+        )
+
+        super().__init__(centre=centre, kappa_s=kappa_s, scale_radius=scale_radius)
+
+    def with_new_normalization(self, normalization):
+
+        raise NotImplementedError()
+
+
+class EllNFWMCRLudlow(EllNFW):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        mass_at_200: float = 1e9,
+        redshift_object: float = 0.5,
+        redshift_source: float = 1.0,
+    ):
+
+        self.mass_at_200 = mass_at_200
+        self.redshift_object = redshift_object
+        self.redshift_source = redshift_source
+
+        (
+            kappa_s,
+            scale_radius,
+            radius_at_200,
+        ) = mcr_util.kappa_s_and_scale_radius_for_ludlow(
+            mass_at_200=mass_at_200,
+            scatter_sigma=0.0,
+            redshift_object=redshift_object,
+            redshift_source=redshift_source,
+        )
+
+        super().__init__(
+            centre=centre,
+            elliptical_comps=elliptical_comps,
+            kappa_s=kappa_s,
+            scale_radius=scale_radius,
+        )
+
+
+class SphNFWMCRLudlow(SphNFWMCRScatterLudlow):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        mass_at_200: float = 1e9,
+        redshift_object: float = 0.5,
+        redshift_source: float = 1.0,
+    ):
+
+        super().__init__(
+            centre=centre,
+            mass_at_200=mass_at_200,
+            scatter_sigma=0.0,
+            redshift_object=redshift_object,
+            redshift_source=redshift_source,
+        )
