### 2022-11-16 d6d0633
remove intensity break
