+++ autogalaxy/profiles/light/light_profiles.py
@@ -138,10 +138,6 @@ class LightProfile(EllProfile, OperateImage):
     def lp_instance_from(self, intensity: float) -> "LightProfile":
         raise NotImplementedError
 
-    @property
-    def _intensity(self):
-        return self.intensity
-
 
 class EllGaussian(LightProfile):
     def __init__(
@@ -724,7 +720,7 @@ class EllSersicCore(EllSersic):
         effective_radius: float = 0.6,
         sersic_index: float = 4.0,
         radius_break: float = 0.01,
-        intensity_break: float = 0.05,
+        intensity: float = 0.05,
         gamma: float = 0.25,
         alpha: float = 3.0,
     ):
@@ -746,7 +742,7 @@ class EllSersicCore(EllSersic):
             Controls the concentration of the profile (lower -> less concentrated, higher -> more concentrated).
         radius_break
             The break radius separating the inner power-law (with logarithmic slope gamma) and outer Sersic function.
-        intensity_break
+        intensity
             The intensity at the break radius.
         gamma
             The logarithmic power-law slope of the inner core profiles
@@ -757,19 +753,15 @@ class EllSersicCore(EllSersic):
         super().__init__(
             centre=centre,
             elliptical_comps=elliptical_comps,
-            intensity=intensity_break,
+            intensity=intensity,
             effective_radius=effective_radius,
             sersic_index=sersic_index,
         )
 
         self.radius_break = radius_break
-        self.intensity_break = intensity_break
+        self.intensity = intensity
         self.alpha = alpha
         self.gamma = gamma
-
-    @property
-    def _intensity(self):
-        return self.intensity_break
 
     @property
     def intensity_prime(self) -> float:
@@ -841,7 +833,7 @@ class SphSersicCore(EllSersicCore):
         effective_radius: float = 0.6,
         sersic_index: float = 4.0,
         radius_break: float = 0.01,
-        intensity_break: float = 0.05,
+        intensity: float = 0.05,
         gamma: float = 0.25,
         alpha: float = 3.0,
     ):
@@ -860,7 +852,7 @@ class SphSersicCore(EllSersicCore):
             Controls the concentration of the profile (lower -> less concentrated, higher -> more concentrated).
         radius_break
             The break radius separating the inner power-law (with logarithmic slope gamma) and outer Sersic function.
-        intensity_break
+        intensity
             The intensity at the break radius.
         gamma
             The logarithmic power-law slope of the inner core profiles
@@ -874,13 +866,13 @@ class SphSersicCore(EllSersicCore):
             effective_radius=effective_radius,
             sersic_index=sersic_index,
             radius_break=radius_break,
-            intensity_break=intensity_break,
+            intensity=intensity,
             gamma=gamma,
             alpha=alpha,
         )
 
         self.radius_break = radius_break
-        self.intensity_break = intensity_break
+        self.intensity = intensity
         self.alpha = alpha
         self.gamma = gamma
 
@@ -892,7 +884,7 @@ class EllExponentialCore(EllSersicCore):
         elliptical_comps: Tuple[float, float] = (0.0, 0.0),
         effective_radius: float = 0.6,
         radius_break: float = 0.01,
-        intensity_break: float = 0.05,
+        intensity: float = 0.05,
         gamma: float = 0.25,
         alpha: float = 3.0,
     ):
@@ -914,7 +906,7 @@ class EllExponentialCore(EllSersicCore):
             Controls the concentration of the profile (lower -> less concentrated, higher -> more concentrated).
         radius_break
             The break radius separating the inner power-law (with logarithmic slope gamma) and outer Sersic function.
-        intensity_break
+        intensity
             The intensity at the break radius.
         gamma
[…]
