+++ autogalaxy/profiles/light/base/exponential_core.py
@@ -1,99 +1,0 @@
-from typing import Tuple
-
-from autogalaxy.profiles.light.base.sersic_core import EllSersicCore
-
-
-class EllExponentialCore(EllSersicCore):
-    def __init__(
-        self,
-        centre: Tuple[float, float] = (0.0, 0.0),
-        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
-        effective_radius: float = 0.6,
-        radius_break: float = 0.01,
-        intensity: float = 0.05,
-        gamma: float = 0.25,
-        alpha: float = 3.0,
-    ):
-        """
-        The elliptical cored-Exponential light profile.
-
-        See `autogalaxy.profiles.light.light_profiles.LightProfile` for a description of light profile objects.
-
-        Parameters
-        ----------
-        centre
-            The (y,x) arc-second coordinates of the profile centre.
-        elliptical_comps
-            The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
-        effective_radius
-            The circular radius containing half the light of this profile.
-        sersic_index
-            Controls the concentration of the profile (lower -> less concentrated, higher -> more concentrated).
-        radius_break
-            The break radius separating the inner power-law (with logarithmic slope gamma) and outer Sersic function.
-        intensity
-            The intensity at the break radius.
-        gamma
-            The logarithmic power-law slope of the inner core profiles
-        alpha :
-            Controls the sharpness of the transition between the inner core / outer Sersic profiles.
-        """
-
-        super().__init__(
-            centre=centre,
-            elliptical_comps=elliptical_comps,
-            intensity=intensity,
-            effective_radius=effective_radius,
-            sersic_index=1.0,
-            radius_break=radius_break,
-            gamma=gamma,
-            alpha=alpha,
-        )
-
-
-class SphExponentialCore(EllExponentialCore):
-    def __init__(
-        self,
-        centre: Tuple[float, float] = (0.0, 0.0),
-        effective_radius: float = 0.6,
-        radius_break: float = 0.01,
-        intensity: float = 0.05,
-        gamma: float = 0.25,
-        alpha: float = 3.0,
-    ):
-        """
-        The elliptical cored-Exponential light profile.
-
-        See `autogalaxy.profiles.light.light_profiles.LightProfile` for a description of light profile objects.
-
-        Parameters
-        ----------
-        centre
-            The (y,x) arc-second coordinates of the profile centre.
-        effective_radius
-            The circular radius containing half the light of this profile.
-        radius_break
-            The break radius separating the inner power-law (with logarithmic slope gamma) and outer Sersic function.
-        intensity
-            The intensity at the break radius.
-        gamma
-            The logarithmic power-law slope of the inner core profiles
-        alpha :
-            Controls the sharpness of the transition between the inner core / outer Sersic profiles.
-        """
-
-        super().__init__(
-            centre=centre,
-            elliptical_comps=(0.0, 0.0),
-            effective_radius=effective_radius,
-            radius_break=radius_break,
-            intensity=intensity,
-            gamma=gamma,
-            alpha=alpha,
-        )
-
-        self.radius_break = radius_break
-        self.intensity = intensity
-        self.alpha = alpha
-        self.gamma = gamma
