+++ autogalaxy/profiles/light/standard/sersic.py
@@ -1,0 +1,211 @@
+import numpy as np
+from typing import Optional, Tuple
+
+import autoarray as aa
+
+from autogalaxy.profiles.light.base.abstract import LightProfile
+from autogalaxy.profiles.light.decorators import (
+    check_operated_only,
+)
+
+
+class AbstractEllSersic(LightProfile):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        effective_radius: float = 0.6,
+        sersic_index: float = 4.0,
+    ):
+        """
+        Abstract base class for elliptical Sersic light profiles.
+
+        See `autogalaxy.profiles.light.light_profiles.LightProfile` for a description of light profile objects.
+
+        Parameters
+        ----------
+        centre
+            The (y,x) arc-second coordinates of the profile centre.
+        elliptical_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        intensity
+            Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
+            the light profile's image is compared too, which is expected to be electrons per second).
+        effective_radius
+            The circular radius containing half the light of this light profile.
+        sersic_index
+            Controls the concentration of the profile (lower -> less concentrated, higher -> more concentrated).
+        """
+        super().__init__(
+            centre=centre, elliptical_comps=elliptical_comps, intensity=intensity
+        )
+        self.effective_radius = effective_radius
+        self.sersic_index = sersic_index
+
+    @property
+    def elliptical_effective_radius(self) -> float:
+        """
+        The `effective_radius` of a Sersic light profile is defined as the circular effective radius, which is the
+        radius within which a circular aperture contains half the profile's total integrated light.
+
+        For elliptical systems, this will not robustly capture the light profile's elliptical shape.
+
+        The elliptical effective radius instead describes the major-axis radius of the ellipse containing
+        half the light, and may be more appropriate for highly flattened systems like disk galaxies.
+        """
+        return self.effective_radius / np.sqrt(self.axis_ratio)
+
+    @property
+    def sersic_constant(self) -> float:
+        """
+        A parameter derived from Sersic index which ensures that effective radius contains 50% of the profile's
+        total integrated light.
+        """
+        return (
+            (2 * self.sersic_index)
+            - (1.0 / 3.0)
+            + (4.0 / (405.0 * self.sersic_index))
+            + (46.0 / (25515.0 * self.sersic_index**2))
+            + (131.0 / (1148175.0 * self.sersic_index**3))
+            - (2194697.0 / (30690717750.0 * self.sersic_index**4))
+        )
+
+    def image_2d_via_radii_from(self, radius: np.ndarray) -> np.ndarray:
+        """
+        Returns the 2D image of the Sersic light profile from a grid of coordinates which are the radial distances of
+        each coordinate from the its `centre`.
+
+        Parameters
+        ----------
+        grid_radii
+            The radial distances from the centre of the profile, for each coordinate on the grid.
+        """
+        return self._intensity * np.exp(
+            -self.sersic_constant
+            * (((radius / self.effective_radius) ** (1.0 / self.sersic_index)) - 1)
+        )
+
+
+class EllSersic(AbstractEllSersic, LightProfile):
+    def __init__(
+        self,
+        centre: Tuple[float, float] = (0.0, 0.0),
+        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        intensity: float = 0.1,
+        effective_radius: float = 0.6,
+        sersic_index: float = 4.0,
+    ):
+        """
+        The elliptical Sersic light profile.
+
[…]
