+++ autogalaxy/profiles/light/operated/sersic.py
@@ -1,4 +1,4 @@
-from autogalaxy.profiles.light import base as lp
+from autogalaxy.profiles.light import standard as lp
 
 from autogalaxy.profiles.light.operated.abstract import LightProfileOperated
 
