+++ autogalaxy/profiles/light/linear/gaussian.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 from autogalaxy.profiles.light.linear.abstract import LightProfileLinear
 
-from autogalaxy.profiles.light import base as lp
+from autogalaxy.profiles.light import standard as lp
 from autogalaxy.profiles import light_and_mass_profiles as lmp
 
 
