+++ autogalaxy/analysis/clump_model.py
@@ -4,7 +4,7 @@ import autofit as af
 import autoarray as aa
 
 from autogalaxy.galaxy.galaxy import Galaxy
-from autogalaxy.profiles.light.base.abstract import LightProfile
+from autogalaxy.profiles.light.standard.abstract import LightProfile
 from autogalaxy.profiles.mass.abstract.abstract import MassProfile
 
 
