+++ autogalaxy/profiles/light/snr/dev_vaucouleurs.py
@@ -1,7 +1,7 @@
 from typing import Tuple
 
 from autogalaxy.profiles.light.snr.abstract import LightProfileSNR
-from autogalaxy.profiles.light import base as lp
+from autogalaxy.profiles.light import standard as lp
 
 
 class EllDevVaucouleurs(lp.EllDevVaucouleurs, LightProfileSNR):
