+++ autogalaxy/profiles/light/standard/eff.py
@@ -3,7 +3,7 @@ from typing import Optional, Tuple
 
 import autoarray as aa
 
-from autogalaxy.profiles.light.standard.abstract import LightProfile
+from autogalaxy.profiles.light.abstract import LightProfile
 from autogalaxy.profiles.light.decorators import (
     check_operated_only,
 )
