### 2022-11-16 4e18ad4
mass renaming of light profiles
