+++ autogalaxy/profiles/light/standard/dev_vaucouleurs.py
@@ -1,9 +1,9 @@
 from typing import Tuple
 
-from autogalaxy.profiles.light.standard.sersic import EllSersic
+from autogalaxy.profiles.light.standard.sersic import Sersic
 
 
-class EllDevVaucouleurs(EllSersic):
+class DevVaucouleurs(Sersic):
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
@@ -40,7 +40,7 @@ class EllDevVaucouleurs(EllSersic):
         )
 
 
-class SphDevVaucouleurs(EllDevVaucouleurs):
+class DevVaucouleursSph(DevVaucouleurs):
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
