+++ autogalaxy/profiles/mass/stellar/__init__.py
@@ -1,7 +1,7 @@
-from .chameleon import EllChameleon, SphChameleon
-from .dev_vaucouleurs import EllDevVaucouleurs, SphDevVaucouleurs
-from .exponential import EllExponential, SphExponential
-from .gaussian import EllGaussian
-from .sersic import EllSersic, SphSersic
-from .sersic_core import EllSersicCore, SphSersicCore
-from .sersic_radial_gradient import EllSersicRadialGradient, SphSersicRadialGradient
+from .chameleon import Chameleon, ChameleonSph
+from .dev_vaucouleurs import DevVaucouleurs, DevVaucouleursSph
+from .exponential import Exponential, ExponentialSph
+from .gaussian import Gaussian
+from .sersic import Sersic, SersicSph
+from .sersic_core import SersicCore, SersicCoreSph
+from .sersic_radial_gradient import SersicRadialGradient, SphSersicRadialGradient
