+++ autogalaxy/profiles/light/operated/gaussian.py
@@ -3,6 +3,6 @@ from autogalaxy.profiles.light import standard as lp
 from autogalaxy.profiles.light.operated.abstract import LightProfileOperated
 
 
-class EllGaussian(lp.EllGaussian, LightProfileOperated):
+class Gaussian(lp.Gaussian, LightProfileOperated):
 
     pass
