+++ autogalaxy/profiles/light/linear_operated/gaussian.py
@@ -2,6 +2,6 @@ from autogalaxy.profiles.light import operated as lp_operated
 from autogalaxy.profiles.light import linear as lp_linear
 
 
-class EllGaussian(lp_linear.EllGaussian, lp_operated.LightProfileOperated):
+class Gaussian(lp_linear.Gaussian, lp_operated.LightProfileOperated):
 
     pass
