+++ autogalaxy/profiles/light/linear_operated/sersic.py
@@ -2,6 +2,6 @@ from autogalaxy.profiles.light import operated as lp_operated
 from autogalaxy.profiles.light import linear as lp_linear
 
 
-class EllSersic(lp_linear.EllSersic, lp_operated.LightProfileOperated):
+class Sersic(lp_linear.Sersic, lp_operated.LightProfileOperated):
 
     pass
