+++ test_autogalaxy/aggregator/test_aggregator_interferometer.py
@@ -21,16 +21,16 @@ def make_path():
 def make_model():
     return af.Collection(
         galaxies=af.Collection(
-            galaxy=af.Model(ag.Galaxy, redshift=0.5, light=ag.lp.EllSersic),
-            source=af.Model(ag.Galaxy, redshift=1.0, light=ag.lp.EllSersic),
+            galaxy=af.Model(ag.Galaxy, redshift=0.5, light=ag.lp.Sersic),
+            source=af.Model(ag.Galaxy, redshift=1.0, light=ag.lp.Sersic),
         )
     )
 
 
 @pytest.fixture(name="samples")
 def make_samples(model):
-    galaxy_0 = ag.Galaxy(redshift=0.5, light=ag.lp.EllSersic(centre=(0.0, 1.0)))
-    galaxy_1 = ag.Galaxy(redshift=1.0, light=ag.lp.EllSersic())
+    galaxy_0 = ag.Galaxy(redshift=0.5, light=ag.lp.Sersic(centre=(0.0, 1.0)))
+    galaxy_1 = ag.Galaxy(redshift=1.0, light=ag.lp.Sersic())
 
     plane = ag.Plane(galaxies=[galaxy_0, galaxy_1])
 
