+++ autogalaxy/profiles/light/linear_operated/moffat.py
@@ -2,6 +2,6 @@ from autogalaxy.profiles.light import operated as lp_operated
 from autogalaxy.profiles.light import linear as lp_linear
 
 
-class EllMoffat(lp_linear.EllMoffat, lp_operated.LightProfileOperated):
+class Moffat(lp_linear.Moffat, lp_operated.LightProfileOperated):
 
     pass
