+++ test_autogalaxy/profiles/mass/dark/test_nfw.py
@@ -7,7 +7,7 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 def test__deflections_via_integral_from():
-    nfw = ag.mp.SphNFW(centre=(0.0, 0.0), kappa_s=1.0, scale_radius=1.0)
+    nfw = ag.mp.NFWSph(centre=(0.0, 0.0), kappa_s=1.0, scale_radius=1.0)
 
     deflections = nfw.deflections_2d_via_integral_from(
         grid=np.array([[0.1625, 0.1625]])
@@ -16,7 +16,7 @@ def test__deflections_via_integral_from():
     assert deflections[0, 0] == pytest.approx(0.56194, 1e-3)
     assert deflections[0, 1] == pytest.approx(0.56194, 1e-3)
 
-    nfw = ag.mp.SphNFW(centre=(0.3, 0.2), kappa_s=2.5, scale_radius=4.0)
+    nfw = ag.mp.NFWSph(centre=(0.3, 0.2), kappa_s=2.5, scale_radius=4.0)
 
     deflections = nfw.deflections_2d_via_integral_from(
         grid=np.array([[0.1875, 0.1625]])
@@ -25,7 +25,7 @@ def test__deflections_via_integral_from():
     assert deflections[0, 0] == pytest.approx(-2.08909, 1e-3)
     assert deflections[0, 1] == pytest.approx(-0.69636, 1e-3)
 
-    nfw = ag.mp.EllNFW(
+    nfw = ag.mp.NFW(
         centre=(0.0, 0.0),
         elliptical_comps=(0.0, 0.0),
         kappa_s=1.0,
@@ -39,7 +39,7 @@ def test__deflections_via_integral_from():
     assert deflections[0, 0] == pytest.approx(0.56194, 1e-3)
     assert deflections[0, 1] == pytest.approx(0.56194, 1e-3)
 
-    nfw = ag.mp.EllNFW(
+    nfw = ag.mp.NFW(
         centre=(0.3, 0.2),
         elliptical_comps=(0.03669, 0.172614),
         kappa_s=2.5,
@@ -56,7 +56,7 @@ def test__deflections_via_integral_from():
 
 def test__deflections_2d_via_cse_from():
 
-    nfw = ag.mp.SphNFW(centre=(0.0, 0.0), kappa_s=1.0, scale_radius=1.0)
+    nfw = ag.mp.NFWSph(centre=(0.0, 0.0), kappa_s=1.0, scale_radius=1.0)
 
     deflections_via_integral = nfw.deflections_2d_via_integral_from(
         grid=np.array([[0.1625, 0.1625]])
@@ -67,7 +67,7 @@ def test__deflections_2d_via_cse_from():
 
     assert deflections_via_integral == pytest.approx(deflections_via_cse, 1.0e-4)
 
-    nfw = ag.mp.SphNFW(centre=(0.3, 0.2), kappa_s=2.5, scale_radius=4.0)
+    nfw = ag.mp.NFWSph(centre=(0.3, 0.2), kappa_s=2.5, scale_radius=4.0)
 
     deflections_via_integral = nfw.deflections_2d_via_integral_from(
         grid=np.array([[0.1625, 0.1625]])
@@ -78,7 +78,7 @@ def test__deflections_2d_via_cse_from():
 
     assert deflections_via_integral == pytest.approx(deflections_via_cse, 1.0e-4)
 
-    nfw = ag.mp.EllNFW(
[…]
