### 2022-11-16 790a02d
remove priors test config
