+++ test_autogalaxy/profiles/mass/total/test_isothermal_cored.py
@@ -8,7 +8,7 @@ grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 def test__deflections_yx_2d_from():
 
-    cored_isothermal = ag.mp.IsothermalCoredSph(
+    cored_isothermal = ag.mp.IsothermalCoreSph(
         centre=(-0.7, 0.5), einstein_radius=1.3, core_radius=0.2
     )
 
@@ -19,7 +19,7 @@ def test__deflections_yx_2d_from():
     assert deflections[0, 0] == pytest.approx(0.98582, 1e-3)
     assert deflections[0, 1] == pytest.approx(-0.37489, 1e-3)
 
-    cored_isothermal = ag.mp.IsothermalCoredSph(
+    cored_isothermal = ag.mp.IsothermalCoreSph(
         centre=(0.2, -0.2), einstein_radius=0.5, core_radius=0.5
     )
 
@@ -30,7 +30,7 @@ def test__deflections_yx_2d_from():
     assert deflections[0, 0] == pytest.approx(-0.00559, 1e-3)
     assert deflections[0, 1] == pytest.approx(0.16216, 1e-3)
 
-    cored_isothermal = ag.mp.IsothermalCored(
+    cored_isothermal = ag.mp.IsothermalCore(
         centre=(-0.7, 0.5),
         elliptical_comps=(0.152828, -0.088235),
         einstein_radius=1.3,
@@ -44,7 +44,7 @@ def test__deflections_yx_2d_from():
     assert deflections[0, 0] == pytest.approx(0.95429, 1e-3)
     assert deflections[0, 1] == pytest.approx(-0.52047, 1e-3)
 
-    cored_isothermal = ag.mp.IsothermalCored(
+    cored_isothermal = ag.mp.IsothermalCore(
         centre=(0.2, -0.2),
         elliptical_comps=(-0.216506, -0.125),
         einstein_radius=0.5,
@@ -58,13 +58,13 @@ def test__deflections_yx_2d_from():
     assert deflections[0, 0] == pytest.approx(0.02097, 1e-3)
     assert deflections[0, 1] == pytest.approx(0.20500, 1e-3)
 
-    elliptical = ag.mp.IsothermalCored(
+    elliptical = ag.mp.IsothermalCore(
         centre=(1.1, 1.1),
         elliptical_comps=(0.0, 0.0),
         einstein_radius=3.0,
         core_radius=1.0,
     )
-    spherical = ag.mp.IsothermalCoredSph(
+    spherical = ag.mp.IsothermalCoreSph(
         centre=(1.1, 1.1), einstein_radius=3.0, core_radius=1.0
     )
 
@@ -75,7 +75,7 @@ def test__deflections_yx_2d_from():
 
 def test__convergence_2d_from():
 
-    cored_isothermal = ag.mp.IsothermalCoredSph(
+    cored_isothermal = ag.mp.IsothermalCoreSph(
         centre=(1, 1), einstein_radius=1.0, core_radius=0.1
     )
 
@@ -83,7 +83,7 @@ def test__convergence_2d_from():
 
     assert convergence == pytest.approx(0.49752, 1e-4)
 
-    cored_isothermal = ag.mp.IsothermalCoredSph(
+    cored_isothermal = ag.mp.IsothermalCoreSph(
         centre=(1, 1), einstein_radius=1.0, core_radius=0.1
     )
 
@@ -91,7 +91,7 @@ def test__convergence_2d_from():
 
     assert convergence == pytest.approx(0.49752, 1e-4)
 
-    cored_isothermal = ag.mp.IsothermalCoredSph(
+    cored_isothermal = ag.mp.IsothermalCoreSph(
[…]
