+++ autogalaxy/profiles/mass/total/__init__.py
@@ -1,6 +1,6 @@
 from .isothermal import Isothermal, IsothermalSph
-from .isothermal_cored import IsothermalCored, IsothermalCoredSph
+from .isothermal_core import IsothermalCore, IsothermalCoreSph
 from .point import PointMass
 from .power_law import PowerLaw, PowerLawSph
 from .power_law_broken import PowerLawBroken, PowerLawBrokenSph
-from .power_law_cored import PowerLawCored, PowerLawCoredSph
+from .power_law_core import PowerLawCore, PowerLawCoreSph
