+++ test_autogalaxy/profiles/mass/total/test_isothermal.py
@@ -171,7 +171,7 @@ def test__compare_to_cored_power_law():
     isothermal = ag.mp.Isothermal(
         centre=(0.0, 0.0), elliptical_comps=(0.333333, 0.0), einstein_radius=1.0
     )
-    cored_power_law = ag.mp.PowerLawCored(
+    cored_power_law = ag.mp.PowerLawCore(
         centre=(0.0, 0.0),
         elliptical_comps=(0.333333, 0.0),
         einstein_radius=1.0,
