### 2022-11-16 #4475 @dfm
@alexandergagliano, @sloneil, @Jammy2211 — Hi folks, I wanted to check in here because we're recently switched to a "tracks" model here at JOSS and I'm now the managing editor of the track where this submission has been assigned. It looks like there hasn't been much activity on this review in a while - can you update me on what the status is and if there are any blockers in place that I can help with? Thanks!
