### 2022-11-25 1972168
aggregator tests simplified
