+++ autogalaxy/quantity/model/analysis.py
@@ -155,7 +155,7 @@ class AnalysisQuantity(Analysis):
 
     def make_result(
         self,
-        samples: af.PDFSamples,
+        samples: af.SamplesPDF,
         model: af.Collection,
         sigma=1.0,
         use_errors=True,
