+++ autogalaxy/interferometer/model/analysis.py
@@ -416,7 +416,7 @@ class AnalysisInterferometer(AnalysisDataset):
 
     def make_result(
         self,
-        samples: af.PDFSamples,
+        samples: af.SamplesPDF,
         model: af.Collection,
         sigma=1.0,
         use_errors=True,
