+++ docs/index.rst
@@ -190,7 +190,7 @@ grid or Voronoi mesh that accounts for irregular galaxy morphologies.
     The results contain information on the fit, for example the maximum likelihood
     model from the Dynesty parameter space search.
     """
-    print(result.samples.max_log_likelihood_instance)
+    print(result.samples.max_log_likelihood())
 
 .. toctree::
    :caption: Overview:
