+++ autogalaxy/profiles/light/linear/dev_vaucouleurs.py
@@ -10,7 +10,7 @@ class DevVaucouleurs(lp.DevVaucouleurs, LightProfileLinear):
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
-        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        ell_comps: Tuple[float, float] = (0.0, 0.0),
         effective_radius: float = 0.6,
     ):
         """
@@ -22,7 +22,7 @@ class DevVaucouleurs(lp.DevVaucouleurs, LightProfileLinear):
         ----------
         centre
             The (y,x) arc-second coordinates of the profile centre.
-        elliptical_comps
+        ell_comps
             The first and second ellipticity components of the elliptical coordinate system, (see the module
             `autogalaxy -> convert.py` for the convention).
         effective_radius
@@ -30,7 +30,7 @@ class DevVaucouleurs(lp.DevVaucouleurs, LightProfileLinear):
         """
         super().__init__(
             centre=centre,
-            elliptical_comps=elliptical_comps,
+            ell_comps=ell_comps,
             intensity=1.0,
             effective_radius=effective_radius,
         )
