+++ autogalaxy/profiles/light/standard/gaussian.py
@@ -13,7 +13,7 @@ class Gaussian(LightProfile):
     def __init__(
         self,
         centre: Tuple[float, float] = (0.0, 0.0),
-        elliptical_comps: Tuple[float, float] = (0.0, 0.0),
+        ell_comps: Tuple[float, float] = (0.0, 0.0),
         intensity: float = 0.1,
         sigma: float = 1.0,
     ):
@@ -26,7 +26,7 @@ class Gaussian(LightProfile):
         ----------
         centre
             The (y,x) arc-second coordinates of the profile centre.
-        elliptical_comps
+        ell_comps
             The first and second ellipticity components of the elliptical coordinate system, (see the module
             `autogalaxy -> convert.py` for the convention).
         intensity
@@ -36,9 +36,7 @@ class Gaussian(LightProfile):
             The sigma value of the Gaussian, corresponding to ~ 1 / sqrt(2 log(2)) the full width half maximum.
         """
 
-        super().__init__(
-            centre=centre, elliptical_comps=elliptical_comps, intensity=intensity
-        )
+        super().__init__(centre=centre, ell_comps=ell_comps, intensity=intensity)
         self.sigma = sigma
 
     def image_2d_via_radii_from(self, grid_radii: np.ndarray) -> np.ndarray:
@@ -112,5 +110,5 @@ class GaussianSph(Gaussian):
             The sigma value of the Gaussian, corresponding to ~ 1 / sqrt(2 log(2)) the full width half maximum.
         """
         super().__init__(
-            centre=centre, elliptical_comps=(0.0, 0.0), intensity=intensity, sigma=sigma
+            centre=centre, ell_comps=(0.0, 0.0), intensity=intensity, sigma=sigma
         )
