+++ test_autogalaxy/galaxy/test_dict.py
@@ -86,14 +86,14 @@ def make_profiles_galaxy_dict():
             "centre": (0.0, 0.0),
             "core_radius_0": 0.01,
             "core_radius_1": 0.05,
-            "elliptical_comps": (0.0, 0.0),
+            "ell_comps": (0.0, 0.0),
             "intensity": 0.1,
             "type": "autogalaxy.profiles.light.standard.chameleon.Chameleon",
         },
         "mass": {
             "centre": (0.0, 0.0),
             "effective_radius": 0.6,
-            "elliptical_comps": (0.0, 0.0),
+            "ell_comps": (0.0, 0.0),
             "intensity": 0.1,
             "mass_to_light_ratio": 1.0,
             "type": "autogalaxy.profiles.mass.stellar.dev_vaucouleurs.DevVaucouleurs",
