### 2022-11-25 12b0731
shear ell comps -> gamma
