+++ test_autogalaxy/profiles/mass/sheets/test_external_shear.py
@@ -1,20 +1,18 @@
 import autogalaxy as ag
 import numpy as np
 import pytest
-from autogalaxy import exc
-
 
 grid = np.array([[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [2.0, 4.0]])
 
 
 def test__deflections_yx_2d_from():
 
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, 0.0))
+    shear = ag.mp.ExternalShear(gamma=(0.1, 0.0))
     deflections = shear.deflections_yx_2d_from(grid=np.array([[0.1625, 0.1625]]))
     assert deflections[0, 0] == pytest.approx(0.01625, 1e-3)
     assert deflections[0, 1] == pytest.approx(0.01625, 1e-3)
 
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, -0.17320))
+    shear = ag.mp.ExternalShear(gamma=(0.1, -0.17320))
     deflections = shear.deflections_yx_2d_from(grid=np.array([[0.1625, 0.1625]]))
     assert deflections[0, 0] == pytest.approx(0.04439, 1e-3)
     assert deflections[0, 1] == pytest.approx(-0.011895, 1e-3)
@@ -47,11 +45,11 @@ def test__deflections_yx_2d_from():
 
 def test__convergence_returns_zeros():
 
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, 0.0))
+    shear = ag.mp.ExternalShear(gamma=(0.1, 0.0))
     convergence = shear.convergence_2d_from(grid=np.array([[0.1, 0.1]]))
     assert (convergence == np.array([0.0])).all()
 
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, 0.0))
+    shear = ag.mp.ExternalShear(gamma=(0.1, 0.0))
     convergence = shear.convergence_2d_from(
         grid=np.array([[0.1, 0.1], [0.2, 0.2], [0.3, 0.3]])
     )
@@ -76,11 +74,11 @@ def test__convergence_returns_zeros():
 
 
 def test__potential_returns_zeros():
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, 0.0))
+    shear = ag.mp.ExternalShear(gamma=(0.1, 0.0))
     potential = shear.potential_2d_from(grid=np.array([[0.1, 0.1]]))
     assert (potential == np.array([[0.0, 0.0]])).all()
 
-    shear = ag.mp.ExternalShear(ell_comps=(0.1, 0.0))
+    shear = ag.mp.ExternalShear(gamma=(0.1, 0.0))
     potential = shear.potential_2d_from(
         grid=np.array([[0.1, 0.1], [0.2, 0.2], [0.3, 0.3]])
     )
