+++ test_autogalaxy/config/grids/radial_minimum.ini
@@ -1,7 +1,7 @@
 [radial_minimum]
 PowerLawBroken=0.0001
 PowerLawBrokenSph=0.0001
-MockSphIsothermal=0.03
+MockIsothermalSph=0.03
 MockIsothermal=0.0001
 EllProfile=0.0001
 MockGridRadialMinimum=2.5
