+++ test_autogalaxy/plane/files/summary/notation/radial_minimum.ini
@@ -1,6 +1,6 @@
 [radial_minimum]
 MockIsothermal=0.01
-MockSphIsothermal=0.01
+MockIsothermalSph=0.01
 EllProfile=1e-8
 SersicSph=1e-8
 Sersic=1e-8
