+++ test_autogalaxy/config/lensing/grids/radial_minimum.ini
@@ -1,5 +1,5 @@
 [radial_minimum]
 MockIsothermal=0.01
-MockSphIsothermal=0.01
+MockIsothermalSph=0.01
 Isothermal=0.01
 IsothermalSph=0.01
