+++ test_autogalaxy/config/non_linear/nest/DynestyDynamic.ini
@@ -41,3 +41,4 @@ use_widths=True
 
 [parallel]
 number_of_cores=1
+force_x1_cpu=False
