### 2022-11-26 0cf2f5c
bypass figure of merit sanity output for test mode
