+++ test_autogalaxy/config/general.ini
@@ -40,7 +40,6 @@ hyper_noise_limit=1.0e8
 stochastic_outputs=False
 
 [test]
-output_figure_of_merit_sanity=False
 check_figure_of_merit_sanity=False
 check_preloads=False
 
