+++ autogalaxy/__init__.py
@@ -106,4 +106,4 @@ from autoconf import conf
 
 conf.instance.register(__file__)
 
-__version__ = "2022.07.11.1"
+__version__ = "2022.11.26.11"
