+++ docs/api/api.rst
@@ -78,7 +78,7 @@ Modeling
 Light Profiles
 --------------
 
-.. currentmodule:: autogalaxy.profiles.light_profiles
+.. currentmodule:: autogalaxy.profiles.light.standard
 
 .. autosummary::
    :toctree: generated/
@@ -261,7 +261,7 @@ They are used by the child project PyAutoLens, but currently do not have a use i
 Mass Profiles
 -------------
 
-.. currentmodule:: autogalaxy.profiles.mass_profiles
+.. currentmodule:: autogalaxy.profiles.mass
 
 **Total Mass Profiles:**
 
