+++ docs/general/configs.rst
@@ -2,8 +2,11 @@ Configs
 =======
 
 **PyAutoGalaxy** uses a number of configuration files that customize the default behaviour of the non-linear searches,
-visualization and other aspects of **PyAutoGalaxy**. Here, we describe how to configure **PyAutoGalaxy** to use the configs
-and describe every configuration file complete with input parameters.
+visualization and other aspects of **PyAutoGalaxy**.
+
+Descriptions of every configuration file and their input parameters are provided in the ``README.rst`` in
+the `config directory of the workspace <https://github.com/Jammy2211/autogalaxy_workspace/tree/release/config>`_
+
 
 Setup
 -----
@@ -22,271 +25,3 @@ command (the path to the ``output`` folder where the results of a non-linear sea
         config_path="path/to/config",
         output_path=f"path/to/output"
     )
-
-general.ini
------------
-
-This config file is found at 'autogalaxy_workspace/config/general.ini' and contains the following sections and variables:
-
-[output]
-    log_to_file
-        If True the outputs of processes like the non-linear search are logged to a file (and not printed to screen).
-    log_file
-        The file name the logged output is written to (in the non-linear search output folder).
-    log_level
-        The level of logging.
-    model_results_decimal_places
-        The number of decimal places the estimated values and errors of all parameters in the model.results file are
-        output to.
-    remove_files
-        If True, all output files of a non-linear search (e.g. samples, samples_backup, model.results, images, etc.)
-        are deleted once the model-fit has completed.
-
-        A .zip file of all output is always created before files are removed, thus results are not lost with this
-        option turned on. If PyAutoGalaxy does not find the output files of a model-fit (because they were removed) but
-        does find this .zip file, it will unzip the contents and continue the analysis as if the files were
-        there all along.
-
-        This feature was implemented because super-computers often have a limit on the number of files allowed per
-        user and the large number of files output by PyAutoGalaxy can exceed this limit. By removing files the
-        number of files is restricted only to the .zip files.
-    force_pickle_overwrite
-        A model-fit outputs pickled files of the model, search, results, etc., which the database feature can load.
-        If this setting it ``True`` these pickle files are recreated when a new model-fit is performed, even if
-        the search is complete.
-
-The following setting flips all images that are loaded by **PyAutoGalaxy** so that they appear the same orinetation as
-the software ds9:
-
-[fits]
-    flip_for_ds9
-        If ``True``, the ndarray of all .fits files containing an image, noise-map, psf, etc, is flipped upside down
-        so its orientation is the same as ds9.
-
-The following settings are specific for High Performance Super computer use with **PyAutoGalaxy**.
-
-[hpc]
-    hpc_mode
-        If ``True``, HPC mode is activated, which disables GUI visualization, logging to screen and other settings which
-        are not suited to running on a super computer.
-    iterations_per_update
-        The number of iterations between every update (visualization, results output, etc) in HPC mode, which may be
-        better suited to being less frequent on a super computer.
-
-The following settings customize how a model is handled by **PyAutoFit**:
-
-[model]
-    ignore_prior_limits
-        If ``True`` the limits applied to priors will be ignored, where limits set upper / lower limits. This should be
-        disabled if one has difficult manipulating results in the database due to a ``PriorLimitException``.
-
-The library `numba <https://github.com/numba/numba>`_ is used to speed up functions, by converting them to C callable
-functions before the Python interpreter runs:
-
-[numba]
-    nopython
-        If True, functions which hve a numba decorator must not convert back to Python during a run and thus must stay
-        100% C. All PyAutoGalaxy functions were developed with nopython mode turned on.
-    cache
[…]
