+++ MANIFEST.in
@@ -1,19 +1,22 @@
 # MANIFEST.in
 exclude .gitignore
-exclude .coverage
-exclude .travis.yml
-exclude docs
 include README.rst
 include setup.cfg
+include CITATIONS.rst
+include LICENSE
+include requirements.txt
+include optional_requirements.txt
+
 prune .cache
 prune .git
 prune build
 prune dist
-recursive-include autogalaxy/config *
+
 recursive-exclude *.egg-info *
-recursive-include tests *
-include requirements.txt
-recursive-include data *
+
+recursive-include autogalaxy/config *
+
+exclude docs
 
 global-exclude test_autogalaxy
 recursive-exclude test_autogalaxy *
