+++ autogalaxy/config/visualize.yaml
@@ -1,0 +1,165 @@
+general: {}
+include:
+  include_1d:
+    einstein_radius: true
+    half_light_radius: true
+  include_2d:
+    caustics: true
+    critical_curves: true
+    light_profile_centres: true
+    mass_profile_centres: true
+    multiple_images: true
+mat_wrap_1d:
+  EinsteinRadiusAXVLine:
+    figure:
+      c: b
+      linestyle: --
+    subplot:
+      c: b
+      linestyle: --
+  HalfLightRadiusAXVLine:
+    figure:
+      c: g
+      linestyle: -.
+    subplot:
+      c: kg
+      linestyle: -.
+  ModelFluxesYXScatter:
+    figure:
+      c: c
+    subplot:
+      c: c
+mat_wrap_2d:
+  CausticsPlot:
+    figure:
+      c: k,r,g,b,m,y
+      linestyle: '-'
+      linewidth: 2
+    subplot:
+      c: k,r,g,b,m,y
+      linestyle: '-'
+      linewidth: 2
+  CriticalCurvesPlot:
+    figure:
+      c: k,r,g,b,m,y
+      linestyle: '-'
+      linewidth: 2
+    subplot:
+      c: k,r,g,b,m,y
+      linestyle: '-'
+      linewidth: 2
+  LightProfileCentresScatter:
+    figure:
+      c: k,r,g,b,m,y
+      marker: +
+      s: 20
+    subplot:
+      c: r,g,b,m,y,k
+      marker: +
+      s: 26
+  MassProfileCentresScatter:
+    figure:
+      c: k,r,g,b,m,y
+      marker: x
+      s: 20
+    subplot:
+      c: r,g,b,m,y,k
+      marker: x
+      s: 26
+  MultipleImagesScatter:
+    figure:
+      c: k,r,g,b,m,y
+      marker: o
+      s: 16
+    subplot:
+      c: r,g,b,m,y,k
+      marker: o
+      s: 16
+plots:
+  dataset:
+    absolute_signal_to_noise_map: false
+    data: false
+    inverse_noise_map: false
+    noise_map: false
+    potential_chi_squared_map: false
+    signal_to_noise_map: false
+    subplot_dataset: true
+  fit:
+    all_at_end_fits: false
+    all_at_end_png: false
+    chi_squared_map: false
+    data: false
+    model_data: false
+    model_images_of_galaxies: true
+    noise_map: false
+    normalized_residual_map: false
+    residual_map: false
+    signal_to_noise_map: false
+    subplot_fit: true
+    subplot_of_galaxies: true
+    subtracted_images_of_galaxies: true
+  fit_imaging: {}
+  fit_interferometer:
+    amplitudes_vs_uv_distances: false
+    dirty_chi_squared_map: false
+    dirty_image: false
+    dirty_inverse_noise_map: false
+    dirty_noise_map: false
+    dirty_normalized_residual_map: false
+    dirty_residual_map: false
+    dirty_signal_to_noise_map: false
+    phases_vs_uv_distances: false
+    uv_wavelengths: false
+  fit_quantity:
+    all_at_end_fits: true
+    all_at_end_png: true
+    chi_squared_map: false
+    image: true
+    model_image: false
+    noise_map: false
+    residual_map: false
+    subplot_fit: true
+  galaxies:
+    convergence: false
+    image: false
+    potential: false
+  hyper:
+    contribution_map_list: true
+    fit_no_hyper: true
+    images_of_galaxies: true
+    model_image: true
+  imaging:
+    psf: false
+  interferometer:
+    amplitudes_vs_uv_distances: false
+    dirty_image: false
+    dirty_inverse_noise_map: false
[…]
