+++ autogalaxy/config/visualize/mat_wrap_2d/MassProfileCentresScatter.ini
@@ -1,9 +1,0 @@
-[figure]
-s=20
-marker=x
-c=k,r,g,b,m,y
-
-[subplot]
-s=26
-marker=x
-c=r,g,b,m,y,k
