+++ autogalaxy/config/visualize/mat_wrap_1d/ModelFluxesYXScatter.ini
@@ -1,5 +1,0 @@
-[figure]
-c=c
-
-[subplot]
-c=c
