+++ autogalaxy/config/grids/radial_minimum.ini
@@ -1,52 +1,0 @@
-[radial_minimum]
-PointMass=1e-8
-PowerLaw=1e-8
-PowerLawSph=1e-8
-Isothermal=1e-8
-IsothermalInitialize=1e-8
-IsothermalSph=1e-8
-PowerLawBroken=1e-8
-PowerLawBrokenSph=1e-8
-PowerLawCore=1e-8
-PowerLawCoreSph=1e-8
-IsothermalCore=1e-8
-IsothermalCoreSph=1e-8
-Moffat=1e-8
-MoffatSph=1e-8
-Gaussian=1e-8
-GaussianSph=1e-8
-Sersic=1e-6
-SersicSph=1e-6
-Exponential=1e-6
-ExponentialSph=1e-6
-DevVaucouleurs=1e-6
-DevVaucouleursSph=1e-6
-SersicCore=1e-6
-SersicCoreSph=1e-6
-ExponentialCore=1e-6
-ExponentialCoreSph=1e-6
-SphSersicRadialGradient=1e-6
-SersicRadialGradient=1e-6
-SphExponentialRadialGradient=1e-6
-EllExponentialRadialGradient=1e-6
-Chameleon=1e-8
-ChameleonSph=1e-8
-ElsonFreeFall=1e-8
-ElsonFreeFallSph=1e-8
-NFW=1e-6
-NFWSph=1e-6
-gNFW=1e-6
-gNFWSph=1e-6
-NFWTruncatedSph=1e-6
-NFWTruncatedMCRDuffySph=1e-6
-NFWTruncatedMCRLudlowSph=1e-6
-NFWTruncatedMCRScatterLudlowSph=1e-6
-NFWMCRDuffySph=1e-6
-NFWMCRLudlow=1e-6
-NFWMCRLudlowSph=1e-6
-NFWMCRScatterLudlowSph=1e-6
-NFWMCRScatterLudlow=1e-6
-gNFWMCRLudlow=1e-6
-ExternalShear=1e-8
-MassSheet=1e-8
-EllProfile=1e-8
