+++ autogalaxy/config/visualize/mat_wrap_2d/MultipleImagesScatter.ini
@@ -1,9 +1,0 @@
-[figure]
-s=16
-marker=o
-c=k,r,g,b,m,y
-
-[subplot]
-s=16
-marker=o
-c=r,g,b,m,y,k
