+++ autogalaxy/config/notation.yaml
@@ -1,0 +1,115 @@
+label:
+  label:
+    alpha: \alpha
+    beta: \beta
+    break_radius: \theta_{\rm B}
+    centre_0: y
+    centre_1: x
+    coefficient: \lambda
+    contribution_factor: \omega_{\rm 0}
+    core_radius: C_{\rm r}
+    core_radius_0: C_{rm r0}
+    core_radius_1: C_{\rm r1}
+    effective_radius: R_{\rm eff}
+    einstein_radius: \theta_{\rm Ein}
+    ell_comps_0: \epsilon_{\rm 1}
+    ell_comps_1: \epsilon_{\rm 2}
+    flux: F
+    gamma: \gamma
+    inner_coefficient: \lambda_{\rm 1}
+    inner_slope: t_{\rm 1}
+    intensity: I_{\rm b}
+    kappa: \kappa
+    kappa_s: \kappa_{\rm s}
+    mass_at_200: M_{\rm 200}
+    mass_to_light_gradient: \Gamma
+    mass_to_light_ratio: \Psi
+    noise_factor: \omega_{\rm 1}
+    noise_power: \omega{\rm 2}
+    noise_scale: \sigma_{\rm 1}
+    normalization_scale: n
+    outer_coefficient: \lambda_{\rm 2}
+    outer_slope: t_{\rm 2}
+    pixels: N_{\rm pix}
+    radius_break: R_{\rm b}
+    redshift: z
+    redshift_object: z_{\rm obj}
+    redshift_source: z_{\rm src}
+    scale_radius: R_{\rm s}
+    scatter: \sigma
+    sersic_index: n
+    shape_0: y_{\rm pix}
+    shape_1: x_{\rm pix}
+    sigma: \sigma
+    signal_scale: V
+    sky_scale: \sigma_{\rm 0}
+    slope: \gamma
+    truncation_radius: R_{\rm t}
+    weight_floor: W_{\rm f}
+    weight_power: W_{\rm p}
+  superscript:
+    externalshear: ext
+    hyperbackgroundnoise: hyper
+    hypergalaxy: hyper
+    hyperimagesky: hyper
+    inputdeflections: defl
+    pixelization: pix
+    point: point
+    redshift: ''
+    regularization: reg
+label_format:
+  format:
+    alpha: '{:.4f}'
+    angular_diameter_distance_to_earth: '{:.4f}'
+    beta: '{:.4f}'
+    centre_0: '{:.4f}'
+    centre_1: '{:.4f}'
+    coefficient: '{:.4f}'
+    concentration: '{:.4f}'
+    contribution_factor: '{:.4f}'
+    core_radius: '{:.4f}'
+    core_radius_0: '{:.4f}'
+    core_radius_1: '{:.4f}'
+    effective_radius: '{:.4f}'
+    einstein_mass: '{:.4e}'
+    einstein_radius: '{:.4f}'
+    ell_comps_0: '{:.4f}'
+    ell_comps_1: '{:.4f}'
+    flux: '{:.4e}'
+    gamma: '{:.4f}'
+    inner_coefficient: '{:.4f}'
+    inner_slope: '{:.4f}'
+    intensity: '{:.4f}'
+    kappa: '{:.4f}'
+    kappa_s: '{:.4f}'
+    kpc_per_arcsec: '{:.4f}'
+    luminosity: '{:.4e}'
+    mass: '{:.4e}'
+    mass_at_200: '{:.4e}'
+    mass_at_truncation_radius: '{:.4e}'
+    mass_to_light_gradient: '{:.4f}'
+    mass_to_light_ratio: '{:.4f}'
[…]
