+++ autogalaxy/config/visualize/mat_wrap_2d/LightProfileCentresScatter.ini
@@ -1,9 +1,0 @@
-[figure]
-s=20
-marker=+
-c=k,r,g,b,m,y
-
-[subplot]
-s=26
-marker=+
-c=r,g,b,m,y,k

[…]
