+++ test_autogalaxy/config/grids/radial_minimum.ini
@@ -1,40 +1,0 @@
-[radial_minimum]
-PowerLawBroken=0.0001
-PowerLawBrokenSph=0.0001
-MockIsothermalSph=0.03
-MockIsothermal=0.0001
-EllProfile=0.0001
-MockGridRadialMinimum=2.5
-IsothermalSph=0.0001
-SersicSph=0.0001
-Sersic=0.0001
-ExternalShear=0.0001
-Gaussian=0.0001
-Isothermal=0.0001
-Exponential=0.0001
-SersicCore=0.0001
-SersicCoreSph=0.0001
-DevVaucouleurs=0.0001
-DevVaucouleursSph=0.0001
-ExponentialSph=0.0001
-GaussianSph=0.0001
-MassSheet=0.0001
-SphSersicRadialGradient=0.0001
-SersicRadialGradient=0.0001
-SphExponentialRadialGradient=0.0001
-EllExponentialRadialGradient=0.0001
-Chameleon=0.0001
-ChameleonSph=0.0001
-gNFWSph=0.0001
-EllMassProfile=0.0001
-PowerLawCore=0.0001
-PowerLawCoreSph=0.0001
-PowerLaw=0.0001
-PowerLawSph=0.0001
-IsothermalCore=0.0001
-IsothermalCoreSph=0.0001
-gNFW=0.0001
-NFWTruncatedSph=0.0001
-NFW=0.0001
-NFWSph=0.0001
-PointMass=0.0001
