+++ test_autogalaxy/config/visualize/plots.ini
@@ -1,101 +1,0 @@
-[dataset]
-subplot_dataset=True
-data=True
-noise_map=False
-inverse_noise_map=True
-signal_to_noise_map=False
-absolute_signal_to_noise_map=False
-potential_chi_squared_map=True
-
-[imaging]
-psf=True
-
-[plane]
-all_at_end_png=False
-all_at_end_fits=True
-subplot_plane=True
-image=True
-source_plane_image=True
-convergence=True
-potential=False
-deflections=False
-magnification=True
-
-[fit]
-all_at_end_png=False
-all_at_end_fits=True
-subplot_fit=True
-subplot_of_galaxies_fits=False
-data=True
-noise_map=False
-signal_to_noise_map=False
-model_data=True
-residual_map=False
-normalized_residual_map=True
-chi_squared_map=True
-subtracted_images_of_galaxies=True
-model_images_of_galaxies=False
-
-[fit_imaging]
-
-[galaxies]
-image=True
-convergence=False
-potential=True
-
-[inversion]
-all_at_end_png=False
-subplot_inversion=True
-reconstructed_image=True
-reconstruction=True
-errors=False
-residual_map=False
-normalized_residual_map=False
-chi_squared_map=True
-regularization_weights=False
-interpolated_reconstruction=True
-interpolated_errors=True
-
-[interferometer]
-uv_wavelengths=False
-amplitudes_vs_uv_distances=False
-phases_vs_uv_distances=False
-dirty_image=False
-dirty_noise_map=False
-dirty_signal_to_noise_map=False
-dirty_inverse_noise_map=False
-
-[fit_interferometer]
-uv_wavelengths=False
-amplitudes_vs_uv_distances=False
-phases_vs_uv_distances=False
-dirty_image=False
-dirty_noise_map=False
-dirty_signal_to_noise_map=False
-dirty_inverse_noise_map=False
-dirty_residual_map=False
-dirty_normalized_residual_map=False
-dirty_chi_squared_map=False
-
-[hyper]
-fit_no_hyper=True
-model_image=True
-images_of_galaxies=True
-contribution_map_list=True
-
-[fit_quantity]
-all_at_end_png=True
-all_at_end_fits=True
-subplot_fit=True
-image=True
-noise_map=False
-model_image=True
-residual_map=False
-normalized_residual_map=True
-chi_squared_map=True
-
-[positions]
-image_with_positions=True
-
-[other]
-stochastic_histogram=False
