+++ test_autogalaxy/config/text/label_format.ini
@@ -1,10 +1,0 @@
-[format]
-param0={:.2f}
-param00={:.2f}
-param000={:.2f}
-param11={:.4f}
-param12={:.2e}
-radius_value=radius_value
-mass_value=mass_value
-radius={:.2f}
-mass={:.4e}

[…]
