+++ autogalaxy/config/README.rst
@@ -1,0 +1,16 @@
+The ``config`` folder contains configuration files which customize default **PyAutoLens**.
+
+Folders
+-------
+
+- ``priors``: Configs defining default priors assumed on every lens model component and set of parameters.
+- ``visualize``: Configs defining what images are output by a lens model fit.
+
+Files
+-----
+
+- ``general.yaml``: Customizes general **PyAutoLens** settings.
+- ``grids.yaml``: Customize default behaviour of grids when used for calculations.
+- ``logging.yaml``: Customizes the logging behaviour.
+- ``notation.yaml``: Configs defining labels and formatting of model parameters when used for visualization.
+
