+++ autogalaxy/config/visualize/README.rst
@@ -1,0 +1,11 @@
+The ``config`` folder contains configuration files which customize default **PyAutoLens**.
+
+Files
+-----
+
+- ``general.yaml``: Customizes general visualization settings (e.g. the matplotlib backend).
+- ``include.yaml``: Customize features that appears on plotted images by default (e.g. a mask, a grid).
+- ``plots.yaml``: Customize which figures are output during a model-fit.
+- ``mat_wrap.yaml``: Specify the default matplotlib settings when figures and subplots are plotted.
+- ``mat_wrap_1d.yaml``: Specify the default matplotlib settings when 1D figures and subplots are plotted.
+- ``mat_wrap_2d.yaml``: Specify the default matplotlib settings when 2D figures and subplots are plotted.
