### 2022-11-30 49a667e
restructing light nad mass prior configs
