### 2022-11-30 e299e8e
dark mass configs
