### 2022-11-30 dfbb5f0
stellar mass configs
