+++ autogalaxy/profiles/mass/stellar/__init__.py
@@ -4,4 +4,4 @@ from .exponential import Exponential, ExponentialSph
 from .gaussian import Gaussian
 from .sersic import Sersic, SersicSph
 from .sersic_core import SersicCore, SersicCoreSph
-from .sersic_radial_gradient import SersicRadialGradient, SphSersicRadialGradient
+from .sersic_radial_gradient import SersicRadialGradient, SersicRadialGradientSph
