+++ autogalaxy/config/priors/mass/stellar/dev_vaucouleurs.yaml
@@ -1,4 +1,4 @@
-gNFW:
+DevVaucouleurs:
   centre_0:
     gaussian_limits:
       lower: -inf
@@ -23,6 +23,16 @@ gNFW:
     width_modifier:
       type: Absolute
       value: 0.05
+  effective_radius:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 0.0
+    type: Uniform
+    upper_limit: 30.0
+    width_modifier:
+      type: Relative
+      value: 1.0
   ell_comps_0:
     gaussian_limits:
       lower: -1.0
@@ -47,37 +57,27 @@ gNFW:
     width_modifier:
       type: Absolute
       value: 0.2
-  inner_slope:
+  intensity:
     gaussian_limits:
-      lower: -1.0
-      upper: 3.0
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 2.0
+      lower: 0.0
+      upper: inf
+    lower_limit: 1.0e-06
+    type: LogUniform
+    upper_limit: 10.0
     width_modifier:
-      type: Absolute
+      type: Relative
+      value: 0.5
+  mass_to_light_ratio:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 1.0e-06
+    type: LogUniform
+    upper_limit: 1000000.0
+    width_modifier:
+      type: Relative
       value: 0.3
-  kappa_s:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 1.0
-    width_modifier:
-      type: Relative
-      value: 0.2
-  scale_radius:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 30.0
-    width_modifier:
-      type: Relative
-      value: 0.2
-gNFWSph:
+DevVaucouleursSph:
   centre_0:
     gaussian_limits:
       lower: -inf
@@ -102,27 +102,7 @@ gNFWSph:
     width_modifier:
       type: Absolute
       value: 0.05
-  inner_slope:
-    gaussian_limits:
-      lower: -1.0
-      upper: 3.0
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 2.0
-    width_modifier:
-      type: Absolute
-      value: 0.3
-  kappa_s:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 1.0
-    width_modifier:
-      type: Relative
-      value: 0.2
-  scale_radius:
+  effective_radius:
     gaussian_limits:
       lower: 0.0
       upper: inf
@@ -131,4 +111,24 @@ gNFWSph:
     upper_limit: 30.0
     width_modifier:
       type: Relative
-      value: 0.2
+      value: 1.0
+  intensity:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 1.0e-06
+    type: LogUniform
+    upper_limit: 10.0
+    width_modifier:
+      type: Relative
+      value: 0.5
+  mass_to_light_ratio:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 1.0e-06
+    type: LogUniform
+    upper_limit: 1000000.0
+    width_modifier:
+      type: Relative
+      value: 0.3

[…]
