+++ autogalaxy/config/priors/mesh/voronoi.yaml
@@ -1,0 +1,52 @@
+VoronoiBrightnessImage:
+  pixels:
+    gaussian_limits:
+      lower: 50.0
+      upper: inf
+    lower_limit: 50.0
+    type: Uniform
+    upper_limit: 2500.0
+    width_modifier:
+      type: Absolute
+      value: 100.0
+  weight_floor:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 0.0
+    type: Uniform
+    upper_limit: 1.0
+    width_modifier:
+      type: Absolute
+      value: 0.1
+  weight_power:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 0.0
+    type: Uniform
+    upper_limit: 20.0
+    width_modifier:
+      type: Absolute
+      value: 5.0
+VoronoiMagnification:
+  shape_0:
+    gaussian_limits:
+      lower: 3.0
+      upper: inf
+    lower_limit: 20.0
+    type: Uniform
+    upper_limit: 45.0
+    width_modifier:
+      type: Absolute
+      value: 8.0
+  shape_1:
+    gaussian_limits:
+      lower: 3.0
+      upper: inf
+    lower_limit: 20.0
+    type: Uniform
+    upper_limit: 45.0
+    width_modifier:
+      type: Absolute
+      value: 8.0
