+++ autogalaxy/config/priors/regularization/zeroth.yaml
@@ -1,0 +1,11 @@
+Zeroth:
+  coefficient:
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
+    lower_limit: 1.0e-06
+    type: LogUniform
+    upper_limit: 1000000.0
+    width_modifier:
+      type: Relative
+      value: 0.5
