+++ autogalaxy/config/priors/regularization.yaml
@@ -1,116 +1,0 @@
-adaptive_brightness.AdaptiveBrightness:
-  inner_coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-  outer_coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-  signal_scale:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 1.0
-    width_modifier:
-      type: Relative
-      value: 0.2
-adaptive_brightness_split.AdaptiveBrightnessSplit:
-  inner_coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-  outer_coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-  signal_scale:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 0.0
-    type: Uniform
-    upper_limit: 1.0
-    width_modifier:
-      type: Relative
-      value: 0.2
-constant.Constant:
-  coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-constant_split.ConstantSplit:
-  coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-constant_zeroth.ConstantZeroth:
-  coefficient_neighbor:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-  coefficient_zeroth:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
-zeroth.Zeroth:
-  coefficient:
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
-    lower_limit: 1.0e-06
-    type: LogUniform
-    upper_limit: 1000000.0
-    width_modifier:
-      type: Relative
-      value: 0.5
