+++ autogalaxy/config/priors/mesh/README.rst
@@ -1,0 +1,3 @@
+The ``mesh`` folder contains configuration files for the default priors assumed for ``mesh`` objects.
+
+These model components reconstruct a galaxy's emission using a rectangular grid, Delaunay triangulation or Voronoi mesh.
