+++ autogalaxy/config/priors/mass/README.rst
@@ -1,0 +1,9 @@
+The ``mass`` folder contains configuration files for the default priors assumed for mass profiles in **PyAutoGalaxy**.
+
+Folders
+-------
+
+- ``dark``: Configs for mass profiles that typically represent dark matter (e.g. ``NFW``, ``gNFW``).
+- ``stellar``: Configs for mass profiles that typically represent stellar matter (e.g. ``Sersic`, ``Gaussian``).
+- ``total``: Configs for mass profiles that typically represent the total (e.g. stellar and dark matter) mass distribution (e.g. ``Isothermal`, ``PowerLaw``).
+- ``sheets``: Configs for mass profiles that represent sheet-like mass components (e.g. ``ExternalShear`, ``MassSheet``).
