+++ test_autogalaxy/config/non_linear.yaml
@@ -140,64 +140,6 @@ nest:
       model_results_every_update: 1
       remove_state_files_at_end: true
       visualize_every_update: 1
-  MultiNest:
-    hyper_combined:
-      acceptance_ratio_threshold: 0.05
-      const_efficiency_mode: true
-      evidence_tolerance: 10.0
-      multimodal: false
-      n_live_points: 50
-      sampling_efficiency: 0.3
-      terminate_at_acceptance_ratio: false
-    hyper_galaxy:
-      const_efficiency_mode: false
-      evidence_tolerance: 0.1
-      multimodal: false
-      n_live_points: 30
-      sampling_efficiency: 0.5
-    initialize:
-      method: prior
-    inversion:
-      acceptance_ratio_threshold: 0.05
-      const_efficiency_mode: true
-      evidence_tolerance: 100.0
-      multimodal: false
-      n_live_points: 50
-      sampling_efficiency: 0.3
-      terminate_at_acceptance_ratio: false
-    printing:
-      silence: false
-    prior_passer:
-      sigma: 3.0
-      use_errors: true
-      use_widths: true
-    search:
-      const_efficiency_mode: false
-      evidence_tolerance: 0.5
-      importance_nested_sampling: false
-      max_iter: 0
-      max_modes: 100
-      mode_tolerance: -1.0e+90
-      multimodal: false
-      n_live_points: 50
-      sampling_efficiency: 0.5
-    settings:
-      context: 0
-      init_mpi: false
-      log_zero: -1.0e+100
-      n_iter_before_update: 5
-      null_log_evidence: -1.0e+90
-      resume: true
-      seed: -1.0
-      stagger_resampling_likelihood: true
-      verbose: false
-      write_output: true
-    updates:
-      iterations_per_update: 2500
-      log_every_update: 1
-      model_results_every_update: 1
-      remove_state_files_at_end: true
-      visualize_every_update: 1
 optimize:
   DownhillSimplex:
     search:
