### 2022-12-08 f234cb8
mass and light docstrings
