+++ autogalaxy/convert.py
@@ -9,9 +9,9 @@ def ell_comps_from(axis_ratio, angle):
     Parameters
     ----------
     axis_ratio
-            Ratio of light profiles ellipse's minor and major axes (b/a).
+        Ratio of light profiles ellipse's minor and major axes (b/a).
     angle
-            Rotation angle of light profile counter-clockwise from positive x-axis.
+        Rotation angle of light profile counter-clockwise from positive x-axis.
     """
     angle *= np.pi / 180.0
     fac = (1 - axis_ratio) / (1 + axis_ratio)
@@ -29,7 +29,7 @@ def axis_ratio_and_angle_from(ell_comps):
     ----------
     ell_comps : (float, float)
         The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
+        `autogalaxy -> convert.py` for the convention).
     """
     angle = np.arctan2(ell_comps[0], ell_comps[1]) / 2
     angle *= 180.0 / np.pi
@@ -50,7 +50,7 @@ def axis_ratio_from(ell_comps):
     ----------
     ell_comps : (float, float)
         The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
+        `autogalaxy -> convert.py` for the convention).
     """
     axis_ratio, angle = axis_ratio_and_angle_from(ell_comps=ell_comps)
     return axis_ratio
@@ -65,7 +65,7 @@ def angle_from(ell_comps):
     ----------
     ell_comps : (float, float)
         The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
+        `autogalaxy -> convert.py` for the convention).
     """
     axis_ratio, angle = axis_ratio_and_angle_from(ell_comps=ell_comps)
     return angle
