+++ autogalaxy/profiles/light_and_mass_profiles.py
@@ -3,15 +3,14 @@ from typing import Tuple
 from autogalaxy.profiles.light import standard as lp
 from autogalaxy.profiles import mass as mp
 
-"""
-Mass and light profiles describe both mass distributions and light distributions with a single set of parameters. This
-means that the light and mass of these profiles are tied together. Galaxy instances interpret these
-objects as being both mass and light profiles. 
-"""
 
 
 class LightMassProfile:
-
+    """
+    Mass and light profiles describe both mass distributions and light distributions with a single set of parameters. This
+    means that the light and mass of these profiles are tied together. Galaxy instances interpret these
+    objects as being both mass and light profiles.
+    """
     pass
 
 
@@ -24,7 +23,25 @@ class Gaussian(lp.Gaussian, mp.Gaussian, LightMassProfile):
         sigma: float = 1.0,
         mass_to_light_ratio: float = 1.0,
     ):
+        """
+        The elliptical Gaussian light and mass profile.
 
+        This simultaneously represents the luminous emission and stellar mass of a galaxy.
+
+        Parameters
+        ----------
+        centre
+            The grid of The (y,x) arc-second coordinates of the profile centre.
+        ell_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        intensity
+            Overall flux intensity normalisation in the light profiles (electrons per second).
+        effective_radius
+            The radius containing half the light of this light profile.
+        mass_to_light_ratio
+            The mass-to-light ratio of the light profiles.
+        """
         lp.Gaussian.__init__(
             self,
             centre=centre,
@@ -52,7 +69,25 @@ class Sersic(lp.Sersic, mp.Sersic, LightMassProfile):
         sersic_index: float = 0.6,
         mass_to_light_ratio: float = 1.0,
     ):
+        """
+        The elliptical Sersic light and mass profile.
 
+        This simultaneously represents the luminous emission and stellar mass of a galaxy.
+
+        Parameters
+        ----------
+        centre
+            The grid of The (y,x) arc-second coordinates of the profile centre.
+        ell_comps
+            The first and second ellipticity components of the elliptical coordinate system, (see the module
+            `autogalaxy -> convert.py` for the convention).
+        intensity
+            Overall flux intensity normalisation in the light profiles (electrons per second).
+        effective_radius
+            The radius containing half the light of this light profile.
+        mass_to_light_ratio
+            The mass-to-light ratio of the light profiles.
+        """
         lp.Sersic.__init__(
             self,
             centre=centre,
@@ -82,19 +117,20 @@ class SersicSph(Sersic, LightMassProfile):
         mass_to_light_ratio: float = 1.0,
     ):
         """
-        The SersicSph mass profile, the mass profiles of the light profiles that are used to fit_normal and
-        subtract the lens model_galaxy's light.
+        The spherical Sersic light and mass profile.
+
+        This simultaneously represents the luminous emission and stellar mass of a galaxy.
 
         Parameters
         ----------
         centre
-            The grid of the origin of the profiles
+            The grid of The (y,x) arc-second coordinates of the profile centre..
         intensity
-            Overall flux intensity normalisation in the light profiles (electrons per second)
+            Overall flux intensity normalisation in the light profiles (electrons per second).
         effective_radius
             The radius containing half the light of this light profile.
         mass_to_light_ratio
-            The mass-to-light ratio of the light profiles
+            The mass-to-light ratio of the light profiles.
         """
         Sersic.__init__(
             self,
@@ -117,17 +153,17 @@ class Exponential(Sersic, LightMassProfile):
         mass_to_light_ratio: float = 1.0,
     ):
[…]
