+++ docs/requirements.txt
@@ -1,4 +1,4 @@
-autoconf==2022.11.26.11
+autoconf
 astropy>=5.0
 colossus==1.3.1
 numba
