### 2022-12-08 ebb6e39
readthedocs should use class __init__
