+++ docs/conf.py
@@ -105,6 +105,8 @@ autosummary_generate = True
 autodoc_member_order = "bysource"
 autodoc_default_flags = ["members"]
 
+autoclass_content = 'both'
+
 numpydoc_show_class_members = False
 numpydoc_show_inherited_class_members = False
 numpydoc_class_members_toctree = False
