### 2022-12-08 9307fbd
docs and remove class method which computes from axis ratio and angle
