+++ autogalaxy/profiles/light/linear/sersic.py
@@ -17,8 +17,6 @@ class Sersic(lp.Sersic, LightProfileLinear):
         """
         The elliptical Sersic light profile.
 
-        See `autogalaxy.profiles.light.light_profiles.LightProfile` for a description of light profile objects.
-
         Parameters
         ----------
         centre
