+++ autogalaxy/profiles/light/linear/exponential_core.py
@@ -18,8 +18,6 @@ class ExponentialCore(lp.ExponentialCore, LightProfileLinear):
         """
         The elliptical cored-Exponential light profile.
 
-        See `autogalaxy.profiles.light.light_profiles.LightProfile` for a description of light profile objects.
-
         Parameters
         ----------
         centre

[…]
