+++ autogalaxy/profiles/mass/total/power_law.py
@@ -146,7 +146,7 @@ class PowerLawSph(PowerLaw):
     @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike):
 
-        eta = self.grid_to_grid_radii(grid)
+        eta = self.radial_grid_from(grid)
         deflection_r = (
             2.0
             * self.einstein_radius_rescaled
