+++ autogalaxy/profiles/mass/sheets/mass_sheet.py
@@ -37,5 +37,5 @@ class MassSheet(MassProfile):
     @aa.grid_dec.transform
     @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike):
-        grid_radii = self.grid_to_grid_radii(grid=grid)
+        grid_radii = self.radial_grid_from(grid=grid)
         return self.grid_to_grid_cartesian(grid=grid, radius=self.kappa * grid_radii)
