+++ autogalaxy/profiles/mass/dark/nfw.py
@@ -305,7 +305,7 @@ class NFWSph(NFW):
             The grid of (y,x) arc-second coordinates the deflection angles are computed on.
         """
 
-        eta = np.multiply(1.0 / self.scale_radius, self.grid_to_grid_radii(grid=grid))
+        eta = np.multiply(1.0 / self.scale_radius, self.radial_grid_from(grid=grid))
 
         deflection_grid = np.multiply(
             (4.0 * self.kappa_s * self.scale_radius / eta),
@@ -331,7 +331,7 @@ class NFWSph(NFW):
             The grid of (y,x) arc-second coordinates the deflection angles are computed on.
 
         """
-        eta = (1.0 / self.scale_radius) * self.grid_to_grid_radii(grid) + 0j
+        eta = (1.0 / self.scale_radius) * self.radial_grid_from(grid) + 0j
         return np.real(
             2.0 * self.scale_radius * self.kappa_s * self.potential_func_sph(eta)
         )
