+++ autogalaxy/profiles/mass/total/point.py
@@ -44,7 +44,7 @@ class PointMass(MassProfile):
     @aa.grid_dec.transform
     @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike):
-        grid_radii = self.grid_to_grid_radii(grid=grid)
+        grid_radii = self.radial_grid_from(grid=grid)
         return self.grid_to_grid_cartesian(
             grid=grid, radius=self.einstein_radius**2 / grid_radii
         )
