+++ autogalaxy/profiles/mass/dark/nfw_truncated.py
@@ -41,7 +41,7 @@ class NFWTruncatedSph(AbstractgNFW):
             The grid of (y,x) arc-second coordinates the deflection angles are computed on.
         """
 
-        eta = np.multiply(1.0 / self.scale_radius, self.grid_to_grid_radii(grid=grid))
+        eta = np.multiply(1.0 / self.scale_radius, self.radial_grid_from(grid=grid))
 
         deflection_grid = np.multiply(
             (4.0 * self.kappa_s * self.scale_radius / eta),
