+++ autogalaxy/profiles/mass/dark/gnfw.py
@@ -383,7 +383,7 @@ class gNFWSph(gNFW):
             The grid of (y,x) arc-second coordinates the deflection angles are computed on.
         """
 
-        eta = np.multiply(1.0 / self.scale_radius, self.grid_to_grid_radii(grid))
+        eta = np.multiply(1.0 / self.scale_radius, self.radial_grid_from(grid))
 
         deflection_grid = np.zeros(grid.shape[0])
 
