+++ autogalaxy/profiles/mass/sheets/mass_sheet.py
@@ -38,4 +38,6 @@ class MassSheet(MassProfile):
     @aa.grid_dec.relocate_to_radial_minimum
     def deflections_yx_2d_from(self, grid: aa.type.Grid2DLike):
         grid_radii = self.radial_grid_from(grid=grid)
-        return self._cartesian_grid_via_radial_from(grid=grid, radius=self.kappa * grid_radii)
+        return self._cartesian_grid_via_radial_from(
+            grid=grid, radius=self.kappa * grid_radii
+        )
