+++ test_autogalaxy/profiles/test_geometry_profiles.py
@@ -80,32 +80,34 @@ def test__spherical__transform_to_reference_frame__coordinate_shifts_using_centr
 
     spherical_profile = geometry_profiles.SphProfile(centre=(0.0, 0.0))
 
-    transformed_grid = spherical_profile.transform_grid_from_reference_frame(
-        np.array([[1.0, 1.0]])
+    transformed_grid = (
+        spherical_profile.grid_transformed_from_profile_reference_frame_from(
+            np.array([[1.0, 1.0]])
+        )
     )
 
     assert (transformed_grid == np.array([[1.0, 1.0]])).all()
 
     spherical_profile1 = geometry_profiles.SphProfile(centre=(0, 0))
 
-    grid1 = spherical_profile1.transform_grid_to_reference_frame(
+    grid1 = spherical_profile1.grid_transformed_to_profile_reference_frame_from(
         grid=np.array([[1.0, 1.0]])
     )
 
     spherical_profile2 = geometry_profiles.SphProfile(centre=(-1, -1))
-    grid2 = spherical_profile2.transform_grid_to_reference_frame(
+    grid2 = spherical_profile2.grid_transformed_to_profile_reference_frame_from(
         grid=np.array([[0.0, 0.0]])
     )
 
     assert (grid1 == grid2).all()
 
     spherical_profile1 = geometry_profiles.SphProfile(centre=(1, 1))
-    grid1 = spherical_profile1.transform_grid_to_reference_frame(
+    grid1 = spherical_profile1.grid_transformed_to_profile_reference_frame_from(
         grid=np.array([[1.0, 1.0]])
     )
 
     spherical_profile2 = geometry_profiles.SphProfile(centre=(-1, -1))
-    grid2 = spherical_profile2.transform_grid_to_reference_frame(
+    grid2 = spherical_profile2.grid_transformed_to_profile_reference_frame_from(
         grid=np.array([[-1.0, -1.0]])
     )
 
@@ -118,7 +120,9 @@ def test__spherical__transform_to_and_from_reference_frame():
 
     grid = np.array([[1.0, 1.0]])
 
-    transformed_grid = spherical_profile.transform_grid_to_reference_frame(grid)
+    transformed_grid = (
+        spherical_profile.grid_transformed_to_profile_reference_frame_from(grid)
+    )
 
     assert transformed_grid == pytest.approx(np.array([[1.0, 1.0]]), 1e-3)
 
@@ -126,10 +130,14 @@ def test__spherical__transform_to_and_from_reference_frame():
 
     grid_original = np.array([[5.2221, 2.6565]])
 
-    grid_spherical = spherical_profile.transform_grid_to_reference_frame(grid_original)
+    grid_spherical = spherical_profile.grid_transformed_to_profile_reference_frame_from(
+        grid_original
+    )
 
-    transformed_grid = spherical_profile.transform_grid_from_reference_frame(
-        grid_spherical
+    transformed_grid = (
+        spherical_profile.grid_transformed_from_profile_reference_frame_from(
+            grid_spherical
+        )
     )
 
     assert transformed_grid == pytest.approx(grid_original, 1e-5)
@@ -139,20 +147,26 @@ def test__elliptical__transform_grid_to_and_from_reference_frame():
 
     elliptical_profile = geometry_profiles.EllProfile(ell_comps=(0.0, 0.0))
 
-    transformed_grid = elliptical_profile.transform_grid_to_reference_frame(
-        grid=np.array([[1.0, 1.0]])
+    transformed_grid = (
+        elliptical_profile.grid_transformed_to_profile_reference_frame_from(
+            grid=np.array([[1.0, 1.0]])
+        )
     )
 
[…]
