### 2022-12-08 f1a7c76
Gaussian light profile docs
