+++ autogalaxy/profiles/mass/stellar/sersic_radial_gradient.py
@@ -137,7 +137,7 @@ class SersicRadialGradient(AbstractSersic):
             The grid of (y,x) arc-second coordinates the convergence is computed on.
 
         """
-        return self.convergence_func(self.grid_to_eccentric_radii(grid))
+        return self.convergence_func(self.eccentric_radii_grid_from(grid))
 
     def convergence_func(self, grid_radius: float) -> float:
         return (
