+++ autogalaxy/profiles/light/standard/eff.py
@@ -68,7 +68,8 @@ class ElsonFreeFall(LightProfile):
         """
         Returns the Eff light profile's 2D image from a 2D grid of Cartesian (y,x) coordinates.
 
-        If the coordinates have not been transformed to the profile's geometry, this is performed automatically.
+        If the coordinates have not been transformed to the profile's geometry (e.g. translated to the
+        profile `centre`), this is performed automatically.
 
         Parameters
         ----------
@@ -80,7 +81,7 @@ class ElsonFreeFall(LightProfile):
         image
             The image of the Eff evaluated at every (y,x) coordinate on the transformed grid.
         """
-        return self.image_2d_via_radii_from(self.grid_to_eccentric_radii(grid))
+        return self.image_2d_via_radii_from(self.eccentric_radii_grid_from(grid))
 
     @property
     def half_light_radius(self) -> float:
