+++ autogalaxy/profiles/mass/stellar/chameleon.py
@@ -135,7 +135,7 @@ class Chameleon(MassProfile, StellarProfile):
         grid
             The grid of (y,x) arc-second coordinates the convergence is computed on.
         """
-        return self.convergence_func(self.grid_to_elliptical_radii(grid))
+        return self.convergence_func(self.elliptical_radii_grid_from(grid))
 
     def convergence_func(self, grid_radius: float) -> float:
         return self.mass_to_light_ratio * self.image_2d_via_radii_from(grid_radius)
