+++ test_autogalaxy/profiles/mass/abstract/test_abstract.py
@@ -20,7 +20,7 @@ def mass_within_radius_of_profile_from_grid_calculation(radius, profile):
     for x in xs:
         for y in ys:
 
-            eta = profile.grid_to_elliptical_radii(grid=np.array([[x, y]]))
+            eta = profile.elliptical_radii_grid_from(grid=np.array([[x, y]]))
 
             if eta < radius:
                 mass_total += profile.convergence_func(eta) * area
