+++ test_autogalaxy/profiles/test_geometry_profiles.py
@@ -81,7 +81,7 @@ def test__spherical__transform_to_reference_frame__coordinate_shifts_using_centr
     spherical_profile = geometry_profiles.SphProfile(centre=(0.0, 0.0))
 
     transformed_grid = (
-        spherical_profile.grid_transformed_from_profile_reference_frame_from(
+        spherical_profile.transformed_from_reference_frame_grid_from(
             np.array([[1.0, 1.0]])
         )
     )
@@ -90,24 +90,24 @@ def test__spherical__transform_to_reference_frame__coordinate_shifts_using_centr
 
     spherical_profile1 = geometry_profiles.SphProfile(centre=(0, 0))
 
-    grid1 = spherical_profile1.grid_transformed_to_profile_reference_frame_from(
+    grid1 = spherical_profile1.transformed_to_reference_frame_grid_from(
         grid=np.array([[1.0, 1.0]])
     )
 
     spherical_profile2 = geometry_profiles.SphProfile(centre=(-1, -1))
-    grid2 = spherical_profile2.grid_transformed_to_profile_reference_frame_from(
+    grid2 = spherical_profile2.transformed_to_reference_frame_grid_from(
         grid=np.array([[0.0, 0.0]])
     )
 
     assert (grid1 == grid2).all()
 
     spherical_profile1 = geometry_profiles.SphProfile(centre=(1, 1))
-    grid1 = spherical_profile1.grid_transformed_to_profile_reference_frame_from(
+    grid1 = spherical_profile1.transformed_to_reference_frame_grid_from(
         grid=np.array([[1.0, 1.0]])
     )
 
     spherical_profile2 = geometry_profiles.SphProfile(centre=(-1, -1))
-    grid2 = spherical_profile2.grid_transformed_to_profile_reference_frame_from(
+    grid2 = spherical_profile2.transformed_to_reference_frame_grid_from(
         grid=np.array([[-1.0, -1.0]])
     )
 
@@ -121,7 +121,7 @@ def test__spherical__transform_to_and_from_reference_frame():
     grid = np.array([[1.0, 1.0]])
 
     transformed_grid = (
-        spherical_profile.grid_transformed_to_profile_reference_frame_from(grid)
+        spherical_profile.transformed_to_reference_frame_grid_from(grid)
     )
 
     assert transformed_grid == pytest.approx(np.array([[1.0, 1.0]]), 1e-3)
@@ -130,12 +130,12 @@ def test__spherical__transform_to_and_from_reference_frame():
 
     grid_original = np.array([[5.2221, 2.6565]])
 
-    grid_spherical = spherical_profile.grid_transformed_to_profile_reference_frame_from(
+    grid_spherical = spherical_profile.transformed_to_reference_frame_grid_from(
         grid_original
     )
 
     transformed_grid = (
-        spherical_profile.grid_transformed_from_profile_reference_frame_from(
+        spherical_profile.transformed_from_reference_frame_grid_from(
             grid_spherical
         )
     )
@@ -148,7 +148,7 @@ def test__elliptical__transform_grid_to_and_from_reference_frame():
     elliptical_profile = geometry_profiles.EllProfile(ell_comps=(0.0, 0.0))
 
     transformed_grid = (
-        elliptical_profile.grid_transformed_to_profile_reference_frame_from(
+        elliptical_profile.transformed_to_reference_frame_grid_from(
             grid=np.array([[1.0, 1.0]])
         )
     )
@@ -156,7 +156,7 @@ def test__elliptical__transform_grid_to_and_from_reference_frame():
     assert transformed_grid == pytest.approx(np.array([[1.0, 1.0]]), 1e-3)
 
     transformed_back_grid = (
[…]
