+++ autogalaxy/profiles/mass/total/power_law_broken.py
@@ -129,7 +129,7 @@ class PowerLawBroken(MassProfile):
             inner_part * (R <= self.break_radius) + outer_part * (R > self.break_radius)
         ).conjugate()
 
-        return self.rotate_grid_from_reference_frame(
+        return self.rotated_grid_from_reference_frame_from(
             grid=np.multiply(
                 1.0, np.vstack((np.imag(deflections), np.real(deflections))).T
             )
