+++ docs/conf.py
@@ -104,6 +104,7 @@ myst_heading_anchors = 3
 autosummary_generate = True
 autodoc_member_order = "bysource"
 autodoc_default_flags = ["members"]
+autodoc_class_signature = "separated"
 
 numpydoc_show_class_members = False
 numpydoc_show_inherited_class_members = False
