+++ docs/api/plot/search.rst
@@ -1,0 +1,15 @@
+Non-linear Search Plotters [aplt]
+---------------------------------
+
+.. currentmodule:: autogalaxy.plot
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   DynestyPlotter
+   UltraNestPlotter
+   EmceePlotter
+   ZeusPlotter
+   PySwarmsPlotter
