+++ docs/api/fitting/fitting.rst
@@ -1,5 +1,5 @@
-Imaging
-=======
+Fitting
+-------
 
 .. currentmodule:: autogalaxy
 
@@ -8,6 +8,5 @@ Imaging
    :template: custom-class-template.rst
    :recursive:
 
-   Imaging
-   SettingsImaging
-   SimulatorImaging
+   FitImaging
+   FitInterferometer
