+++ docs/api/data.rst
@@ -1,0 +1,10 @@
+===============
+Data Structures
+===============
+
+.. toctree::
+   :maxdepth: 1
+
+   data/structures
+   data/imaging
+   data/interferometer
