+++ docs/api/light.rst
@@ -1,0 +1,8 @@
+==============
+Light Profiles
+==============
+
+.. toctree::
+   :maxdepth: 1
+
+   light/standard
