+++ docs/api/data/structures.rst
@@ -1,24 +1,0 @@
-Data Structures
-===============
-
-.. currentmodule:: autogalaxy
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Mask1D
-   Mask2D
-   Array1D
-   Array2D
-   ValuesIrregular
-   Grid1D
-   Grid2D
-   Grid2DIterate
-   Grid2DIrregular
-   Kernel2D
-   Convolver
-   Visibilities
-   TransformerDFT
-   TransformerNUFFT
