+++ docs/api/data.rst
@@ -2,9 +2,54 @@
 Data Structures
 ===============
 
-.. toctree::
-   :maxdepth: 1
+Data Structures
+---------------
 
-   data/structures
-   data/imaging
-   data/interferometer
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Mask1D
+   Mask2D
+   Array1D
+   Array2D
+   ValuesIrregular
+   Grid1D
+   Grid2D
+   Grid2DIterate
+   Grid2DIrregular
+   Kernel2D
+   Convolver
+   Visibilities
+   TransformerDFT
+   TransformerNUFFT
+
+Imaging
+-------
+
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Imaging
+   SettingsImaging
+   SimulatorImaging
+
+
+Interferometer
+--------------
+
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+
+   SettingsInterferometer
+   Interferometer
+   SimulatorInterferometer
