+++ docs/api/light/standard.rst
@@ -1,26 +1,0 @@
-Standard [``ag.lp``]
---------------------
-
-.. currentmodule:: autogalaxy.profiles.light.standard
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Gaussian
-   GaussianSph
-   Sersic
-   SersicSph
-   Exponential
-   ExponentialSph
-   DevVaucouleurs
-   DevVaucouleursSph
-   SersicCore
-   SersicCoreSph
-   ExponentialCore
-   ExponentialCoreSph
-   Chameleon
-   ChameleonSph
-   ElsonFreeFall
-   ElsonFreeFallSph
