+++ docs/api/data/imaging.rst
@@ -1,13 +1,0 @@
-Imaging
-=======
-
-.. currentmodule:: autogalaxy
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Imaging
-   SettingsImaging
-   SimulatorImaging
