+++ docs/api/data/interferometer.rst
@@ -1,11 +1,0 @@
-Interferometer
-==============
-
-.. currentmodule:: autogalaxy
-
-.. autosummary::
-   :toctree: _autosummary
-
-   SettingsInterferometer
-   Interferometer
-   SimulatorInterferometer
