+++ docs/api/light.rst
@@ -2,7 +2,29 @@
 Light Profiles
 ==============
 
-.. toctree::
-   :maxdepth: 1
+Standard [``ag.lp``]
+--------------------
 
-   light/standard
+.. currentmodule:: autogalaxy.profiles.light.standard
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Gaussian
+   GaussianSph
+   Sersic
+   SersicSph
+   Exponential
+   ExponentialSph
+   DevVaucouleurs
+   DevVaucouleursSph
+   SersicCore
+   SersicCoreSph
+   ExponentialCore
+   ExponentialCoreSph
+   Chameleon
+   ChameleonSph
+   ElsonFreeFall
+   ElsonFreeFallSph
