+++ docs/api/inversion/mesh.rst
@@ -1,17 +1,0 @@
-Mesh [ag.mesh]
---------------
-
-.. currentmodule:: autoarray.inversion.pixelization.mesh
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Rectangular
-   DelaunayMagnification
-   DelaunayBrightnessImage
-   VoronoiMagnification
-   VoronoiBrightnessImage
-   VoronoiNNMagnification
-   VoronoiNNBrightnessImage
