+++ docs/api/inversion/settings.rst
@@ -1,12 +1,0 @@
-Settings
---------
-
-.. currentmodule:: autogalaxy
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   SettingsPixelization
-   SettingsInversion
