+++ docs/api/inversion/regularization.rst
@@ -1,14 +1,0 @@
-Regularization [ag.reg]
------------------------
-
-.. currentmodule:: autoarray.inversion.regularization
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Constant
-   ConstantSplit
-   AdaptiveBrightness
-   AdaptiveBrightnessSplit
