+++ docs/api/mass/stellar.rst
@@ -1,21 +1,0 @@
-Stellar [ag.mp]
----------------
-
-.. currentmodule:: autogalaxy.profiles.mass
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Gaussian
-   Sersic
-   SersicSph
-   Exponential
-   ExponentialSph
-   DevVaucouleurs
-   DevVaucouleursSph
-   SersicRadialGradient
-   SersicRadialGradientSph
-   Chameleon
-   ChameleonSph
