+++ docs/api/source/geometry.rst
@@ -1,12 +1,0 @@
-Geometry Profiles
------------------
-
-.. currentmodule:: autogalaxy.profiles.geometry_profiles
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   EllProfile
-   SphProfile
