+++ docs/api/pixelization.rst
@@ -1,0 +1,73 @@
+=============
+Pixelizations
+=============
+
+Mesh [ag.mesh]
+--------------
+
+.. currentmodule:: autoarray.inversion.pixelization.mesh
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Rectangular
+   DelaunayMagnification
+   DelaunayBrightnessImage
+   VoronoiMagnification
+   VoronoiBrightnessImage
+   VoronoiNNMagnification
+   VoronoiNNBrightnessImage
+
+Regularization [ag.reg]
+-----------------------
+
+.. currentmodule:: autoarray.inversion.regularization
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Constant
+   ConstantSplit
+   AdaptiveBrightness
+   AdaptiveBrightnessSplit
+
+Pixelization
+------------
+
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Pixelization
+
+Settings
+--------
+
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   SettingsPixelization
+   SettingsInversion
+
+Mapper
+------
+
+.. currentmodule:: autogalaxy
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Mapper
