+++ docs/api/mass/total.rst
@@ -1,21 +1,0 @@
-Total [ag.mp]
--------------
-
-.. currentmodule:: autogalaxy.profiles.mass
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   PointMass
-   PowerLawCore
-   PowerLawCoreSph
-   PowerLawBroken
-   PowerLawBrokenSph
-   IsothermalCore
-   IsothermalCoreSph
-   PowerLaw
-   PowerLawSph
-   Isothermal
-   IsothermalSph
