+++ docs/api/galaxy.rst
@@ -1,3 +1,7 @@
+==============
+Galaxy / Plane
+==============
+
 Galaxy / Plane
 --------------
 
