+++ docs/api/modeling/analysis.rst
@@ -1,12 +1,0 @@
-Analysis
-========
-
-.. currentmodule:: autogalaxy
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   AnalysisImaging
-   AnalysisInterferometer
