+++ docs/api/plot/mat_wrap.rst
@@ -1,58 +1,0 @@
-Matplot Lib Wrappers [aplt]
----------------------------
-
-.. currentmodule:: autogalaxy.plot
-
-**Matplotlib Wrapper Base Objects:**
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   Units
-   Figure
-   Axis
-   Cmap
-   Colorbar
-   ColorbarTickParams
-   TickParams
-   YTicks
-   XTicks
-   Title
-   YLabel
-   XLabel
-   Legend
-   Output
-
-**Matplotlib Wrapper 1D Objects:**
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   YXPlot
-
-**Matplotlib Wrapper 2D Objects:**
-
-.. autosummary::
-   :toctree: _autosummary
-   :template: custom-class-template.rst
-   :recursive:
-
-   ArrayOverlay
-   GridScatter
-   GridPlot
-   VectorYXQuiver
-   PatchOverlay
-   VoronoiDrawer
-   OriginScatter
-   MaskScatter
-   BorderScatter
-   PositionsScatter
-   IndexScatter
-   MeshGridScatter
-   ParallelOverscanPlot
-   SerialPrescanPlot
-   SerialOverscanPlot
