### 2022-12-15 3edfa20
no visualizaiton in test mode
