+++ autogalaxy/imaging/model/analysis.py
@@ -1,4 +1,5 @@
 import numpy as np
+import os
 from typing import Dict, Optional
 
 import autofit as af
@@ -94,14 +95,16 @@ class AnalysisImaging(AnalysisDataset):
 
         if not paths.is_complete:
 
-            visualizer = VisualizerImaging(visualize_path=paths.image_path)
+            if not os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
 
-            visualizer.visualize_imaging(imaging=self.imaging)
+                visualizer = VisualizerImaging(visualize_path=paths.image_path)
 
-            visualizer.visualize_hyper_images(
-                hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
-                hyper_model_image=self.hyper_model_image,
-            )
+                visualizer.visualize_imaging(imaging=self.imaging)
+
+                visualizer.visualize_hyper_images(
+                    hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
+                    hyper_model_image=self.hyper_model_image,
+                )
 
             self.set_preloads(paths=paths, model=model)
 
@@ -301,6 +304,9 @@ class AnalysisImaging(AnalysisDataset):
             If True the visualization is being performed midway through the non-linear search before it is finished,
             which may change which images are output.
         """
+
+        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
+            return
 
         instance = self.instance_with_associated_hyper_images_from(instance=instance)
         plane = self.plane_via_instance_from(instance=instance)
