+++ autogalaxy/interferometer/model/analysis.py
@@ -1,5 +1,6 @@
 import logging
 import numpy as np
+import os
 from typing import Optional
 
 import autofit as af
@@ -172,14 +173,16 @@ class AnalysisInterferometer(AnalysisDataset):
 
         if not paths.is_complete:
 
-            visualizer = VisualizerInterferometer(visualize_path=paths.image_path)
+            if not os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
 
-            visualizer.visualize_interferometer(interferometer=self.interferometer)
+                visualizer = VisualizerInterferometer(visualize_path=paths.image_path)
 
-            visualizer.visualize_hyper_images(
-                hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
-                hyper_model_image=self.hyper_model_image,
-            )
+                visualizer.visualize_interferometer(interferometer=self.interferometer)
+
+                visualizer.visualize_hyper_images(
+                    hyper_galaxy_image_path_dict=self.hyper_galaxy_image_path_dict,
+                    hyper_model_image=self.hyper_model_image,
+                )
 
             logger.info(
                 "PRELOADS - Setting up preloads, may take a few minutes for fits using an inversion."
@@ -363,6 +366,10 @@ class AnalysisInterferometer(AnalysisDataset):
             If True the visualization is being performed midway through the non-linear search before it is finished,
             which may change which images are output.
         """
+
+        if os.environ.get("PYAUTOFIT_TEST_MODE") == "1":
+            return
+
         instance = self.instance_with_associated_hyper_images_from(instance=instance)
         plane = self.plane_via_instance_from(instance=instance)
         hyper_background_noise = self.hyper_background_noise_via_instance_from(
