### 2022-12-16 5a39b0f
copy and pasted multipole code over
