### 2022-12-16 f30e1e1
multipole power law m4 added
