+++ autogalaxy/config/notation.yaml
@@ -15,6 +15,8 @@ label:
     einstein_radius: \theta_{\rm Ein}
     ell_comps_0: \epsilon_{\rm 1}
     ell_comps_1: \epsilon_{\rm 2}
+    ell_comps_multipole_0: M_{\rm 1}
+    ell_comps_multipole_1: M_{\rm 2}
     flux: F
     gamma: \gamma
     inner_coefficient: \lambda_{\rm 1}
@@ -76,6 +78,8 @@ label_format:
     einstein_radius: '{:.4f}'
     ell_comps_0: '{:.4f}'
     ell_comps_1: '{:.4f}'
+    ell_comps_multipole_0: '{:.4f}'
+    ell_comps_multipole_1: '{:.4f}'
     flux: '{:.4e}'
     gamma: '{:.4f}'
     inner_coefficient: '{:.4f}'
