### 2022-12-16 2fb5efe
refactor multipole
