+++ autogalaxy/profiles/light/standard/exponential.py
@@ -21,8 +21,7 @@ class Exponential(Sersic):
         centre
             The (y,x) arc-second centre of the light profile.
         ell_comps
-            The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
+            The first and second ellipticity components of the elliptical coordinate system.
         intensity
             Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
             the light profile's image is compared too, which is expected to be electrons per second).
