+++ autogalaxy/profiles/light/snr/dev_vaucouleurs.py
@@ -22,8 +22,7 @@ class DevVaucouleurs(lp.DevVaucouleurs, LightProfileSNR):
         centre
             The (y,x) arc-second coordinates of the profile centre.
         ell_comps
-            The first and second ellipticity components of the elliptical coordinate system, (see the module
-            `autogalaxy -> convert.py` for the convention).
+            The first and second ellipticity components of the elliptical coordinate system.
         intensity
             Overall intensity normalisation of the light profile (units are dimensionless and derived from the data
             the light profile's image is compared too, which is expected to be electrons per second).
