+++ autogalaxy/profiles/mass/multipole/power_law_m4.py
@@ -98,6 +98,8 @@ class MultipolePowerLawM4(MassProfile):
                 \phi^{\rm mass}_m = \arctan{\frac{\epsilon_{\rm 2}^{\rm mp}}{\epsilon_{\rm 2}^{\rm mp}}}, \, \,
                 k^{\rm mass}_m = \sqrt{{\epsilon_{\rm 1}^{\rm mp}}^2 + {\epsilon_{\rm 2}^{\rm mp}}^2} \, .
 
+        This mass profile is described fully in the following paper: https://arxiv.org/abs/1302.5482
+
         Parameters
         ----------
         centre
