+++ docs/api/source.rst
@@ -32,3 +32,97 @@ Operators
 
    OperateImage
    OperateDeflections
+
+Total [ag.mp]
+-------------
+
+.. currentmodule:: autogalaxy.profiles.mass
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   PointMass
+   PowerLawCore
+   PowerLawCoreSph
+   PowerLawBroken
+   PowerLawBrokenSph
+   IsothermalCore
+   IsothermalCoreSph
+   PowerLaw
+   PowerLawSph
+   Isothermal
+   IsothermalSph
+
+Mass Sheets [ag.mp]
+-------------------
+
+.. currentmodule:: autogalaxy.profiles.mass
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   ExternalShear
+   MassSheet
+
+Multipoles [ag.mp]
+------------------
+
+.. currentmodule:: autogalaxy.profiles.mass
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   MultipolePowerLawM4
+
+Stellar [ag.mp]
+---------------
+
+.. currentmodule:: autogalaxy.profiles.mass
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Gaussian
+   Sersic
+   SersicSph
+   Exponential
+   ExponentialSph
+   DevVaucouleurs
+   DevVaucouleursSph
+   SersicRadialGradient
+   SersicRadialGradientSph
+   Chameleon
+   ChameleonSph
+
+Dark [ag.mp]
+------------
+
+.. currentmodule:: autogalaxy.profiles.mass
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   gNFW
+   gNFWSph
+   NFWTruncatedSph
+   NFWTruncatedMCRDuffySph
+   NFWTruncatedMCRLudlowSph
+   NFWTruncatedMCRScatterLudlowSph
+   NFW
+   NFWSph
+   NFWMCRDuffySph
+   NFWMCRLudlowSph
+   NFWMCRScatterLudlow
+   NFWMCRScatterLudlowSph
+   NFWMCRLudlow
+   gNFWMCRLudlow
