+++ autogalaxy/config/priors/README.rst
@@ -26,8 +26,7 @@ The sections of this example config set the following:
         set in an analogous fashion (see the example configs).
     width_modifier
         When the results of a search are passed to a subsequent search to set up the priors of its non-linear search,
-        this entry describes how the Prior is passed. For a full description of prior passing, checkout the examples
-        in 'autolens_workspace/examples/complex/linking'.
+        this entry describes how the Prior is passed.
     gaussian_limits
         When the results of a search are passed to a subsequent search, they are passed using a GaussianPrior. The
         gaussian_limits set the physical lower and upper limits of this GaussianPrior, such that parameter samples
