### 2022-12-16 4bf8a9c
raw string literal multipole docstring
