+++ docs/overview/overview_3_modeling.rst
@@ -9,7 +9,7 @@ We can use a ``Plane`` to fit data of a galaxy and quantify its goodness-of-fit 
 Of course, when observe an image of a galaxy, we have no idea what combination of
 ``LightProfile``'s will produce a model-image that looks like the galaxy we observed:
 
-.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/master/docs/overview/images/fitting/image.png
+.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/main/docs/overview/images/fitting/image.png
   :width: 400
   :alt: Alternative text
 
@@ -282,7 +282,7 @@ make a corner plot of the probability density function (PDF):
 
 Here is an example of how a PDF estimated for a model appears:
 
-.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/master/docs/overview/images/modeling/cornerplot.png
+.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/main/docs/overview/images/modeling/cornerplot.png
   :width: 600
   :alt: Alternative text
 
@@ -300,7 +300,7 @@ plotted.
 Here's what the model-fit of the model which maximizes the log likelihood looks like, providing good residuals and
 low chi-squared values:
 
-.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/master/docs/overview/images/modeling/subplot_fit.png
+.. image:: https://raw.githubusercontent.com/Jammy2211/PyAutoGalaxy/main/docs/overview/images/modeling/subplot_fit.png
   :width: 600
   :alt: Alternative text
 
