+++ setup.py
@@ -48,6 +48,7 @@ setup(
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
     ],
+    python_requires='>=3.7',
     keywords="cli",
     packages=find_packages(exclude=["docs", "test_autogalaxy", "test_autogalaxy*"])
     + config_packages("autogalaxy/config"),
