+++ readthedocs.yaml
@@ -11,3 +11,4 @@ python:
 
 sphinx:
   configuration: docs/conf.py
+  fail_on_warning: false
