+++ .github/workflows/main.yml
@@ -68,7 +68,7 @@ jobs:
         pip3 install --upgrade pip
         pip3 install setuptools
         pip3 install wheel
-        pip3 install pytest
+        pip3 install pytest coverage pytest-cov
         pip3 install -r PyAutoConf/requirements.txt
         pip3 install -r PyAutoFit/requirements.txt
         pip3 install -r PyAutoArray/requirements.txt
@@ -82,7 +82,7 @@ jobs:
         export PYTHONPATH=$PYTHONPATH:$ROOT_DIR/PyAutoArray
         export PYTHONPATH=$PYTHONPATH:$ROOT_DIR/PyAutoGalaxy
         pushd PyAutoGalaxy
-        python3 -m pytest
+        python3 -m pytest --cov autogalaxy --cov-report xml:coverage.xml
     - name: Slack send
       if: ${{ failure() }}
       id: slack
