### 2022-12-20 01d3f60
plane.to_inversion
