### 2022-12-20 6edbe43
agreegator fit interferometer
