### 2022-12-20 d84ae82
more aggregator typing
