+++ autogalaxy/aggregator/fit_interferometer.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Optional, List
 
 if TYPE_CHECKING:
+    from autogalaxy.galaxy.galaxy import Galaxy
     from autogalaxy.interferometer.fit_interferometer import FitInterferometer
 
 import autofit as af
@@ -15,7 +16,7 @@ from autogalaxy.aggregator.plane import _plane_from
 
 def _fit_interferometer_from(
     fit: af.Fit,
-    galaxies: List["Galaxy"],
+    galaxies: List[Galaxy],
     real_space_mask: Optional[aa.Mask2D] = None,
     settings_interferometer: aa.SettingsInterferometer = None,
     settings_pixelization: aa.SettingsPixelization = None,
