### 2022-12-20 506e9d7
abstract_fit
