+++ autogalaxy/aggregator/abstract.py
@@ -1,5 +1,4 @@
 from __future__ import annotations
-
 from abc import ABC, abstractmethod
 from functools import partial
 from typing import TYPE_CHECKING, List, Generator
