### 2022-12-20 f35ad9a
move PixelixaitonGalaxy to test_galaxy
