### 2022-12-20 8610b19
Redshift refactored and docs added
