+++ autogalaxy/galaxy/hyper.py
@@ -1,0 +1,135 @@
+from itertools import count
+from typing import Dict, List, Optional, Type, Union
+
+import numpy as np
+
+import autoarray as aa
+
+
+class HyperGalaxy:
+    _ids = count()
+
+    def __init__(
+            self,
+            contribution_factor: float = 0.0,
+            noise_factor: float = 0.0,
+            noise_power: float = 1.0,
+    ):
+        """
+        By using a ``HyperGalaxy``, the noise-map value in the regions of the image that the galaxy is located
+        are increased.
+
+        This prevents over-fitting regions of the data where the model does not provide a good fit
+        (e.g. where a high chi-squared is inferred).
+
+        This uses the parent ``Galaxy``'s `'contribution_map`', which determines the fraction of flux in every pixel
+        of the image that is associated with that galaxy.
+
+        The ``HyperGalaxy`` class contains free parameters which using the ``contribution_map`` then increase
+        the noise.
+
+        Using ``HyperGalaxy``'s to perform noise-scaling is described fully in the following ``HowToGalaxy``
+        and ``HowToLens`` chapters:
+
+        https://pyautogalaxy.readthedocs.io/en/latest/howtogalaxy/chapter_optional.html
+        https://pyautolens.readthedocs.io/en/latest/howtolens/chapter_5_hyper_mode.html
+
+        Parameters
+        ----------
+        contribution_factor
+            Factor that adjusts how much of the galaxy's light is attributed to the
+            contribution map.
+        noise_factor
+            Factor by which the noise-map is increased in the regions of the galaxy's
+            contribution map.
+        noise_power
+            The power to which the contribution map is raised when scaling the
+            noise-map.
+        """
+        self.contribution_factor = contribution_factor
+        self.noise_factor = noise_factor
+        self.noise_power = noise_power
+
+        self.component_number = next(self._ids)
+
+    def __eq__(self, other):
+        if isinstance(other, HyperGalaxy):
+            return (
+                    self.contribution_factor == other.contribution_factor
+                    and self.noise_factor == other.noise_factor
+                    and self.noise_power == other.noise_power
+            )
+        return False
+
+    def __str__(self):
+        return "\n".join(["{}: {}".format(k, v) for k, v in self.__dict__.items()])
+
+    def contribution_map_from(self, hyper_model_image : aa.Array2D, hyper_galaxy_image : aa.Array2D) -> aa.Array2D:
+        """
+        Returns the contribution map of a galaxy, which represents the fraction of
+        flux in each pixel that the galaxy is attributed to contain. This uses the
+        `contribution_factor` free parameter.
+
+        This is computed by dividing that galaxy's flux by the total flux in that
+        pixel and then scaling by the maximum flux such that the contribution map
+        ranges between 0 and 1.
+
+        The contribution map is described in full in the following ``HowToGalaxy``
+        and ``HowToLens`` chapters:
+
+        https://pyautogalaxy.readthedocs.io/en/latest/howtogalaxy/chapter_optional.html
+        https://pyautolens.readthedocs.io/en/latest/howtolens/chapter_5_hyper_mode.html
+
+        Parameters
+        ----------
+        hyper_model_image
+            The best-fit model image to the observed image from a previous analysis
+            search. This provides the total light attributed to each image pixel by the
+            model.
+        hyper_galaxy_image
+            A model image of the galaxy (from light profiles or an inversion) from a
+            previous analysis search.
+        """
+        try:
+            contribution_map = np.divide(
+                hyper_galaxy_image, np.add(hyper_model_image, self.contribution_factor)
+            )
+            return np.divide(contribution_map, np.max(contribution_map))
[…]
