### 2022-12-20 a2b59d2
remove HyperGalaxy from galaxy.py
