+++ autogalaxy/galaxy/galaxy.py
@@ -565,118 +565,5 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
             )
 
 
-class HyperGalaxy:
-    _ids = count()
-
-    def __init__(
-        self,
-        contribution_factor: float = 0.0,
-        noise_factor: float = 0.0,
-        noise_power: float = 1.0,
-    ):
-        """
-        If a `Galaxy` is given a *HyperGalaxy* as an attribute, the noise-map in \
-        the regions of the image that the galaxy is located will be hyper, to prevent \
-        over-fitting of the galaxy.
-        
-        This is performed by first computing the hyper_galaxies-galaxy's 'contribution-map', \
-        which determines the fraction of flux in every pixel of the image that can be \
-        associated with this particular hyper_galaxies-galaxy. This is computed using \
-        hyper_galaxies-hyper_galaxies set (e.g. fitting.fit_data.FitDataHyper), which includes  best-fit \
-        unblurred_image_1d of the galaxy's light from a previous analysis search.
-         
-        The *HyperGalaxy* class contains the hyper_galaxies-parameters which are associated \
-        with this galaxy for scaling the noise-map.
-        
-        Parameters
-        ----------
-        contribution_factor
-            Factor that adjusts how much of the galaxy's light is attributed to the
-            contribution map.
-        noise_factor
-            Factor by which the noise-map is increased in the regions of the galaxy's
-            contribution map.
-        noise_power
-            The power to which the contribution map is raised when scaling the
-            noise-map.
-        """
-        self.contribution_factor = contribution_factor
-        self.noise_factor = noise_factor
-        self.noise_power = noise_power
-
-        self.component_number = next(self._ids)
-
-    def __eq__(self, other):
-        if isinstance(other, HyperGalaxy):
-            return (
-                self.contribution_factor == other.contribution_factor
-                and self.noise_factor == other.noise_factor
-                and self.noise_power == other.noise_power
-            )
-        return False
-
-    def __str__(self):
-        return "\n".join(["{}: {}".format(k, v) for k, v in self.__dict__.items()])
-
-    def contribution_map_from(self, hyper_model_image, hyper_galaxy_image):
-        """
-        Returns the contribution map of a galaxy, which represents the fraction of
-        flux in each pixel that the galaxy is attributed to contain, hyper to the
-        *contribution_factor* hyper_galaxies-parameter.
-
-        This is computed by dividing that galaxy's flux by the total flux in that \
-        pixel and then scaling by the maximum flux such that the contribution map \
-        ranges between 0 and 1.
-
-        Parameters
-        ----------
-        hyper_model_image
-            The best-fit model image to the observed image from a previous analysis
-            search. This provides the total light attributed to each image pixel by the
-            model.
-        hyper_galaxy_image
-            A model image of the galaxy (from light profiles or an inversion) from a
-            previous analysis search.
-        """
-        try:
-            contribution_map = np.divide(
-                hyper_galaxy_image, np.add(hyper_model_image, self.contribution_factor)
-            )
-            return np.divide(contribution_map, np.max(contribution_map))
-        except TypeError:
-            raise
-
-    def hyper_noise_map_via_hyper_images_from(
-        self,
-        hyper_model_image: aa.Array2D,
-        hyper_galaxy_image: aa.Array2D,
-        noise_map: aa.Array2D,
-    ) -> aa.Array2D:
-        contribution_map = self.contribution_map_from(
-            hyper_model_image=hyper_model_image, hyper_galaxy_image=hyper_galaxy_image
-        )
-
-        return self.hyper_noise_map_from(
[…]
