+++ autogalaxy/__init__.py
@@ -70,7 +70,9 @@ from .interferometer.model.analysis import AnalysisInterferometer
 from .quantity.fit_quantity import FitQuantity
 from .quantity.model.analysis import AnalysisQuantity
 from .quantity.dataset_quantity import DatasetQuantity
-from .galaxy.galaxy import Galaxy, HyperGalaxy, Redshift
+from .galaxy.galaxy import Galaxy
+from .galaxy.hyper import HyperGalaxy
+from .galaxy.redshift import Redshift
 from .galaxy.stellar_dark_decomp import StellarDarkDecomp
 from .hyper import hyper_data
 from .analysis.setup import SetupHyper
