### 2022-12-20 dcc87a0
separate hyper galaxy unit tests
