### 2022-12-20 8f8cab5
flesh out data.rst docs
