+++ docs/api/data.rst
@@ -2,8 +2,11 @@
 Data Structures
 ===============
 
-Data Structures
----------------
+2D Data Structures
+------------------
+
+Two-dimensional data structures store and mask 2D arrays containing data (e.g. images) and
+grids of (y,x) Cartesian coordinates (which are used for evaluating light profiles).
 
 .. currentmodule:: autogalaxy
 
@@ -12,25 +15,19 @@ Data Structures
    :template: custom-class-template.rst
    :recursive:
 
-   Mask1D
    Mask2D
-   Array1D
    Array2D
-   ValuesIrregular
-   Grid1D
    Grid2D
    Grid2DIterate
    Grid2DIrregular
-   Kernel2D
-   Convolver
-   Visibilities
-   TransformerDFT
-   TransformerNUFFT
+
+
 
 Imaging
 -------
 
-.. currentmodule:: autogalaxy
+For datasets taken with a CCD (or similar imaging device), including objects which perform
+2D convolution.
 
 .. autosummary::
    :toctree: _autosummary
@@ -40,12 +37,15 @@ Imaging
    Imaging
    SettingsImaging
    SimulatorImaging
+   Kernel2D
+   Convolver
 
 
 Interferometer
 --------------
 
-.. currentmodule:: autogalaxy
+For datasets taken with an interferometer (E.g. ALMA), including objects which perform
+a fast Fourier transform to map data to the uv-plane.
 
 .. autosummary::
    :toctree: _autosummary
@@ -53,3 +53,25 @@ Interferometer
    Interferometer
    SettingsInterferometer
    SimulatorInterferometer
+   Visibilities
+   TransformerDFT
+   TransformerNUFFT
+
+1D Data Structures
+------------------
+
+One-dimensional data structures store and mask 1D arrays and grids of (x) Cartesian
+coordinates.
+
+Their most common use is manipulating 1D representations of a light profile (e.g. computing
+the intensity versus radius in 1D).
+
+.. autosummary::
+   :toctree: _autosummary
+   :template: custom-class-template.rst
+   :recursive:
+
+   Mask1D
+   Array1D
+   ValuesIrregular
+   Grid1D
