+++ docs/api/plot.rst
@@ -2,8 +2,20 @@
 Plotting
 ========
 
+**PyAutoGalaxy** custom visualization library.
+
+Step-by-step Juypter notebook guides illustrating all objects listed on this page are
+provided on the `autogalaxy_workspace: plot tutorials <https://github.com/Jammy2211/autogalaxy_workspace/tree/release/notebooks/plot>`_ and
+it is strongly recommended you use those to learn plot customization.
+
+**Examples / Tutorials:**
+
+- `autogalaxy_workspace: plot tutorials <https://github.com/Jammy2211/autogalaxy_workspace/tree/release/notebooks/plot>`_
+
 Plotters [aplt]
 ---------------
+
+Create figures and subplots showing quantities of standard **PyAutoGalaxy** objects.
 
 .. currentmodule:: autogalaxy.plot
 
@@ -34,6 +46,9 @@ Plotters [aplt]
 Non-linear Search Plotters [aplt]
 ---------------------------------
 
+Create figures and subplots of non-linear search specific visualization of every search algorithm supported
+by **PyAutoGalaxy**.
+
 .. currentmodule:: autogalaxy.plot
 
 .. autosummary::
@@ -49,6 +64,9 @@ Non-linear Search Plotters [aplt]
 
 Plot Customization [aplt]
 -------------------------
+
+Customize figures created via ``Plotter`` objects, including changing ``matplotlib`` settings and adding
+visuals to figures.
 
 .. currentmodule:: autogalaxy.plot
 
@@ -66,6 +84,9 @@ Plot Customization [aplt]
 
 Matplot Lib Wrappers [aplt]
 ---------------------------
+
+Wrappers for every ``matplotlib`` function used by a ``Plotter``, allowing for detailed customizaiton of
+every figure and subplot.
 
 .. currentmodule:: autogalaxy.plot
 
