+++ docs/api/modeling.rst
@@ -5,6 +5,10 @@ Modeling
 Analysis
 ========
 
+The ``Analysis`` objects define the ``log_likelihood_function`` of how a galaxy model is fitted to a dataset.
+
+It acts as an interface between the data, model and the non-linear search.
+
 .. currentmodule:: autogalaxy
 
 .. autosummary::
@@ -17,6 +21,11 @@ Analysis
 
 Non-linear Searches
 -------------------
+
+A non-linear search is an algorithm which fits a model to data.
+
+**PyAutoGalaxy** currently supports three types of non-linear search algorithms: nested samplers,
+Markov Chain Monte Carlo (MCMC) and optimizers.
 
 .. currentmodule:: autofit
 
@@ -33,6 +42,8 @@ Non-linear Searches
 
 Priors
 ------
+
+The priors of parameters of every component of a mdoel, which is fitted to data, are customized using ``Prior`` objects.
 
 .. autosummary::
    :toctree: _autosummary
