+++ docs/conf.py
@@ -120,7 +120,16 @@ numpydoc_class_members_toctree = True
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
+exclude_patterns = [
+    "_build",
+    "Thumbs.db",
+    ".DS_Store",
+    "CODE_OF_CONDUCT.md",
+    "CONTRIBUTING.md",
+    "CITATIONS.rst",
+    "README.rst",
+    "README.md"
+]
 
 
 # -- Options for HTML output -------------------------------------------------
