+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,307 @@
+# PyAutoGalaxy Code of Conduct
+
+**Table of Contents**
+
+- [The Short Version](#the-short-version)
+- [The Longer Version](#the-longer-version)
+  - [PyAutoGalaxy Diversity Statement](#project-diversity-statement)
+  - [PyAutoGalaxy Code of Conduct: Introduction & Scope](#project-code-of-conduct-introduction--scope)
+  - [Standards for Behavior](#standards-for-behavior)
+    - [Unacceptable Behavior](#unacceptable-behavior)
+  - [Reporting Guidelines](#reporting-guidelines)
+    - [How to Submit a Report](#how-to-submit-a-report)
+    - [Person(s) Responsible for Resolving Complaints](#persons-responsible-for-resolving-complaints)
+    - [Conflicts of Interest](#conflicts-of-interest)
+    - [What to Include in a Report](#what-to-include-in-a-report)
+  - [Enforcement: What Happens After a Report is Filed?](#enforcement-what-happens-after-a-report-is-filed)
+    - [Acknowledgment and Responding to Immediate Needs](#acknowledgment-and-responding-to-immediate-needs)
+    - [Reviewing the Report](#reviewing-the-report)
+    - [Contacting the Person Reported](#contacting-the-person-reported)
+    - [Response and Potential Consequences](#response-and-potential-consequences)
+    - [Appealing a Decision](#appealing-a-decision)
+  - [Timeline Summary:](#timeline-summary)
+    - [Confirming Receipt](#confirming-receipt)
+    - [Reviewing the Report](#reviewing-the-report-1)
+    - [Consequences & Resolution](#consequences--resolution)
+- [License](#license)
+
+## The Short Version
+
+Be kind to others. Do not insult or put down others. Behave professionally. Remember that harassment and sexist, 
+racist, or exclusionary jokes are not appropriate for PyAutoGalaxy.
+
+All communication should be appropriate for a professional audience including people of many different backgrounds. 
+Sexual language and imagery is not appropriate.
+
+PyAutoGalaxy is dedicated to providing a harassment-free community for everyone, regardless of gender, sexual orientation, 
+gender identity and expression, disability, physical appearance, body size, race, or religion. We do not tolerate 
+harassment of community members in any form.
+
+Thank you for helping make this a welcoming, friendly community for all.
+
+## The Longer Version
+
+### PyAutoGalaxy Diversity Statement
+
+PyAutoGalaxy welcomes and encourages participation in our community by people of all backgrounds and identities. We 
+are committed to promoting and sustaining a culture that values mutual respect, tolerance, and learning, and we work 
+together as a community to help each other live out these values.
+
+We have created this diversity statement because we believe that a diverse community is stronger, more vibrant, 
+and produces better software and better science. A diverse community where people treat each other with respect has 
+more potential contributors, more sources for ideas, and fewer shared assumptions that might hinder development 
+or research.
+
+Although we have phrased the formal diversity statement generically to make it all-inclusive, we recognize that there 
+are specific identities that are impacted by systemic discrimination and marginalization. We welcome all people to 
+participate in the PyAutoGalaxy community regardless of their identity or background.
+
+### PyAutoGalaxy Code of Conduct: Introduction & Scope
+
+This code of conduct should be honored by everyone who participates in the PyAutoGalaxy community. It should be 
+honored in any PyAutoGalaxy-related activities, by anyone claiming affiliation with PyAutoGalaxy, and especially when 
+someone is representing PyAutoGalaxy in any role (including as an event volunteer or speaker).
+
+This code of conduct applies to all spaces managed by PyAutoGalaxy, including all public and private mailing lists, 
[…]
