### 2022-12-20 754f7dd
class template
