+++ autogalaxy/plane/to_inversion.py
@@ -208,8 +208,8 @@ class PlaneToInversion(AbstractToInversion):
             return None
 
         return [
-            pixelization.mesh.data_mesh_grid_from(
-                data_grid_slim=self.grid_pixelization,
+            pixelization.mesh.image_plane_mesh_grid_from(
+                image_plane_data_grid=self.grid_pixelization,
                 hyper_data=hyper_galaxy_image,
                 settings=self.settings_pixelization,
             )
@@ -223,15 +223,15 @@ class PlaneToInversion(AbstractToInversion):
         self,
         mesh: aa.AbstractMesh,
         regularization: aa.AbstractRegularization,
-        source_mesh_grid: aa.Grid2DSparse,
+        source_plane_mesh_grid: aa.Grid2DSparse,
         hyper_galaxy_image: aa.Array2D,
-        data_mesh_grid: aa.Grid2DSparse = None,
+        image_plane_mesh_grid: aa.Grid2DSparse = None,
     ) -> aa.AbstractMapper:
 
         mapper_grids = mesh.mapper_grids_from(
-            source_grid_slim=self.grid_pixelization,
-            source_mesh_grid=source_mesh_grid,
-            data_mesh_grid=data_mesh_grid,
+            source_plane_data_grid=self.grid_pixelization,
+            source_plane_mesh_grid=source_plane_mesh_grid,
+            image_plane_mesh_grid=image_plane_mesh_grid,
             hyper_data=hyper_galaxy_image,
             settings=self.settings_pixelization,
             preloads=self.preloads,
@@ -261,9 +261,9 @@ class PlaneToInversion(AbstractToInversion):
             mapper = self.mapper_from(
                 mesh=pixelization_list[mapper_index].mesh,
                 regularization=pixelization_list[mapper_index].regularization,
-                source_mesh_grid=sparse_grid_list[mapper_index],
+                source_plane_mesh_grid=sparse_grid_list[mapper_index],
                 hyper_galaxy_image=hyper_galaxy_image_list[mapper_index],
-                data_mesh_grid=sparse_grid_list[mapper_index],
+                image_plane_mesh_grid=sparse_grid_list[mapper_index],
             )
 
             galaxy = galaxies_with_pixelization_list[mapper_index]
