+++ test_autogalaxy/config/visualize.yaml
@@ -15,9 +15,9 @@ include:
     critical_curves: false
     grid: false
     light_profile_centres: true
-    mapper_data_mesh_grid: true
-    mapper_source_grid_slim: false
-    mapper_source_mesh_grid: false
+    mapper_image_plane_mesh_grid: true
+    mapper_source_plane_data_grid: false
+    mapper_source_plane_mesh_grid: false
     mask: true
     mass_profile_centres: true
     multiple_images: false
