+++ test_autogalaxy/plot/mat_wrap/config/visualize/include.ini
@@ -8,9 +8,9 @@ caustics=False
 multiple_images=False
 light_profile_centres=True
 mass_profile_centres=True
-mapper_data_mesh_grid=True
-mapper_source_mesh_grid=False
-mapper_source_grid_slim=False
+mapper_image_plane_mesh_grid=True
+mapper_source_plane_mesh_grid=False
+mapper_source_plane_data_grid=False
 parallel_overscan=True
 serial_prescan=True
 serial_overscan=True
