+++ docs/_templates/custom-class-template.rst
@@ -14,6 +14,7 @@
 
    .. autosummary::
       :nosignatures:
+      :exclude-members: ndarray
    {% for item in methods %}
       {%- if not item.startswith('_') %}
       ~{{ name }}.{{ item }}
@@ -27,6 +28,7 @@
    .. rubric:: {{ _('Attributes') }}
 
    .. autosummary::
+      :exclude-members: ndarray
    {% for item in attributes %}
       {%- if not item.startswith('_') %}
       ~{{ name }}.{{ item }}
