### 2022-12-21 d54a8a5
update galaxy priors
