+++ autogalaxy/config/priors/galaxy/README.rst
@@ -1,0 +1,7 @@
+The ``galaxy`` folder contains configuration files for the default priors assumed for ``Galaxy`` object properties **PyAutoGalaxy**.
+
+Folders
+-------
+
+- ``redshift``: Configs for priors on the redshift of a galaxy, when treated as a free parameter.
+- ``hyper``: Configs for hyper-galaxy model components which scale the noise-map.
