+++ autogalaxy/config/priors/galaxy/redshift.yaml
@@ -1,0 +1,11 @@
+Redshift:
+  redshift:
+    type: Uniform
+    lower_limit: 0.0
+    upper_limit: 3.0
+    width_modifier:
+      type: Absolute
+      value: 1.0
+    gaussian_limits:
+      lower: 0.0
+      upper: inf
