+++ autogalaxy/config/priors/galaxy/hyper.yaml
@@ -29,14 +29,3 @@ HyperGalaxy:
     gaussian_limits:
       lower: -inf
       upper: inf
-Redshift:
-  redshift:
-    type: Uniform
-    lower_limit: 0.0
-    upper_limit: 3.0
-    width_modifier:
-      type: Absolute
-      value: 1.0
-    gaussian_limits:
-      lower: 0.0
-      upper: inf
