### 2022-12-21 eab9f2b
more docstring alignment
