### 2022-12-23 e3117f8
moved grid_scaled_2d_for_marching_squares_from
