+++ test_autogalaxy/plane/test_plane.py
@@ -28,8 +28,8 @@ def critical_curve_via_magnification_via_plane_from(plane, grid):
         contour_x, contour_y = contours[jj].T
         pixel_coord = np.stack((contour_x, contour_y), axis=-1)
 
-        critical_curve = grid.mask.grid_scaled_for_marching_squares_from(
-            grid_pixels_1d=pixel_coord, shape_native=magnification.sub_shape_native
+        critical_curve = grid.mask.grid_scaled_2d_for_marching_squares_from(
+            grid_pixels_2d=pixel_coord, shape_native=magnification.sub_shape_native
         )
 
         critical_curve = np.array(grid=critical_curve)
