+++ autogalaxy/gui/clicker.py
@@ -50,8 +50,8 @@ class Clicker:
                         y_pixels_max = y
                         x_pixels_max = x
 
-            grid_arcsec = self.image.mask.grid_scaled_from(
-                grid_pixels_1d=aa.Grid2D.manual_native(
+            grid_arcsec = self.image.mask.grid_scaled_2d_from(
+                grid_pixels_2d=aa.Grid2D.manual_native(
                     grid=[[[y_pixels_max + 0.5, x_pixels_max + 0.5]]],
                     pixel_scales=self.pixel_scales,
                 )
