+++ test_autogalaxy/plot/mat_wrap/test_get_visuals.py
@@ -124,7 +124,9 @@ def test__2d__via_light_obj_from(lp_0, grid_2d_7x7):
 
     assert visuals_2d_via.origin.in_list == [(0.0, 0.0)]
     assert (visuals_2d_via.mask == grid_2d_7x7.mask).all()
-    assert (visuals_2d_via.border == grid_2d_7x7.mask.border_grid_sub_1.binned).all()
+    assert (
+        visuals_2d_via.border == grid_2d_7x7.mask.derived_grids.border_sub_1.binned
+    ).all()
     assert visuals_2d_via.light_profile_centres.in_list == [lp_0.centre]
     assert visuals_2d_via.vectors == 2
 
@@ -160,7 +162,9 @@ def test__2d__via_mass_obj(mp_0, grid_2d_7x7):
 
     assert visuals_2d_via.origin.in_list == [(0.0, 0.0)]
     assert (visuals_2d_via.mask == grid_2d_7x7.mask).all()
-    assert (visuals_2d_via.border == grid_2d_7x7.mask.border_grid_sub_1.binned).all()
+    assert (
+        visuals_2d_via.border == grid_2d_7x7.mask.derived_grids.border_sub_1.binned
+    ).all()
     assert visuals_2d_via.mass_profile_centres.in_list == [mp_0.centre]
     assert (
         visuals_2d_via.critical_curves[0]
@@ -208,7 +212,9 @@ def test__2d__via_light_mass_obj(gal_x1_lp_x1_mp, grid_2d_7x7):
 
     assert visuals_2d_via.origin.in_list == [(0.0, 0.0)]
     assert (visuals_2d_via.mask == grid_2d_7x7.mask).all()
-    assert (visuals_2d_via.border == grid_2d_7x7.mask.border_grid_sub_1.binned).all()
+    assert (
+        visuals_2d_via.border == grid_2d_7x7.mask.derived_grids.border_sub_1.binned
+    ).all()
     assert visuals_2d_via.light_profile_centres.in_list == [
         gal_x1_lp_x1_mp.light_profile_0.centre
     ]
@@ -264,7 +270,8 @@ def test__via_fit_imaging_from(fit_imaging_x2_galaxy_7x7, grid_2d_7x7):
     assert visuals_2d_via.origin == (1.0, 1.0)
     assert (visuals_2d_via.mask == fit_imaging_x2_galaxy_7x7.mask).all()
     assert (
-        visuals_2d_via.border == fit_imaging_x2_galaxy_7x7.mask.border_grid_sub_1.binned
+        visuals_2d_via.border
+        == fit_imaging_x2_galaxy_7x7.mask.derived_grids.border_sub_1.binned
     ).all()
     assert visuals_2d_via.light_profile_centres.in_list == [(0.0, 0.0), (0.0, 0.0)]
     assert visuals_2d_via.mass_profile_centres.in_list == [(0.0, 0.0)]
