### 2022-12-29 806d5fc
remove Mask2D.manual / Mask1D.manual
