+++ test_autogalaxy/util/test_plane_util.py
@@ -22,7 +22,7 @@ class TestPlaneImageFromGrid:
             shape=(3, 3), grid=grid, galaxies=[galaxy], buffer=0.0
         )
 
-        mask = ag.Mask2D.manual(
+        mask = ag.Mask2D(
             mask=np.full(shape=(3, 3), fill_value=False), pixel_scales=1.0, sub_size=1
         )
 
@@ -67,7 +67,7 @@ class TestPlaneImageFromGrid:
             shape=(3, 3), grid=grid, galaxies=[galaxy], buffer=0.0
         )
 
-        mask = ag.Mask2D.manual(
+        mask = ag.Mask2D(
             mask=np.full(shape=(3, 3), fill_value=False), pixel_scales=1.0, sub_size=1
         )
 
@@ -101,7 +101,7 @@ class TestPlaneImageFromGrid:
             shape=(2, 3), grid=grid, galaxies=[galaxy], buffer=0.0
         )
 
-        mask = ag.Mask2D.manual(
+        mask = ag.Mask2D(
             mask=np.full(shape=(2, 3), fill_value=False), pixel_scales=1.0, sub_size=1
         )
 
@@ -132,7 +132,7 @@ class TestPlaneImageFromGrid:
             shape=(3, 2), grid=grid, galaxies=[galaxy], buffer=0.0
         )
 
-        mask = ag.Mask2D.manual(
+        mask = ag.Mask2D(
             mask=np.full(shape=(3, 2), fill_value=False), pixel_scales=1.0, sub_size=1
         )
 
@@ -163,7 +163,7 @@ class TestPlaneImageFromGrid:
             shape=(3, 3), grid=grid_without_buffer, galaxies=[galaxy], buffer=0.02
         )
 
-        mask = ag.Mask2D.manual(
+        mask = ag.Mask2D(
             mask=np.full(shape=(3, 3), fill_value=False), pixel_scales=1.0, sub_size=1
         )
 
