+++ test_autogalaxy/interferometer/test_simulate_and_fit_interferometer.py
@@ -51,7 +51,7 @@ def test__perfect_fit__chi_squared_0():
         uv_wavelengths_path=path.join(file_path, "uv_wavelengths.fits"),
     )
 
-    real_space_mask = ag.Mask2D.unmasked(
+    real_space_mask = ag.Mask2D.all_false(
         shape_native=(51, 51), pixel_scales=0.1, sub_size=2
     )
 

[…]
