+++ autogalaxy/quantity/dataset_quantity.py
@@ -91,7 +91,7 @@ class DatasetQuantity(AbstractDataset):
 
             if data.shape[0:-1] == noise_map.shape[0:]:
 
-                noise_map = aa.VectorYX2D.manual_native(
+                noise_map = aa.VectorYX2D._manual_native(
                     vectors=np.stack((noise_map, noise_map), axis=-1),
                     pixel_scales=data.pixel_scales,
                     sub_size=data.sub_size,
