+++ test_autogalaxy/profiles/test_radial_minima.py
@@ -1,10 +1,8 @@
 from __future__ import division, print_function
 
 from os import path
-from autoconf import conf
 import autogalaxy as ag
 
-import numpy as np
 import pytest
 
 directory = path.dirname(path.realpath(__file__))
@@ -12,8 +10,8 @@ directory = path.dirname(path.realpath(__file__))
 
 def test__grid_2d__moves_radial_coordinates__does_not_double_transform():
 
-    grid_2d = ag.Grid2D.manual_native(grid=[[[0.0, 0.0]]], pixel_scales=1.0)
-    grid_2d_offset = ag.Grid2D.manual_native(grid=[[0.0001, 0.0001]], pixel_scales=1)
+    grid_2d = ag.Grid2D.without_mask(grid=[[[0.0, 0.0]]], pixel_scales=1.0)
+    grid_2d_offset = ag.Grid2D.without_mask(grid=[[[0.0001, 0.0001]]], pixel_scales=1)
 
     isothermal = ag.mp.Isothermal(centre=(0.0, 0.0), einstein_radius=1.0)
 
@@ -22,8 +20,8 @@ def test__grid_2d__moves_radial_coordinates__does_not_double_transform():
 
     assert convergence_0 == pytest.approx(convergence_1, 1.0e-8)
 
-    grid_2d = ag.Grid2D.manual_native(grid=[[[0.5, 0.5]]], pixel_scales=1.0)
-    grid_2d_offset = ag.Grid2D.manual_native(grid=[[0.5001, 0.5001]], pixel_scales=1)
+    grid_2d = ag.Grid2D.without_mask(grid=[[[0.5, 0.5]]], pixel_scales=1.0)
+    grid_2d_offset = ag.Grid2D.without_mask(grid=[[[0.5001, 0.5001]]], pixel_scales=1)
 
     isothermal = ag.mp.Isothermal(centre=(0.0, 0.0), einstein_radius=1.0)
 
