### 2023-01-12 e738264
fix blurring image unit test
