+++ test_autogalaxy/imaging/test_fit_imaging.py
@@ -560,35 +560,19 @@ def test___unmasked_blurred_images(masked_imaging_7x7):
 
 def test__subtracted_images_of_galaxies(masked_imaging_7x7_no_blur):
 
-    g0 = ag.Galaxy(redshift=0.5, bulge=ag.m.MockLightProfile(image_2d=np.ones(9)))
+    g0 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=1.0))
 
-    g1 = ag.Galaxy(redshift=0.5, bulge=ag.m.MockLightProfile(image_2d=2.0 * np.ones(9)))
+    g1 = ag.Galaxy(redshift=0.5)
 
-    g2 = ag.Galaxy(redshift=0.5, bulge=ag.m.MockLightProfile(image_2d=3.0 * np.ones(9)))
+    g2 = ag.Galaxy(redshift=0.5, bulge=ag.lp.Sersic(intensity=3.0))
 
     plane = ag.Plane(redshift=0.5, galaxies=[g0, g1, g2])
 
     fit = ag.FitImaging(dataset=masked_imaging_7x7_no_blur, plane=plane)
 
-    fit.subtracted_images_of_galaxies_list  # This stops a nan from being computed in the assertion, which is weird.
-
-    assert fit.subtracted_images_of_galaxies_list[0].slim[0] == -4.0 or np.nan
-    assert fit.subtracted_images_of_galaxies_list[1].slim[0] == -3.0 or np.nan
-    assert fit.subtracted_images_of_galaxies_list[2].slim[0] == -2.0 or np.nan
-
-    # g0 = ag.Galaxy(redshift=0.5, bulge=ag.m.MockLightProfile(image_2d=np.ones(16)))
-    #
-    # g1 = ag.Galaxy(redshift=0.5)
-    #
-    # g2 = ag.Galaxy(redshift=0.5, bulge=ag.m.MockLightProfile(image_2d=3.0 * np.ones(16)))
-    #
-    # plane = ag.Plane(redshift=0.5, galaxies=[g0, g1, g2])
-    #
-    # fit = ag.FitImaging(dataset=masked_imaging_7x7_no_blur, plane=plane)
-    #
-    # assert fit.subtracted_images_of_galaxies_list[0].slim[0] == -2.0 or np.nan
-    # assert fit.subtracted_images_of_galaxies_list[1].slim[0] == -3.0 or np.nan
-    # assert fit.subtracted_images_of_galaxies_list[2].slim[0] == 0.0 or np.nan
+    assert fit.subtracted_images_of_galaxies_list[0].slim[0] == pytest.approx(0.520383, 1.0e-4)
+    assert fit.subtracted_images_of_galaxies_list[1].slim[0] == pytest.approx(0.360511, 1.0e-4)
+    assert fit.subtracted_images_of_galaxies_list[2].slim[0] == pytest.approx(0.840127, 1.0e-4)
 
 
 def test__light_profile_linear__intensity_dict(masked_imaging_7x7):
