### 2023-01-13 99e34c4
all strcuture inputs now values
