+++ test_autogalaxy/quantity/test_dataset_quantity.py
@@ -7,9 +7,9 @@ from autogalaxy import exc
 
 def test_via_signal_to_noise_map(dataset_quantity_7x7_array_2d, sub_mask_2d_7x7):
 
-    data = ag.Array2D.without_mask(array=[[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
-    signal_to_noise_map = ag.Array2D.without_mask(
-        array=[[1.0, 5.0], [15.0, 40.0]], pixel_scales=1.0
+    data = ag.Array2D.no_mask(values=[[1.0, 2.0], [3.0, 4.0]], pixel_scales=1.0)
+    signal_to_noise_map = ag.Array2D.no_mask(
+        values=[[1.0, 5.0], [15.0, 40.0]], pixel_scales=1.0
     )
 
     dataset_quantity = ag.DatasetQuantity.via_signal_to_noise_map(
@@ -23,11 +23,11 @@ def test_via_signal_to_noise_map(dataset_quantity_7x7_array_2d, sub_mask_2d_7x7)
         np.array([[1.0, 0.4], [0.2, 0.1]]), 1.0e-4
     )
 
-    data = ag.VectorYX2D._manual_native(
-        vectors=[[[1.0, 1.0], [2.0, 2.0]], [[3.0, 3.0], [4.0, 4.0]]], pixel_scales=1.0
+    data = ag.VectorYX2D.no_mask(
+        values=[[[1.0, 1.0], [2.0, 2.0]], [[3.0, 3.0], [4.0, 4.0]]], pixel_scales=1.0
     )
-    signal_to_noise_map = ag.Array2D.without_mask(
-        array=[[1.0, 5.0], [15.0, 40.0]], pixel_scales=1.0
+    signal_to_noise_map = ag.Array2D.no_mask(
+        values=[[1.0, 5.0], [15.0, 40.0]], pixel_scales=1.0
     )
 
     dataset_quantity = ag.DatasetQuantity.via_signal_to_noise_map(
@@ -97,14 +97,14 @@ def test__grid(
 
 def test__vector_data__y_x():
 
-    data = ag.VectorYX2D._manual_native(
-        vectors=[[[1.0, 5.0], [2.0, 6.0]], [[3.0, 7.0], [4.0, 8.0]]],
+    data = ag.VectorYX2D.no_mask(
+        values=[[[1.0, 5.0], [2.0, 6.0]], [[3.0, 7.0], [4.0, 8.0]]],
         pixel_scales=1.0,
         sub_size=1,
     )
 
-    noise_map = ag.VectorYX2D._manual_native(
-        vectors=[[[1.1, 5.1], [2.1, 6.1]], [[3.1, 7.1], [4.1, 8.1]]],
+    noise_map = ag.VectorYX2D.no_mask(
+        values=[[[1.1, 5.1], [2.1, 6.1]], [[3.1, 7.1], [4.1, 8.1]]],
         pixel_scales=1.0,
         sub_size=1,
     )
