+++ autogalaxy/analysis/result.py
@@ -159,7 +159,7 @@ class ResultDataset(Result):
         The hyper model image is the sum of the hyper galaxy image of every individual galaxy.
         """
         hyper_model_image = aa.Array2D(
-            array=np.zeros(self.mask.derive_mask.sub_1.pixels_in_mask),
+            values=np.zeros(self.mask.derive_mask.sub_1.pixels_in_mask),
             mask=self.mask.derive_mask.sub_1,
         )
 
