+++ autogalaxy/galaxy/galaxy.py
@@ -472,7 +472,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
 
         would return:
 
-        ValuesIrregular(values=[axis_ratio_0, axis_ratio_1])
+        ArrayIrregular(values=[axis_ratio_0, axis_ratio_1])
 
         If a galaxy has three mass profiles and we want the `MassProfile` centres, the following:
 
@@ -503,7 +503,7 @@ class Galaxy(af.ModelObject, OperateImageList, OperateDeflections, Dictable):
         if attributes == []:
             return None
         elif isinstance(attributes[0], float):
-            return aa.ValuesIrregular(values=attributes)
+            return aa.ArrayIrregular(values=attributes)
         elif isinstance(attributes[0], tuple):
             return aa.Grid2DIrregular(values=attributes)
 
