+++ test_autogalaxy/util/test_shear_field.py
@@ -13,20 +13,20 @@ def test__elliptical_properties_and_patches():
         pixel_scales=1.0,
     )
 
-    assert isinstance(vectors.ellipticities, ag.ValuesIrregular)
+    assert isinstance(vectors.ellipticities, ag.ArrayIrregular)
     assert vectors.ellipticities.in_list == [1.0, 1.0, np.sqrt(2.0), 0.0]
 
-    assert isinstance(vectors.semi_major_axes, ag.ValuesIrregular)
+    assert isinstance(vectors.semi_major_axes, ag.ArrayIrregular)
     assert vectors.semi_major_axes.in_list == pytest.approx(
         [6.0, 6.0, 7.242640, 3.0], 1.0e-4
     )
 
-    assert isinstance(vectors.semi_minor_axes, ag.ValuesIrregular)
+    assert isinstance(vectors.semi_minor_axes, ag.ArrayIrregular)
     assert vectors.semi_minor_axes.in_list == pytest.approx(
         [0.0, 0.0, -1.242640, 3.0], 1.0e-4
     )
 
-    assert isinstance(vectors.phis, ag.ValuesIrregular)
+    assert isinstance(vectors.phis, ag.ArrayIrregular)
     assert vectors.phis.in_list == pytest.approx([0.0, 45.0, 22.5, 0.0], 1.0e-4)
 
     assert isinstance(vectors.elliptical_patches[0], Ellipse)
@@ -40,20 +40,20 @@ def test__elliptical_properties_and_patches():
         grid=[[1.0, -1.0], [1.0, 1.0], [0.0, 0.0]],
     )
 
-    assert isinstance(vectors.ellipticities, ag.ValuesIrregular)
+    assert isinstance(vectors.ellipticities, ag.ArrayIrregular)
     assert vectors.ellipticities.in_list == [1.0, 1.0, np.sqrt(2.0)]
 
-    assert isinstance(vectors.semi_major_axes, ag.ValuesIrregular)
+    assert isinstance(vectors.semi_major_axes, ag.ArrayIrregular)
     assert vectors.semi_major_axes.in_list == pytest.approx(
         [6.0, 6.0, 7.242640], 1.0e-4
     )
 
-    assert isinstance(vectors.semi_minor_axes, ag.ValuesIrregular)
+    assert isinstance(vectors.semi_minor_axes, ag.ArrayIrregular)
     assert vectors.semi_minor_axes.in_list == pytest.approx(
         [0.0, 0.0, -1.242640], 1.0e-4
     )
 
-    assert isinstance(vectors.phis, ag.ValuesIrregular)
+    assert isinstance(vectors.phis, ag.ArrayIrregular)
     assert vectors.phis.in_list == pytest.approx([0.0, 45.0, 22.5], 1.0e-4)
 
     assert isinstance(vectors.elliptical_patches[0], Ellipse)
