+++ autogalaxy/quantity/dataset_quantity.py
@@ -188,8 +188,8 @@ class DatasetQuantity(AbstractDataset):
         else:
             unmasked_dataset = self.unmasked
 
-        data = self.data.apply_mask(mask=mask.derive_mask.sub_1).native
-        noise_map = self.noise_map.apply_mask(mask=mask.derive_mask.sub_1).native
+        data = self.data.apply_mask(mask=mask.derive_mask.sub_1)
+        noise_map = self.noise_map.apply_mask(mask=mask.derive_mask.sub_1)
 
         dataset = DatasetQuantity(
             data=data, noise_map=noise_map, settings=self.settings
