### 2023-01-14 9f82c74
shear refactored to gammas
