### 2023-01-17 #4475 @dfm
Thanks @Jammy2211! That DOI actually just includes a PDF of the paper, which isn't what we want. It should include the current version of the _software_.
