### 2023-01-19 #4475 @Jammy2211
I added a `zenodo.json` file to the repo with the correct information.

However, when I attempt a re-release in order to use it I get an error on the zenodo build:

```
{
    "errors": "New version's files must differ from all previous versions."
}
```

I have updated a few random source code files, but still get it.

I'll try again later via a full pypi rerelease, unless you know a quick fix for this?
