### 2023-01-19 #4475 @dfm
Sorry for the confusion! You can edit the metadata directly on the Zenodo website. If you're logged in, you should see an "Edit" button in the top right corner of https://zenodo.org/record/7546915 which lets you update the info.
