### 2023-01-23 #4475 @xuanxu
The error comes from invalid ISBN values in the paper.bib file, detected when creating the paper metadata for depositing.
Like these for example: [1](https://github.com/Jammy2211/PyAutoGalaxy/blob/main/paper/paper.bib#L556), [2](https://github.com/Jammy2211/PyAutoGalaxy/blob/main/paper/paper.bib#L914)
