### 2023-01-25 #4475 @Jammy2211
Would it be appropriate to simple delete the isbn field from everything in the .bib file?

For a lot of these papers I cant find the true ISBN (I guess it came somewhere from the metadata) and I dont think its an important part of citing?
