### 2023-01-25 8efeb78
another fix to pixels upper limit
