+++ autogalaxy/analysis/model_util.py
@@ -121,7 +121,51 @@ def set_upper_limit_of_pixelization_pixels_prior(
         the number of pixels in the mask below which the lower limit is set.
     """
 
-    # TODO : I'm sorry Rich
+    # if not hasattr(model, "galaxies"):
+    #     return
+    #
+    # mesh_list = model.galaxies.models_with_type(
+    #     cls=(
+    #         aa.mesh.DelaunayBrightnessImage, aa.mesh.VoronoiBrightnessImage, aa.mesh.VoronoiNNBrightnessImage))
+    #
+    # print(mesh_list)
+    #
+    # if not mesh_list:
+    #     return
+    #
+    # for mesh in mesh_list:
+    #
+    #     if pixels_in_mask < mesh.pixels.upper_limit:
+    #
+    #         lower_limit = mesh.pixels.lower_limit
+    #
+    #         log_str = (
+    #             "MODIFY BEFORE FIT -  A pixelization mesh's pixel UniformPrior upper limit"
+    #             "was greater than the number of pixels in the mask. It has been "
+    #             "reduced to the number of pixels in the mask.\,"
+    #         )
+    #
+    #         if lower_limit > pixels_in_mask:
+    #
+    #             lower_limit = (
+    #                 pixels_in_mask
+    #                 - lower_limit_no_pixels_below_mask
+    #             )
+    #
+    #             logger.info(
+    #                 log_str
+    #                 + "MODIFY BEFORE FIT - The pixelization's mesh's pixel UniformPrior lower_limit was "
+    #                 "also above the number of pixels in the mask, and has been reduced"
+    #                 "to the number of pixels in the mask minus 10."
+    #             )
+    #         else:
+    #             logger.info(log_str)
+    #
+    #         mesh.pixels = af.UniformPrior(
+    #             lower_limit=lower_limit,
+    #             upper_limit=pixels_in_mask,
+    #         )
+
 
     if not hasattr(model, "galaxies"):
         return
@@ -144,36 +188,38 @@ def set_upper_limit_of_pixelization_pixels_prior(
 
                         if hasattr(mesh, "pixels"):
 
-                            if pixels_in_mask < mesh.pixels.upper_limit:
+                            if hasattr(mesh.pixels, "upper_limit"):
 
-                                lower_limit = mesh.pixels.lower_limit
+                                if pixels_in_mask < mesh.pixels.upper_limit:
 
-                                log_str = (
-                                    "MODIFY BEFORE FIT -  A pixelization mesh's pixel UniformPrior upper limit"
-                                    "was greater than the number of pixels in the mask. It has been "
-                                    "reduced to the number of pixels in the mask.\,"
-                                )
+                                    lower_limit = mesh.pixels.lower_limit
 
-                                if lower_limit > pixels_in_mask:
-
-                                    lower_limit = (
-                                        pixels_in_mask
-                                        - lower_limit_no_pixels_below_mask
+                                    log_str = (
+                                        "MODIFY BEFORE FIT -  A pixelization mesh's pixel UniformPrior upper limit"
+                                        "was greater than the number of pixels in the mask. It has been "
+                                        "reduced to the number of pixels in the mask.\,"
                                     )
 
-                                    logger.info(
-                                        log_str
-                                        + "MODIFY BEFORE FIT - The pixelization's mesh's pixel UniformPrior lower_limit was "
-                                        "also above the number of pixels in the mask, and has been reduced"
-                                        "to the number of pixels in the mask minus 10."
-                                    )
-                                else:
-                                    logger.info(log_str)
+                                    if lower_limit > pixels_in_mask:
 
-                                mesh.pixels = af.UniformPrior(
-                                    lower_limit=lower_limit,
-                                    upper_limit=pixels_in_mask,
[…]
