+++ test_autogalaxy/analysis/test_model_util.py
@@ -49,13 +49,9 @@ def test__mesh_list_from_model():
 def test__set_upper_limit_of_pixelization_pixels_prior():
 
     mesh = af.Model(ag.mesh.DelaunayBrightnessImage)
-
     mesh.pixels = af.UniformPrior(lower_limit=5.0, upper_limit=10.0)
-
     pixelization = ag.Pixelization(mesh=mesh)
-
     galaxies = af.Collection(source=ag.Galaxy(redshift=0.5, pixelization=pixelization))
-
     model = af.Collection(galaxies=galaxies)
 
     ag.util.model.set_upper_limit_of_pixelization_pixels_prior(
