+++ paper/paper.bib
@@ -125,7 +125,6 @@ abstract = {Dynamic, interpreted languages, like Python, are attractive for doma
 author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
 doi = {10.1145/2833157.2833162},
 file = {:home/jammy/Documents/Papers/Software/numba{\_}sc15.pdf:pdf},
-isbn = {9781450340052},
 journal = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC - LLVM '15},
 keywords = {2,a jit for numeric,com-,compiler,jit,just-in-time,llvm,numba is a function-at-a-time,python},
 pages = {1--6},
@@ -162,7 +161,6 @@ arxivId = {0809.3437},
 author = {Feroz, F. and Hobson, M. P. and Bridges, M.},
 doi = {10.1111/j.1365-2966.2009.14548.x},
 eprint = {0809.3437},
-isbn = {0035-8711},
 issn = {00358711},
 journal = {Monthly Notices of the Royal Astronomical Society},
 keywords = {Methods: Data analysis,Methods: Statistical},
@@ -222,14 +220,6 @@ year = {2018}
     issue   = {21},
     doi     = {10.21105/joss.00433},
     url     = {https://doi.org/10.21105/joss.00433}
-}
- @book{python,
- author = {Van Rossum, Guido and Drake, Fred L.},
- title = {Python 3 Reference Manual},
- year = {2009},
- isbn = {1441412697},
- publisher = {CreateSpace},
- address = {Scotts Valley, CA}
 }
 @article{scikit-image,
   title={scikit-image: image processing in Python},
@@ -314,7 +304,6 @@ arxivId = {1507.02657},
 author = {Collett, Thomas E.},
 doi = {10.1088/0004-637X/811/1/20},
 eprint = {1507.02657},
-isbn = {0004-637x},
 issn = {15384357},
 journal = {Astrophysical Journal},
 keywords = {gravitational lensing: strong},
@@ -332,7 +321,6 @@ arxivId = {1108.0577},
 author = {Czoske, Oliver and Barnab{\`{e}}, Matteo and Koopmans, L{\'{e}}on V.E. and Treu, Tommaso and Bolton, Adam S.},
 doi = {10.1111/j.1365-2966.2011.19726.x},
 eprint = {1108.0577},
-isbn = {00358711},
 issn = {00358711},
 journal = {Monthly Notices of the Royal Astronomical Society},
 keywords = {galaxies: elliptical and lenticular, cD,galaxies: kinematics and dynamics,galaxies: structure,Gravitational lensing: strong,Techniques: spectroscopic},
@@ -395,7 +383,6 @@ arxivId = {0906.1349},
 author = {Koopmans, L. V.E. and Bolton, A. and Treu, T. and Czoske, O. and Auger, M. W. and Barnab{\`{e}}, M. and Vegetti, S. and Gavazzi, R. and Moustakas, L. A. and Burles, S.},
 doi = {10.1088/0004-637X/703/1/L51},
 eprint = {0906.1349},
-isbn = {1522-1601 (Electronic)$\backslash$r0161-7567 (Linking)},
 issn = {15384357},
 journal = {Astrophysical Journal},
 keywords = {galaxies: structure,Gravitational lensing},
@@ -553,7 +540,6 @@ arxivId = {0805.0201},
 author = {Vegetti, S. and Koopmans, L. V.E.},
 doi = {10.1111/j.1365-2966.2008.14005.x},
 eprint = {0805.0201},
[…]
