### 2023-01-25 f2dd33a
Removing abstracts from .bib
Ref: https://github.com/openjournals/joss-reviews/issues/4475
